### 2023-09-21 676a7ce
Fix lost changes on reload after loading from URL
