### 2023-09-21 79d201c
Added DOM function to get key signature of a note
