### 2023-09-21 130a23d
Store key.sig as attribute in speed mode
