### 2023-09-21 db765ef
Simplify addKeySigElement()
