### 2023-09-21 59c9736
Show validation report after manual initiation
