### 2023-09-22 596328a
Refactor check accid with helper functions
