### 2023-09-23 01b3a30
Add converters btw pname/accidGes and base-40 int
