### 2023-09-23 2e97e87
Optimize conversion for base-40 system
