### 2023-09-23 c6a68eb
Comment base-40 converters
