### 2023-09-23 8fd116d
Refactor base-40 system into separate file
