+++ app/static/lib/attribute-classes.js
@@ -282,142 +282,3 @@ export const dataDurationMensural = [
   'fusa',
   'semifusa',
 ];
-
-// See https://wiki.ccarh.org/wiki/Base_40 (accessed 23 Sept 2023)
-// C4 is 162
-// pitch = octave * 40 + chroma
-// octave = value / 40
-// chroma = value - octave * 40
-
-// base-40 constants
-const diatonicSteps = [2, 8, 14, 19, 25, 31, 37]; // numbers of diatonic steps in base-40 system
-const accidGesturalAndAlternation = {
-  s: 1,
-  f: -1,
-  ss: 2,
-  ff: -2,
-  ts: 3,
-  tf: -3,
-  n: 0,
-};
-const accidWrittenAndAlternation = {
-  s: 1,
-  f: -1,
-  x: 2,
-  ss: 2,
-  ff: -2,
-  xs: 3,
-  sx: 3,
-  ts: 3,
-  tf: -3,
-  n: 0,
-  nf: -1,
-  ns: 1,
-};
-
-/**
- * Returns accid.ges conform string for alteration integer (in semi-tones)
- * @param {string|number} alteration (e.g., -3, -2, -1, 0, 1, 2, 3)
- * @returns {string|null}
- */
-export function alterationToAccidGes(alteration = 0) {
-  if (alteration <= 3 && alteration >= -3) {
-    let i = Object.values(accidGesturalAndAlternation).indexOf(alteration);
-    return Object.keys(accidGesturalAndAlternation)[i];
-  } else {
-    console.log('Provide numbers between -2 and +2.');
-    return null;
-  }
-} // alterationToAccidGes()
-
-/**
- * Returns accid.written conform string for alteration integer (in semi-tones). 
- * For 2 it returns 'x', for 3 'xs' among the choices.
- * @param {string|number} alteration (e.g., -3, -2, -1, 0, 1, 2, 3)
- * @returns {string|null}
- */
-export function alterationToAccidWritten(alteration = 0) {
-  if (alteration <= 3 && alteration >= -3) {
-    let i = Object.values(accidWrittenAndAlternation).indexOf(alteration);
-    return Object.keys(accidWrittenAndAlternation)[i];
-  } else {
-    console.log('Provide numbers between -3 and +3.');
-    return null;
-  }
-} // alterationToAccidWritten()
-
-/**
- * Returns alteration integer (-2 for 'ff', 1 for 's') for accid.ges values
- * @param {string} accidGes
- * @returns {number|null}
- */
-export function accidGesToAlteration(accidGes = 'n') {
-  if (dataAccidentalGestural.includes(accidGes)) {
-    return accidGesturalAndAlternation[accidGes];
-  } else {
-    console.log('Please provide valid values for accid.ges.');
-    return null;
-  }
-} // accidGesToAlteration()
-
-/**
- * Returns alteration integer (-2 for 'ff', 3 for 'xs') for accid.written values
- * @param {string} accidGes
- * @returns {number|null}
- */
-export function accidWrittenToAlteration(accid = 'n') {
-  if (dataAccidentalWritten.includes(accid)) {
-    return accidWrittenAndAlternation[accid];
-  } else {
-    console.log('Please provide valid values for accid.written.');
-    return null;
-  }
-} // accidWrittenToAlteration()
-
-/**
- * Converts base-40 integer to object with keys: pname, accidGes, oct
- * @param {number} base40int
- * @returns {Object|null} with keys pname, accid.gestural.basic, and oct
- * When extracting base-40 chroma, just ignore oct=0.
- * Return null for negative input values.
- *
- * Usage: const { oct, pname, accidGes } = base40ToPitch(value)
- */
-export function base40ToPitch(base40int = 0) {
-  if (base40int < 0) {
[…]
