+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # mei-friend-online CHANGELOG.md
 
-### 1.0.1 Version 1.0.1 patch
+### 1.0.1 patch
 * Reduce mei-friend-specific Verovio defaults
 * Add auto-translated draft of Japanese language pack
 * Add keyboard shortcuts to editor for saving, loading and printing files
 * Generate URL with all available parameters and lower precision
 * Fix for facsimile images not loading without zone elements (#45)
+* Fix bug when modified MEI encoding was not restored from local storage when working from GitHub
 * Always show validation report upon manual validation, but not during auto validation
 
 ## 1.0.0 Version 1.0 (released 4 September 2023)
