### 2023-09-26 8f42907
Refactor and comment utils
