+++ app/static/lib/utils.js
@@ -6,7 +6,12 @@ const xmlIdString = /(?:xml:id=)(?:['"])(\S+?)(?:['"])/;
 const numberLikeString = /(?:n=)(?:['"])(\d+?)(?:['"])/;
 const closingTag = /(?:<[/])(\S+?)(?:[>])/;
 
-// returns the object named key of the JSON object obj
+/**
+ * JSON helper function: returns the object named key of the JSON object obj
+ * @param {*} key
+ * @param {*} obj
+ * @returns
+ */
 export function findKey(key, obj) {
   // console.info('findKey: ' + key + '. ', obj);
   if (typeof obj === 'object') {
@@ -22,7 +27,13 @@ export function findKey(key, obj) {
   }
 } // findKey()
 
-// checks whether note noteId is inside a chord. Returns false or the chord id.
+/**
+ * SVG search: checks whether note noteId is inside a chord.
+ * Returns false or the chord id.
+ * @param {string} noteId
+ * @param {string} what
+ * @returns {string|boolean}
+ */
 export function insideParent(noteId, what = 'chord') {
   if (noteId) {
     let note = document.querySelector('g#' + escapeXmlId(noteId));
@@ -33,11 +44,25 @@ export function insideParent(noteId, what = 'chord') {
   return false;
 } // insideParent()
 
-// finds all chords/notes inside an element and return array of id strings
+/**
+ * SVG search: finds all chords/notes inside an element
+ * and returns array of id strings
+ * @param {string} elId
+ * @returns {string}
+ */
 export function findNotes(elId) {
   let idArray = [];
   if (elId) {
-    let notes = document.querySelector('g#' + escapeXmlId(elId)).querySelectorAll('.note');
+    let notes = Array.from(document.querySelector('g#' + escapeXmlId(elId)).querySelectorAll('.note'));
+    // if no child notes, but a plist attribute such as in beamSpan, go through those elements
+    if (notes.length <= 0) {
+      let el = document.querySelector('[*|id="' + elId + '"]');
+      if (el && el.hasAttribute('data-plist')) {
+        el.getAttribute('data-plist')
+          .split(' #')
+          .map((e) => notes.push(document.querySelector('[*|id="' + rmHash(e) + '"]')));
+      }
+    }
     for (let note of notes) {
       let noteId = note.getAttribute('id');
       let chordId = insideParent(noteId);
@@ -51,8 +76,16 @@ export function findNotes(elId) {
   return idArray;
 } // findNotes()
 
-// look for elementname (e.g., 'staff') upwards in the xml file and return
-// attribute value (searchString defaults to the "@n" attribute).
+/**
+ * CodeMirror search: look for elementname (e.g., 'staff') upwards
+ * in the xml file and return attribute value (searchString defaults
+ * to the "@n" attribute).
+ * @param {CodeMirror} cm
+ * @param {number} row
+ * @param {string} elementName
+ * @param {RegExp} searchString
+ * @returns
+ */
 export function getElementAttributeAbove(
   cm,
   row,
@@ -69,8 +102,16 @@ export function getElementAttributeAbove(
   return [null, null];
 } // getElementAttributeAbove()
 
-// look for elementname (e.g., 'staff') downwards in the xml file and return
-// attribute value (searchString defaults to the "@n" attribute).
+/**
+ * CodeMirror search:  look for elementname (e.g., 'staff') downwards
+ * in the xml file and return attribute value (searchString defaults
+ * to the "@n" attribute).
+ * @param {CodeMirror} cm
+ * @param {number} row
+ * @param {string} elementName
[…]
