### 2023-09-26 c35348d
Filter input for insterting control elements #48
