+++ app/static/lib/speed.js
@@ -1008,7 +1008,7 @@ function addKeySigElement(staffDefs, keysigValue) {
     let k = staffDef.querySelector('keySig');
     if (k) {
       k.remove();
-    } 
+    }
     staffDef.setAttribute('key.sig', keysigValue);
   }
 } // addKeySigElement()
@@ -1216,14 +1216,25 @@ export function countStaves(scoreDef) {
 } // countStaves()
 
 /**
- * Filter selected elements and keep only highest in DOM
+ * Filter selected elements and keep only highest in DOM.
+ * If keepOnly array of strings is provided (such as ['note','chord','rest']),
+ * keep only those elements and ignore others.
+ *
  * @param {string[]} ids is modified by this function
  * @param {Document} xmlDoc
+ * @param {string[]} keepOnly
  * @returns {string[]} The modified `ids` array, with all elements removed that
  * have an ancestor that is also in the `ids` array.
  */
-export function filterElements(ids, xmlDoc) {
+export function filterElements(ids, xmlDoc, keepOnly = []) {
   for (let i = 0; i < ids.length; i++) {
+    let eli = xmlDoc.querySelector('[*|id="' + ids[i] + '"]');
+    if ((eli && keepOnly.length > 0 && !keepOnly.includes(eli.nodeName)) || !eli) {
+      // Remove elements in the list, or when not in encoding (such as @accid)
+      console.debug('Speed.filterElements: removed: ', eli ? eli : ids[i]);
+      ids.splice(i--, 1);
+      continue;
+    }
     for (let j = i + 1; j < ids.length; j++) {
       const elj = xmlDoc.querySelector('[*|id="' + ids[j] + '"]');
       if (!elj) continue;
