### 2023-09-26 fe65f59
ScoreDef@meter.sym read (addresses #49)
