+++ app/static/lib/viewer.js
@@ -673,10 +673,13 @@ export default class Viewer {
     } else if (env === environments.testing) {
       owlSrc += 'testing-';
     }
+    let splashOwl = document.getElementById('splashLogo');
+    let splashOwlSrc = splashOwl.getAttribute('src'); // splash logo does not adjust with environment
     if (j < 128) {
       // dark
       // wake up owl
       owlSrc += 'menu-logo' + (isSafari ? '.png' : '.svg');
+      splashOwlSrc += 'menu-logo' + (isSafari ? '.png' : '.svg');
       els.forEach((el) => el.style.setProperty('filter', 'invert(.8)'));
       rt.style.setProperty('--settingsLinkBackgroundColor', utils.brighter(cm.backgroundColor, 21));
       rt.style.setProperty('--settingsLinkHoverColor', utils.brighter(cm.backgroundColor, 36));
@@ -719,6 +722,7 @@ export default class Viewer {
       // bright mode
       // sleepy owl
       owlSrc += 'menu-logo-asleep' + (isSafari ? '.png' : '.svg');
+      splashOwlSrc += 'menu-logo-asleep' + (isSafari ? '.png' : '.svg');
       els.forEach((el) => el.style.removeProperty('filter'));
       rt.style.setProperty('--settingsLinkBackgroundColor', utils.brighter(cm.backgroundColor, -16));
       rt.style.setProperty('--settingsLinkHoverColor', utils.brighter(cm.backgroundColor, -24));
