+++ app/static/lang/lang.it.js
@@ -30,7 +30,7 @@ export const lang = {
   openUrlText: { text: 'Apri URL' },
   openExample: {
     text: 'Repertorio pubblico',
-    description: 'Apri una lista del repertorio pubblico',
+    description: 'Apri un repertorio di pubblico dominio',
   },
   importMusicXml: { text: 'Importa MusicXML' },
   importHumdrum: { text: 'Importa Humdrum' },
@@ -78,14 +78,14 @@ export const lang = {
   betweenPlacementText: { text: 'Posiziona in mezzo' },
   addVerticalGroupText: { text: 'Aggiungi gruppo verticale' },
   deleteText: { text: 'Elimina elemento' },
-  pitchChromUpText: { text: 'Alza cromatica' },
-  pitchChromDownText: { text: 'Abbassa cromatica' },
-  pitchUpDiatText: { text: 'Alza diatonica' },
-  pitchDownDiatText: { text: 'Abbassa diatonica' },
+  pitchChromUpText: { text: 'Alza cromaticamente' },
+  pitchChromDownText: { text: 'Abbassa cromaticamente' },
+  pitchUpDiatText: { text: 'Alza diatonicamente' },
+  pitchDownDiatText: { text: 'Abbassa diatonicamente' },
   pitchOctaveUpText: { text: "Alza di un'ottava" },
   pitchOctaveDownText: { text: "Abbassa di un'ottava" },
-  staffUpText: { text: 'Elemento su 1 rigo' },
-  staffDownText: { text: 'Elemento giù 1 rigo' },
+  staffUpText: { text: 'Alza di un rigo' },
+  staffDownText: { text: 'Abbassa di un rigo' },
   increaseDurText: { text: 'Aumenta durata' },
   decreaseDurText: { text: 'Riduci durata' },
   cleanAccidText: { text: 'Verificare @accid.ges' },
@@ -146,10 +146,10 @@ export const lang = {
   // HELP MENU ITEM / VOCE DI MENU AIUTO
   helpMenuTitle: { text: 'Aiuto' },
   goToHelpPage: { text: 'Pagina di aiuto di mei-friend' },
-  goToCheatSheet: { text: 'Scheda informativa di mei-friend' },
+  goToCheatSheet: { text: 'Documentatione di mei-friend' },
   showChangelog: { text: 'Registro delle modifiche di mei-friend' },
   goToGuidelines: { text: 'Linee guida MEI' },
-  consultGuidelinesForElementText: { text: "Linee guida per l'elemento corrente" },
+  consultGuidelinesForElementText: { text: "Informazioni dell'elemento corrente" },
   provideFeedback: { text: 'Fornisci un feedback' },
   resetDefault: { text: 'Ripristina impostazioni predefinite' },
 
@@ -169,12 +169,12 @@ export const lang = {
       '<a href="https://iwk.mdw.ac.at/impressum">Imprint</a>.',
   },
   loadingVerovio: { text: 'Caricamento Verovio' },
-  verovioLoaded: { text: 'caricato' },
-  convertedToPdf: { text: 'convertito in PDF' },
+  verovioLoaded: { text: 'Caricato' },
+  convertedToPdf: { text: 'Convertito in PDF' },
   statusBarCompute: { text: 'Calcola' },
-  middleFooterPage: { text: 'pagina' },
-  middleFooterOf: { text: 'di' },
-  middleFooterLoaded: { text: 'caricato' },
+  middleFooterPage: { text: 'Pagina' },
+  middleFooterOf: { text: 'Di' },
+  middleFooterLoaded: { text: 'Caricato' },
 
   // control menu/ menu di controllo
   verovioIcon: {
@@ -184,10 +184,10 @@ export const lang = {
   },
[…]
