### 2023-09-26 a661b9d
Rename zoom function names
