### 2023-09-27 7bad407
Finish implementing splash screen
