### 2023-09-27 6dcc498
Adapt language packs for menu item change
