### 2023-09-27 f3fe302
Fix and add keymap to editor panel (#50)
