### 2023-09-27 ec61692
Improve positioning, default showSplash to false
