### 2023-09-27 91669a8
Ensure splash screen shown on first ever load
