### 2023-09-27 2f90b17
hide splash on esc or background click
