+++ app/static/lib/main.js
@@ -612,8 +612,7 @@ function onLanguageLoaded() {
   }
 
   // show splash screen if required
-  if((storage.supported && !storage.splashAcknowledged) || 
-    document.getElementById('showSplashScreen').checked) {
+  if ((storage.supported && !storage.splashAcknowledged) || document.getElementById('showSplashScreen').checked) {
     showSplashScreen();
   }
 
@@ -1327,7 +1326,7 @@ function downloadSpeedMei() {
 function showSplashScreen() {
   document.getElementById('splashOverlay').style.display = 'block';
   document.getElementById('splashAlwaysShow').checked = document.getElementById('showSplashScreen').checked;
-}  
+}
 
 function togglePdfMode() {
   console.log('Toggle PDF mode');
@@ -1663,6 +1662,10 @@ export let cmd = {
   openHelp: () => window.open(`./help`, '_blank'),
   consultGuidelines: () => consultGuidelines(),
   escapeKeyPressed: () => {
+    // hide overlays
+    // TODO refactor logic for all overlays below. For now only splash overlay...
+    document.getElementById('splashOverlay').style.display = 'none';
+
     // reset settings filter, if settings have focus
     if (
       document.getElementById('settingsPanel') &&
@@ -2027,7 +2030,10 @@ function addEventListeners(v, cm) {
   document.getElementById('toggleSpicc').addEventListener('click', cmd.toggleSpicc);
 
   // show splash screen
-  document.getElementById('aboutMeiFriend').addEventListener('click', showSplashScreen)
+  document.getElementById('aboutMeiFriend').addEventListener('click', showSplashScreen);
+  document.getElementById('splashOverlay').addEventListener('click', (e) => {
+    if (e.target.id === 'splashOverlay') document.getElementById('splashOverlay').style.display = "none"; // dismiss splash when user clicks on black background
+  });
 
   // consult guidelines
   document.getElementById('consultGuidelinesForElement').addEventListener('click', cmd.consultGuidelines);
@@ -2210,14 +2216,15 @@ export function drawRightFooter() {
     }
   }
   let rf = document.querySelector('.rightfoot');
-  const versionHtml = "<a href='https://github.com/mei-friend/mei-friend' target='_blank'>mei-friend " +
+  const versionHtml =
+    "<a href='https://github.com/mei-friend/mei-friend' target='_blank'>mei-friend " +
     (env === environments.production ? version : `${env}-${version}`) +
     '</a> (' +
     translatedVersioDate +
     ').&nbsp;';
   rf.innerHTML = versionHtml;
   // also update version string in splash screen
-  document.getElementById("splashVersionNumber").innerHTML = versionHtml;
+  document.getElementById('splashVersionNumber').innerHTML = versionHtml;
   if (tkVersion) {
     let githubUrl = 'https://github.com/rism-digital/verovio/releases/tag/version-' + tkVersion.split('-')[0];
     if (tkVersion.includes('dev')) {
