### 2023-09-27 0671065
Add alert at clicking/drag-sel element w/o id #10
