### 2023-09-27 3e0f746
Internationalize splash screen for EN and DE
