+++ app/static/lang/lang.de.js
@@ -6,6 +6,71 @@ import * as att from '../lib/attribute-classes.js';
 import { heart } from '../css/icons.js';
 
 export const lang = {
+  // Splash screen
+  splashBody: {
+    html: `
+    <p>
+      mei-friend ist ein Online-Editor für <a href="https://music-encoding.org">Musik-Codierungen</a>, gehostet an der
+      <a href="https://mdw.ac.at" target="_blank">mdw &ndash; Universität für Musik und darstellende Kunst Wien</a>. 
+      Bitte konsultieren Sie unsere
+      <a href="https://mei-friend.github.io" target="_blank">umfangreiche Dokumentation</a> für weitere
+      Informationen.
+    </p>
+    <p>
+      Obwohl mei-friend eine browserbasierte Anwendung ist, werden keine Ihrer persönlichen Daten (einschließlich der 
+      von Ihnen bearbeiteten Codierung, Ihrer Anwendungseinstellungen und Ihrer aktuellen Anmeldeinformationen, falls 
+      vorhanden) auf unseren Servern gespeichert. Stattdessen werden sie im 
+      <a href="https://developer.mozilla.org/de/docs/Web/API/Window/localStorage" target="_blank">lokalen Speicher 
+        Ihres Browsers gespeichert</a>.
+    </p>
+    <p>
+      Daten werden nur dann an GitHub übertragen, wenn Sie dies ausdrücklich anfordern (z. B. wenn Sie sich bei GitHub 
+      anmelden, Ihre Codierung aus einem GitHub-Repository laden oder an ein GitHub-Repository übertragen oder wenn 
+      Sie eine GitHub-Action-Workflow für Sie ausführen lassen). Ebenso werden Daten nur dann an Ihren ausgewählten 
+      Solid-Anbieter übertragen, wenn Sie dies ausdrücklich anfordern (z.B. wenn Sie sich bei Solid anmelden oder 
+      Stand-off-Anmerkungen laden oder speichern).
+    </p>
+    <p>
+      Wir verwenden <a href="https://matomo.org/" target="_blank">Matomo</a> zur Erfassung anonymer Nutzungsstatistiken. 
+      Dazu gehören Ihre gekürzte IP-Adresse (die Geolokalisierung auf Länderebene, aber keine weitere Identifizierung 
+      ermöglicht), Ihr Browser und Ihr Betriebssystem, von wo aus Sie gekommen sind (d.h. die verweisende Website), 
+      die Uhrzeit und Dauer Ihres Besuchs und die von Ihnen besuchten Seiten. Diese Informationen werden auf der 
+      Matomo-Instanz gespeichert, die auf den Servern der mdw &ndash; Universität für Musik und darstellende Kunst 
+      Wien läuft, und werden nicht an Dritte weitergegeben.
+    </p>
+    <p>
+      Schließlich wird die MIDI-Wiedergabe mit dem von Google Magenta bereitgestellten SGM_plus-Soundfont präsentiert 
+      und über googleapis.com bereitgestellt. Ihre IP-Adresse wird dabei automatisch an Google übertragen. Wenn Sie 
+      dies nicht möchten, verzichten Sie bitte auf die Verwendung der MIDI-Wiedergabefunktion.
+    </p>
+    <p>
+      Mei-Friend wurde entwickelt von
+      <a href="https://iwk.mdw.ac.at/werner-goebl" target="_blank">Werner Goebl</a> und
+      <a href="https://iwk.mdw.ac.at/david-weigl" target="_blank">David M. Weigl</a> am Institut für Musikakustik &ndash; 
[…]
