### 2023-09-27 776505d
Complete splash-screen internationalisation
