### 2023-09-28 271f4de
Consistently translate encoding as Kodierung in DE
