### 2023-09-28 b52a676
DE splash improvements
