### 2023-09-28 6b24faa
Ensure prompt loading of splash screen text
