### 2023-09-28 cdae1d7
Update splash screen texts in EN and DE
