### 2023-09-28 3ffa2aa
Update splash screen texts for all languages
