### 2023-09-28 9c68eb6
CSS fixes for splash screen
