### 2023-09-29 d79865d
Update lang.it.js
up to line 250 and pannello to finestra for annotazioni e spartito
