+++ app/static/lang/lang.it.js
@@ -64,8 +64,8 @@ export const lang = {
   notationLeft: { text: 'Notazione a sinistra' },
   notationRight: { text: 'Notazione a destra' },
   showSettingsMenuText: { text: 'Pannello di impostazioni' },
-  showAnnotationMenuText: { text: 'Pannello di annotazioni' },
-  showFacsimileMenuText: { text: 'Pannello di facsimili' },
+  showAnnotationMenuText: { text: 'Finestra di annotazioni' },
+  showFacsimileMenuText: { text: 'Finestra di facsimili' },
   showPlaybackControlsText: { text: 'Controlli di riproduzione' },
   facsimileTop: { text: 'Facsimile in alto' },
   facsimileBottom: { text: 'Facsimile in basso' },
@@ -155,8 +155,8 @@ export const lang = {
 
   // panel icons/ icone pannello
   showMidiPlaybackControlBarButton: { description: 'Attiva/Disattiva barra di controllo della riproduzione MIDI' },
-  showFacsimileButton: { description: 'Attiva/Disattiva pannello Facsimile' },
-  showAnnotationsButton: { description: 'Attiva/Disattiva pannello Annotazioni' },
+  showFacsimileButton: { description: 'Attiva/Disattiva la finestra di Facsimile' },
+  showAnnotationsButton: { description: 'Attiva/Disattiva la finestra Annotazioni' },
   showSettingsButton: { description: 'Mostra il pannello delle impostazioni' },
 
   // Footer texts/ Testi di piè di pagina
@@ -223,7 +223,7 @@ export const lang = {
 
   // PDF/print preview panel / Pannello di anteprima PDF/stampa
   pdfSaveButton: { text: 'Salva PDF', description: 'Salva come PDF' },
-  pdfCloseButton: { description: 'Chiudi vista di stampa' },
+  pdfCloseButton: { description: 'Chiudi anteprima di stampa' },
   pagesLegendLabel: { text: 'Intervallo di pagine', singlePage: 'pagina', multiplePages: 'Pagine' },
   selectAllPagesLabel: { text: 'Tutte' },
   selectCurrentPageLabel: { text: 'Pagina corrente' },
@@ -232,23 +232,23 @@ export const lang = {
   selectPageRangeLabel: { text: 'Intervallo di pagine:' },
   pdfPreviewSpeedModeWarning: {
     text:
-      'Solamente la pagina corrente viene renderizzata in PDF, perché è attivata la modalità veloce. ' +
+      'In modalità veloce solamente la pagina corrente viene renderizzata in PDF.' +
       'Deseleziona la modalità veloce per selezionare tutte le pagine.',
   },
   pdfPreviewNormalModeTitle: { text: "Seleziona l'intervallo di pagine da salvare in PDF." },
 
-  // facsimile panel/ pannello facsimile
-  facsimileIcon: { description: 'Pannello facsimile' },
-  facsimileDecreaseZoomButton: { description: "Riduci il pannello dello spartito" },
-  facsimileZoom: { description: "Regola la dimensione del pannello dello spartito" },
-  facsimileIncreaseZoomButton: { description: "Ingrandisci il pannello dello spartito" },
+  // facsimile panel/ finestra facsimile
+  facsimileIcon: { description: 'Finestra facsimile' },
+  facsimileDecreaseZoomButton: { description: "Riduci la finestra dello spartito" },
+  facsimileZoom: { description: "Regola la dimensione la finestra dello spartito" },
+  facsimileIncreaseZoomButton: { description: "Ingrandisci la finestra dello spartito" },
   facsimileFullPageLabel: {
     text: 'Pagina intera',
-    description: "Mostra il pannello del facsimile a pagina intera",
+    description: "Mostra la finestra del facsimile a pagina intera",
   },
[…]
