### 2023-09-29 ee9112b
Update lang.it.js
until line 500
