### 2023-10-02 14ad35f
Polish keyboard shortcuts, bind to body
