+++ app/static/lib/main.js
@@ -1,6 +1,6 @@
 // mei-friend version and date
-export const version = '1.0.3';
-export const versionDate = '28 September 2023'; // use full or 3-character english months, will be translated
+export const version = '1.0.2';
+export const versionDate = '2 October 2023'; // use full or 3-character english months, will be translated
 
 var vrvWorker;
 var spdWorker;
@@ -142,23 +142,6 @@ const defaultCodeMirrorOptions = {
     "' '": completeIfInTag,
     "'='": completeIfInTag,
     'Ctrl-Space': 'autocomplete',
-    // 'Alt-.': consultGuidelines,
-    "'Ï'": indentSelection, // TODO: overcome strange bindings on MAC Shift-Alt-F
-    'Shift-Alt-f': indentSelection,
-    'Shift-Ctrl-G': toMatchingTag,
-    'Cmd-E': encloseSelectionWithTag, // TODO: make OS modifier keys dynamic
-    'Ctrl-E': encloseSelectionWithTag,
-    'Cmd-/': encloseSelectionWithLastTag,
-    'Ctrl-/': encloseSelectionWithLastTag,
-    // 'Cmd-,': cmd.toggleSettingsPanel,
-    // 'Cmd-L': generateUrl,
-    // 'Ctrl-L': generateUrl,
-    // 'Cmd-O': openFileDialog,
-    // 'Ctrl-O': openFileDialog,
-    // 'Cmd-P': togglePdfMode,
-    // 'Ctrl-P': togglePdfMode,
-    // 'Cmd-S': downloadMei,
-    // 'Ctrl-S': downloadMei,
   },
   lint: {
     caller: cm,
@@ -1416,17 +1399,7 @@ function consultGuidelines() {
       }
     }
   }
-}
-
-// wrapper for indentSelection to be called inside CodeMirror
-function indentSelection() {
-  e.indentSelection(v, cm);
-} // indentSelection()
-
-// wrapper for toMatchingTag
-function toMatchingTag() {
-  e.toMatchingTag(v, cm);
-}
+} // consultGuidelines()
 
 let tagEncloserNode; // context menu to choose node name to enclose selected text
 
@@ -1669,6 +1642,10 @@ export let cmd = {
   removeIds: () => e.manipulateXmlIds(v, cm, true),
   ingestFacsimile: () => ingestFacsimile(),
   addFacsimile: () => e.addFacsimile(v, cm),
+  encloseSelectionWithTag: encloseSelectionWithTag,
+  encloseSelectionWithLastTag: encloseSelectionWithLastTag,
+  toMatchingTag: () => e.toMatchingTag(v, cm),
+  indentSelection: () => e.indentSelection(v, cm),
   resetDefault: () => {
     // we're in a clickhandler, so our storage object is out of scope
     // but we only need to clear it, so just grab the window's storage
@@ -2341,22 +2318,20 @@ function fillInSampleEncodings() {
 }
 
 /**
- * Sets keymap JSON information to target element and defines listeners
- * It loads all bindings in `#notation` to notation, and the platform-specific
+ * Sets keymap JSON information to target element and defines listeners.
+ * It loads all bindings in `#notation` to document.body, and the platform-specific
  * to both notation and editor panels (i.e. friendContainer).
  * @param {string} keyMapFilePath
  */
 function setKeyMap(keyMapFilePath) {
-  // let keyMapParent = document.getElementById('notation');
-  let keyMapParent = document.getElementById('friendContainer');
   if (platform.startsWith('mac')) {
-    keyMapParent.classList.add('platform-darwin');
+    document.body.classList.add('platform-darwin');
   }
   if (platform.startsWith('win')) {
-    keyMapParent.classList.add('platform-win32');
[…]
