+++ CHANGELOG.md
@@ -1,8 +1,5 @@
 # mei-friend-online CHANGELOG.md
 
-### 1.0.3 patch
-*
-
 ### 1.0.2 Splash screen 
 * Add splash screen on application load explaining data storage (#41)
 * Fix bug when drag-selected elements contain non-note elements and blocked inserting control elements on them (such as slurs) (#48)
