### 2023-10-02 3f9eb10
Remove unused code
