+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Support for @meter.sym/@sym information (#49). No slur (spanning element) inserted when on same time stamps.
 * Support keyboard shortcuts (with os-dependent modifier keys) in encoding editor (#50)
 * Show warning message when clicking on or drag-selecting elements without xml:ids (#10)
+* Drag'n'drop on CodeMirror editor now shows drag overlay screen
 ### 1.0.1 patch
 * Reduce mei-friend-specific Verovio defaults
 * Add auto-translated draft of Japanese language pack
