### 2023-10-03 d3f0f89
Handle splash screen and Safari warning
