### 2023-10-03 7ee0fdd
Tighten splashscreen behaviour for compliance
