+++ app/static/lib/main.js
@@ -409,6 +409,21 @@ function onLanguageLoaded() {
   }
   // build language selection menu
   buildLanguageSelection();
+
+  // show splash screen if required
+  if (storage.supported) {
+    storage.read();
+    if (!storage.splashAcknowledged || storage.showSplashScreen) {
+      showSplashScreen(true);
+    } else {
+      completeInitialLoad();
+    }
+  } else {
+    completeInitialLoad();
+  }
+} // onLanguageLoaded()
+
+function completeInitialLoad() {
   // link to changelog page according to env settings (develop/staging/production)
   let changeLogUrl;
   switch (env) {
@@ -523,7 +538,6 @@ function onLanguageLoaded() {
     : 'none';
 
   if (storage.supported) {
-    storage.read();
     if (storage.github) {
       // use github object from local storage if available
       isLoggedIn = true;
@@ -562,7 +576,6 @@ function onLanguageLoaded() {
     storage.safelySetStorageItem('fileLocation', url.href);
     storage.safelySetStorageItem('fileName', url.pathname.substring(url.pathname.lastIndexOf('/') + 1));
     storage.safelySetStorageItem('fileLocationType', 'url');
-    storage.read();
     console.log('Have set local storage: ', storage);
   }
 
@@ -602,11 +615,6 @@ function onLanguageLoaded() {
     urlFetchInProgress = true;
   }
 
-  // show splash screen if required
-  if ((storage.supported && !storage.splashAcknowledged) || document.getElementById('showSplashScreen').checked) {
-    showSplashScreen();
-  }
-
   // fill sample encodings
   fillInSampleEncodings();
 
@@ -615,7 +623,6 @@ function onLanguageLoaded() {
 
   // restore localStorage if we have it
   if (storage.supported) {
-    storage.read();
     // save (most) URL parameters in storage
     if (orientationParam !== null) storage.notationOrientation = orientationParam;
     if (notationProportionParam !== null) storage.notationProportion = notationProportionParam;
@@ -769,7 +776,7 @@ function onLanguageLoaded() {
       loginAndFetch(getSolidIdP(), populateSolidTab);
     }, restoreSolidTimeoutDelay);
   }
-} // onLanguageLoaded
+} // completeInitialLoad()
 
 export async function openUrlFetch(url = '', updateAfterLoading = true) {
   let urlInput = document.querySelector('#openUrlInput');
@@ -1314,9 +1321,14 @@ function downloadSpeedMei() {
   }, 0);
 } // downloadSpeedMei()
 
-function showSplashScreen() {
+function showSplashScreen(initialLoad = false) {
   document.getElementById('splashOverlay').style.display = 'flex';
-  document.getElementById('splashAlwaysShow').checked = document.getElementById('showSplashScreen').checked;
+  document.getElementById('splashAlwaysShow').checked = storage.showSplashScreen;
+  document.getElementById('splashConfirmButton').addEventListener('click', () => {
+    document.getElementById('splashOverlay').style.display = 'none';
+    window.localStorage.setItem('splashAcknowledged', 'true');
+    if (initialLoad) completeInitialLoad();
+  });
 }
 
 function togglePdfMode() {
