### 2023-10-03 768d957
Translate all browser exceptions
