### 2023-10-03 b48e649
Avoid reinitialising app on splash button click
