### 2023-10-03 60605cc
Fix alwaysShowSplash behaviour
