### 2023-10-03 a5b3f6a
Explain using Verovio in splash text in 8 langs
