### 2023-10-06 db5974a
Bump to 1.0.4 and update changelog & date
