### 2023-10-06 fe9edfc
Start drafting enrichtment_panel.js
