### 2023-10-06 510ca4c
Review annotations.js
