### 2023-10-09 36e28a6
Add function to spin the icon when loading web annotations
