+++ app/static/lang/lang.eo.js
@@ -1,0 +1,975 @@
+/**
+ * Language file for Esperanto.
+ */
+
+import * as att from '../lib/attribute-classes.js';
+import { heart } from '../css/icons.js';
+
+export const lang = {
+  // Splash screen
+  aboutMeiFriend: { text: 'Pri mei-friend ' },
+  showSplashScreen: {
+    text: 'Montri komenca ekrano ĉe ŝarĝo',
+    description: 'Montru la mei-friend komenca ekrano kiam la aplikaĵo estas ŝargita',
+  },
+  splashBody: {
+    html: `
+    <p>
+      mei-friend estas redaktanto por <a href="https://music-encoding.org">muzikaj kodadoj</a>, gastigita ĉe
+      <a href="https://mdw.ac.at" target="_blank">mdw &ndash; Universitato de Muziko kaj Artaj Artoj de Vieno</a>. 
+      Bonvolu konsulti nian <a href="https://mei-friend.github.io" target="_blank">ampleksan dokumentadon</a> por 
+      pliaj informoj.
+    </p>
+    <p>
+      Kvankam mei-friend estas retumila aplikaĵo, viaj personaĵaj datumoj (inkluzive la kodo kiun
+      vi redaktas, viaj aplikaĵaj agordoj, kaj aktualegaj ensalutinformoj se ili ekzistas) estas konservitaj en via retumila navigilo en
+      <a href="https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage" target="_blank"
+        >lokal memoro</a
+      > kaj ne estas senditaj al aŭ konservitaj sur niaj serviloj.
+    </p>
+    <p>
+      Datumoj estas senditaj al GitHub nur kiam vi eksplice petas tion (ekzemple, kiam vi ensalutas al GitHub, ŝargas
+      vian kodon de aŭ al GitHub-repozitorio, aŭ kiam vi petas ke GitHub-agonaj laborfluo estu ekigita por vi). Sammaniere, datumoj estas senditaj al via elektita Solid provizanto nur kiam vi eksplice petas tion (ekzemple, kiam vi ensalutas al Solid, aŭ ŝargas aŭ konservas standoff-anotaciojn).
+    </p>
+    <p>
+      Ni uzas <a href="https://matomo.org/" target="_blank">Matomo</a>
+      por kolekti anonimajn uzo-statistikojn. Tio inkluzivas vian mallongan IP-adreson (permesanta geolokigon ĝis landnivelo, sed neniun plian identigon), vian foliumilon kaj operaciumon, de kie vi venis (t.e., la referanta retejo), la tempon kaj daŭron de via vizito, kaj la paĝojn kiujn vi vizitis. Tiu informo estas konservita en la Matomo-okaziga servilo situanta sur serviloj de mdw &ndash; Universitato de Muziko kaj Artaj Artoj de Vieno, kaj ne estas dividata kun triaj partioj.
+    </p>
+    <p>
+      La Verovio-ilareto estas ŝargita el <a href="https://verovio.org" target="_blank">https://verovio.org</a>, gastigata de
+      <a href="https://rism.digital/" target="_blank">RISM Digital Switzerland</a>. 
+      Tio ebligas al mei-friend resti aktualigita kun la plej nova ilaretoversio kaj
+      provizi elekton de ĉiuj subtenataj versioj tra la agordaro. 
+      Kiam uzante mei-friend, do via IP-adreso estas videbla por RISM Digital.
+    </p>
+    <p>
[…]
