+++ app/static/lang/lang.uk.js
@@ -1,0 +1,968 @@
+/**
+ * Language file for Ukranian
+ */
+
+import * as att from '../lib/attribute-classes.js';
+import { heart } from '../css/icons.js';
+
+export const lang = {
+  // Splash screen
+  aboutMeiFriend: { text: 'Про mei-friend ' },
+  showSplashScreen: {
+    text: 'Показувати початковий екран при завантаженні',
+    description: 'Показувати початковий екран mei-friend при завантаженні додатку',
+  },
+  splashBody: {
+    html: `
+    <p>
+      mei-friend - це редактор для <a href="https://music-encoding.org">музичних кодувань</a>, розміщений на
+      <a href="https://mdw.ac.at" target="_blank">mdw &ndash; Університеті музики та виконавських мистецтв у Відні</a>. 
+      Будь ласка, ознайомтеся з нашою <a href="https://mei-friend.github.io" target="_blank">розширеною документацією</a> для 
+      отримання додаткової інформації.
+    </p>
+    <p>
+      Незважаючи на те, що mei-friend є веб-додатком, ваші особисті дані (включаючи кодування, яке ви редагуєте, налаштування додатку та поточні дані входу, якщо такі є) зберігаються в локальному сховищі вашого веб-переглядача
+      <a href="https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage" target="_blank"
+        >local storage</a
+      > і не передаються або не зберігаються на наших серверах.
+    </p>
+    <p>
+      Дані передаються на GitHub лише тоді, коли ви це явно запитуєте (наприклад, коли ви входите в систему на GitHub, завантажуєте ваше кодування з репозиторію GitHub або фіксуєте його, або коли ви запитуєте виконання робочого процесу GitHub Action). Точно так само дані передаються на обраний вами провайдер Solid лише тоді, коли ви це явно запитуєте (наприклад, коли ви входите в систему Solid або завантажуєте або зберігаєте анотації станд-офу).
+    </p>
+    <p>
+      Ми використовуємо <a href="https://matomo.org/" target="_blank">Matomo</a>
+      для збору анонімних статистичних даних про використання. До них належать скорочена IP-адреса (дозволяє визначити геолокацію лише на рівні країни, але не дозволяє вище), ваш веб-переглядач та операційна система, з якої ви прийшли (тобто посилаючий сайт), час і тривалість вашого візиту та сторінки, які ви відвідали. Ця інформація зберігається на інсталяції Matomo, яка працює на серверах mdw &ndash; Університет музики та виконавських мистецтв у Відні, і не передається жодній третій стороні.
+    </p>
+    <p>
+      Набір інструментів Verovio завантажується зі сторінки <a href="https://verovio.org" target="_blank">https://verovio.org</a>, розміщеної на
+      <a href="https://rism.digital/" target="_blank">RISM Digital Switzerland</a>. 
+      Це дозволяє mei-friend бути в актуальному стані з останньою версією і
+      надавати вибір всіх підтримуваних версій через панель налаштувань. 
+      Під час використання mei-friend ваша IP-адреса тому видима для RISM Digital.
+    </p>
+    <p>
[…]

[…]
