### 2023-10-10 #5863 @faroit
👋  Hi @wergo, thanks for your submission to JOSS! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues.

Thank you already for providing a list of reviewers with domain knowledge. I would also like to ask Stefan Balke (@ stefan-balke). Can you please check if this would be fine wrt to the JOSS [COI policy](<https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy>)?
