### 2023-10-11 161246d
Start refactoring of annotation.js into enrichment_panel.js
