+++ app/static/lib/annotation.js
@@ -1,7 +1,7 @@
 import { v, cm, log, translator, setStandoffAnnotationEnabledStatus } from './main.js';
 import { convertCoords, generateXmlId, rmHash, setCursorToId } from './utils.js';
 import { meiNameSpace, xmlNameSpace, xmlToString } from './dom-utils.js';
-import {
+/*import {
   circle,
   diffRemoved,
   highlight,
@@ -12,7 +12,7 @@ import {
   rdf,
   speechBubble,
   symLinkFile,
-} from '../css/icons.js';
+} from '../css/icons.js';*/
 import { removeInEditor } from './editor.js';
 import {
   loginAndFetch,
@@ -30,11 +30,12 @@ import {
   safelyPatchResource,
 } from './solid.js';
 import { nsp, traverseAndFetch } from './linked-data.js';
+import { deleteAnnotation } from './enrichment_panel.js';
 
-export let annotations = [];
+//export let annotations = [];
 
 // to enrichment_panel.js
-export function situateAndRefreshAnnotationsList(forceRefresh = false) {
+/*export function situateAndRefreshAnnotationsList(forceRefresh = false) {
   situateAnnotations();
   if (forceRefresh || !document.getElementsByClassName('annotationListItem').length) refreshAnnotationsList();
 }
@@ -169,10 +170,10 @@ export function refreshAnnotationsList() {
     annoDiv.appendChild(annoListItemButtons);
     list.appendChild(annoDiv);
   });
-}
+}*/
 
 // to enrichment_panel.js
-export function generateAnnotationLocationLabel(a) {
+/*export function generateAnnotationLocationLabel(a) {
   console.log('generating anno label for ', a);
   const annotationLocationLabel = document.createElement('span');
   if (a.type === 'annotateIdentify') {
@@ -196,58 +197,11 @@ export function generateAnnotationLocationLabel(a) {
   annotationLocationLabel.classList.add('annotationLocationLabel');
   annotationLocationLabel.dataset.id = 'loc-' + a.id;
   return annotationLocationLabel;
-}
+}*/
 
-/**
- * Adds page locations in rendering to annotation object in annotation list.
- * Call whenever layout reflows to re-situate annotations appropriately.
- */
-export function situateAnnotations() {
-  annotations.forEach(async (a) => {
-    // for each element in a.selection, ask Verovio for the page number
-    // set a.firstPage and a.lastPage to min/max page numbers returned
-    a.firstPage = 'unsituated';
-    a.lastPage = -1;
-    if ('selection' in a) {
-      let selectionStart =
-        a.selection[0].indexOf('#') < 0 ? a.selection[0] : a.selection[0].substr(a.selection[0].indexOf('#') + 1);
-      let selLen = a.selection.length - 1;
-      let selectionEnd =
-        a.selection[selLen].indexOf('#') < 0
-          ? a.selection[selLen]
-          : a.selection[selLen].substr(a.selection[selLen].indexOf('#') + 1);
-      a.firstPage = await v.getPageWithElement(selectionStart);
-      a.lastPage = await v.getPageWithElement(selectionEnd);
-      if (a.firstPage < 0 && v.speedMode) {
-        if (v.xmlDoc.querySelector('[*|id=' + a.selection[0] + ']')?.closest('meiHead')) a.firstPage = 'meiHead';
-        else console.warn('Cannot locate annotation ', a);
-      } else {
-        // if not speed mode, asynchronous return of page numbers after we are finished here
-        const annotationLocationLabelElement = document.querySelector(
-          `.annotationLocationLabel[data-id=loc-${CSS.escape(a.id)}`
-        );
-        if (annotationLocationLabelElement) {
-          annotationLocationLabelElement.innerHTML = generateAnnotationLocationLabel(a).innerHTML;
-        }
[…]
