+++ app/static/lib/enrichment_panel.js
@@ -2,13 +2,115 @@
  * Provides gui elements and functionality for enrichment bar
  */
 
+//import { v, cm, log, translator, setStandoffAnnotationEnabledStatus } from './main.js';
+import { v, cm, translator, setStandoffAnnotationEnabledStatus } from './main.js';
+//import { convertCoords, generateXmlId, rmHash, setCursorToId } from './utils.js';
+import { setCursorToId } from './utils.js';
+//import { meiNameSpace, xmlNameSpace, xmlToString } from './dom-utils.js';
 import * as annot from './annotation.js';
+import {
+  circle,
+  diffRemoved,
+  highlight,
+  fileCode,
+  identify,
+  link,
+  pencil,
+  rdf,
+  speechBubble,
+  symLinkFile,
+} from '../css/icons.js';
+import {
+  loginAndFetch,
+  solid,
+  solidLogout,
+  provider,
+  getSolidStorage,
+  friendContainer,
+  annotationContainer,
+  establishContainerResource,
+  establishResource,
+  createMAOMusicalObject,
+  establishDiscoveryResource,
+  getCurrentFileUri,
+  safelyPatchResource,
+} from './solid.js';
+
+export let listItems = [];
+//export let annotations = [];
 
 /**
  * List of all things tab
  */
 
-export let listItems = [];
+/**
+ * Clear list of annotations
+ */
+export function clearListItems() {
+  listItems = [];
+}
+
+/**
+ * read List Items from XML
+ */
+export function readListItems(flagLimit = false) {
+  annotations = annot.readAnnots(flagLimit);
+  listItems.push(...annotations);
+  refreshAnnotations();
+}
+// TODO: 
+// * Add method to read markup elements 
+
+/**
+ * Draws annotations in Verovio notation panel.
+ * For annotations, the selected elements needs to by highlighted manually by adding a css class.
+ * In the case of markup, this is not necessary because Verovio already adds classes for markup elements to the wrapped elements.
+ * @param {boolean} [forceListRefresh=false]
+ */
+export function refreshAnnotations(forceListRefresh = false) {
+  // clear rendered annotations container
+  const rac = document.getElementById('renderedAnnotationsContainer');
+  rac.innerHTML = '';
+  // reset annotations-containing svg
+  const scoreSvg = document.querySelector('#verovio-panel svg');
+  if (!scoreSvg) return;
+  const annoSvg = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
+  annoSvg.setAttribute('width', scoreSvg.getAttribute('width'));
+  annoSvg.setAttribute('height', scoreSvg.getAttribute('height'));
+  annoSvg.setAttribute('id', 'renderedAnnotationsSvg');
+  annoSvg.setAttribute('xmlns', 'http://www.w3.org/2000/svg');
+  annoSvg.setAttribute('xmlnsXlink', 'http://www.w3.org/1999/xlink');
+  rac.appendChild(annoSvg);
+  if (document.getElementById('showAnnotations')?.checked) {
+    // drawing handlers can draw into renderedAnnotationsSvg if they need to
+    annotations.forEach((a) => {
+      if ('type' in a) {
+        switch (a.type) {
+          case 'annotateIdentify':
+            annot.drawIdentify(a);
+            break;
+          case 'annotateHighlight':
+            annot.drawHighlight(a);
+            break;
+          case 'annotateCircle':
+            annot.drawCircle(a);
+            break;
+          case 'annotateLink':
+            annot.drawLink(a);
+            break;
+          case 'annotateDescribe':
+            annot.drawDescribe(a);
+            break;
+          default:
+            console.warn("Don't have a drawing function for this type of annotation", a);
+        }
+      } else {
[…]
