### 2023-10-12 82517ba
Reformat and reorder enrichment_panel.js
