### 2023-10-12 c89ef62
Rename refreshAnnotations() to refreshAnnotationsInRendering()
