### 2023-10-12 5f6bf9a
Rename readListItems to readListItemsFromXML
