### 2023-10-12 1010814
Redo readAnnots()
