### 2023-10-12 5744b94
Start adding controller functions to access the array of all things
