### 2023-10-12 1fd04a5
First attempt of refactoring before debugging
