### 2023-10-12 cc9760e
Added missing list div to addLoadWebAnnotationButton()
