### 2023-10-12 91fef75
Added missing reference to copyIdToClipboard()
