### 2023-10-12 0d37b7a
Start refactoring creation of UI elements + Prettier
