+++ app/static/lib/enrichment_panel.js
@@ -45,7 +45,7 @@ export function readListItemsFromXML(flagLimit = false) {
 
 /**
  * Looks if an item with itemId exists in listItems
- * @param {string} itemId 
+ * @param {string} itemId
  * @returns {boolean} item is in listItems
  */
 export function isItemInList(itemId) {
@@ -61,7 +61,7 @@ export function isItemInList(itemId) {
  */
 export function addListItem(listItemObject, forceRefreshAnnotations = false) {
   let addedSuccessfully = false;
-  if(!isItemInList(listItemObject.id)) {
+  if (!isItemInList(listItemObject.id)) {
     listItems.push(listItemObject);
     addedSuccessfully = true;
     if (forceRefreshAnnotations === true) refreshAnnotationsInRendering(true);
@@ -85,7 +85,7 @@ export function deleteAnnotation(uuid) {
 
 /**
  * Retrieve a list item by id
- * @param {string} itemId 
+ * @param {string} itemId
  */
 export function retrieveListItem(itemId) {}
 
@@ -224,8 +224,9 @@ export function situateAndRefreshAnnotationsList(forceRefresh = false) {
   if (forceRefresh || !document.getElementsByClassName('annotationListItem').length) refreshAnnotationsList();
 }
 
-// export function refreshAnnotationsList()
-// to enrichment_panel.js
+/**
+ * Creates the list of all things in the enrichment panel
+ */
 export function refreshAnnotationsList() {
   const list = document.getElementById('listAnnotations');
   // clear list
@@ -237,110 +238,7 @@ export function refreshAnnotationsList() {
   // add web annotation button
   addLoadWebAnnotatationButton(list);
   listItems.forEach((a) => {
-    const annoDiv = document.createElement('div');
-    annoDiv.classList.add('annotationListItem');
-    annoDiv.id = a.id;
-    const details = document.createElement('details');
-    details.setAttribute('open', '');
-    const summary = document.createElement('summary');
-    const annoListItemButtons = document.createElement('div');
-    annoListItemButtons.classList.add('annotationListItemButtons');
-    const flipToAnno = document.createElement('a');
-    flipToAnno.insertAdjacentHTML('afterbegin', symLinkFile); //flipToEncoding;
-    flipToAnno.classList.add('flipPageToAnnotationText');
-    flipToAnno.title = translator.lang.flipPageToAnnotationText.description;
-    flipToAnno.classList.add('icon');
-    if (!'selection' in a) flipToAnno.classList.add('disabled');
-    const addObservation = document.createElement('a');
-    addObservation.insertAdjacentHTML('afterbegin', speechBubble);
-    addObservation.classList.add('addObservation');
-    addObservation.title = 'Add observation to extract';
-    addObservation.classList.add('icon');
-    if (a.type !== !'annotateIdentify') {
-      addObservation.classList.add('disabled');
-    }
-    const deleteAnno = document.createElement('a');
-    deleteAnno.classList.add('deleteAnnotation');
-    deleteAnno.insertAdjacentHTML('afterbegin', diffRemoved);
-    deleteAnno.title = translator.lang.deleteAnnotation.description;
-    const isStandoff = document.createElement('a');
-    isStandoff.classList.add('makeStandOffAnnotation');
-    isStandoff.insertAdjacentHTML('afterbegin', rdf);
-    isStandoff.title = translator.lang.makeStandOffAnnotation.description;
-    isStandoff.classList.add('icon');
-    isStandoff.style.filter = 'grayscale(100%)';
-    if (!a.isStandoff) {
[…]
