### 2023-10-16 fd45143
Added function to generate bubbel buttons
