### 2023-10-16 3906a88
Fix icon spinning
Don't use a function for this
