+++ app/static/lib/annotation.js
@@ -651,19 +651,12 @@ export function loadWebAnnotation(prev = '') {
   }
 }
 
-/**
- * Spins the icon to indicate loading activity
- */
-function spinIcon() {
-  const icon = document.getElementById('addWebAnnotationIcon');
-  const svgs = Array.from(icon.getElementsByTagName('svg'));
-  svgs.forEach((t) => t.classList.add('clockwise'));
-}
-
 // Wrapper around traverseAndFetch that reports back errors / progress to 'Load linked data' UI
 export function attemptFetchExternalResource(url, targetTypes, configObj) {
   // spin the icon to indicate loading activity
-  spinIcon();
+  const icon = document.getElementById('addWebAnnotationIcon');
+  const svgs = Array.from(icon.getElementsByTagName('svg'));
+  svgs.forEach((t) => t.classList.add('clockwise'));
   traverseAndFetch(url, targetTypes, configObj)
     .catch((resp) => {
       if (userProvided) {
@@ -681,7 +674,9 @@ export function attemptFetchExternalResource(url, targetTypes, configObj) {
 
 export function fetchWebAnnotations(url, userProvided = true, jumps = 10) {
   // spin the icon to indicate loading activity
-  spinIcon();
+  const icon = document.getElementById('addWebAnnotationIcon');
+  const svgs = Array.from(icon.getElementsByTagName('svg'));
+  svgs.forEach((t) => t.classList.add('clockwise'));
   let fetchAppropriately = solid.getDefaultSession().info.isLoggedIn ? solid.fetch : fetch;
   fetchAppropriately(url, {
     headers: {
