### 2023-10-16 #5863 @wergo
Dear @faroit, sorry for the late response. While I was reading your message, I was under the impression that you ask Stefan Balke whether he has a COI. But probably you addressed this question to us.

I can declare no conflict of interest with Stefan Balke. Both authors do not know Stefan Balke in person and have not collaborated with him.
