### 2023-10-16 33e4593
Finally removed out-commented sections
