+++ app/static/lib/annotation.js
@@ -1,24 +1,9 @@
-import { v, cm, log, translator, /*setStandoffAnnotationEnabledStatus*/ } from './main.js';
+import { v, cm, log, translator } from './main.js';
 import { convertCoords, generateXmlId, rmHash, setCursorToId } from './utils.js';
 import { meiNameSpace, xmlNameSpace, xmlToString } from './dom-utils.js';
-/*import {
-  circle,
-  diffRemoved,
-  highlight,
-  fileCode,
-  identify,
-  link,
-  pencil,
-  rdf,
-  speechBubble,
-  symLinkFile,
-} from '../css/icons.js';*/
 import { removeInEditor } from './editor.js';
 import {
-  //loginAndFetch,
   solid,
-  //solidLogout,
-  //provider,
   getSolidStorage,
   friendContainer,
   annotationContainer,
@@ -31,183 +16,6 @@ import {
 } from './solid.js';
 import { nsp, traverseAndFetch } from './linked-data.js';
 import { deleteAnnotation, isItemInList, addListItem } from './enrichment_panel.js';
-
-//export let annotations = [];
-
-// to enrichment_panel.js
-/*export function situateAndRefreshAnnotationsList(forceRefresh = false) {
-  situateAnnotations();
-  if (forceRefresh || !document.getElementsByClassName('annotationListItem').length) refreshAnnotationsList();
-}
-
-// to enrichment_panel.js
-export function refreshAnnotationsList() {
-  const list = document.getElementById('listAnnotations');
-  // clear list
-  while (list.firstChild) {
-    list.removeChild(list.lastChild);
-  }
-  // console.log("Annotations: ",annotations);
-
-  // add web annotation button
-  const addWebAnnotation = document.createElement('span');
-  const rdfIcon = document.createElement('span');
-  addWebAnnotation.textContent = translator.lang.addWebAnnotation.text;
-  rdfIcon.id = 'addWebAnnotationIcon';
-  rdfIcon.insertAdjacentHTML('beforeend', rdf);
-  addWebAnnotation.appendChild(rdfIcon);
-  addWebAnnotation.id = 'addWebAnnotation';
-  addWebAnnotation.addEventListener('click', () => loadWebAnnotation());
-  list.appendChild(addWebAnnotation);
-  list.insertAdjacentHTML(
-    'beforeend',
-    annotations.length ? '' : '<p>' + translator.lang.noAnnotationsToDisplay.text + '.</p>'
-  );
-  annotations.forEach((a) => {
-    const annoDiv = document.createElement('div');
-    annoDiv.classList.add('annotationListItem');
-    annoDiv.id = a.id;
-    const details = document.createElement('details');
-    details.setAttribute('open', '');
-    const summary = document.createElement('summary');
-    const annoListItemButtons = document.createElement('div');
-    annoListItemButtons.classList.add('annotationListItemButtons');
-    const flipToAnno = document.createElement('a');
-    flipToAnno.insertAdjacentHTML('afterbegin', symLinkFile); //flipToEncoding;
-    flipToAnno.classList.add('flipPageToAnnotationText');
-    flipToAnno.title = translator.lang.flipPageToAnnotationText.description;
-    flipToAnno.classList.add('icon');
-    if (!'selection' in a) flipToAnno.classList.add('disabled');
-    const addObservation = document.createElement('a');
-    addObservation.insertAdjacentHTML('afterbegin', speechBubble);
-    addObservation.classList.add('addObservation');
-    addObservation.title = 'Add observation to extract';
-    addObservation.classList.add('icon');
-    if (a.type !== !'annotateIdentify') {
-      addObservation.classList.add('disabled');
-    }
-    const deleteAnno = document.createElement('a');
[…]
