### 2023-10-16 859201e
Reorder annotation.js and add documentation
