+++ app/static/lib/annotation.js
@@ -17,8 +17,13 @@ import {
 import { nsp, traverseAndFetch } from './linked-data.js';
 import { deleteAnnotation, isItemInList, addListItem } from './enrichment_panel.js';
 
-// functions to draw annotations
+//#region functions to draw annotations
 
+/**
+ * Takes an annotation object and highlights
+ * the selected elements in the notation panel
+ * @param {Object} a annotation object
+ */
 export async function drawIdentify(a) {
   if ('selection' in a) {
     const els = a.selection.map((s) => document.getElementById(s));
@@ -30,6 +35,11 @@ export async function drawIdentify(a) {
   }
 }
 
+/**
+ * Takes an annotation object and highlights
+ * the selected elements in the notation panel
+ * @param {Object} a annotation object
+ */
 export function drawHighlight(a) {
   if ('selection' in a) {
     const els = a.selection.map((s) => document.getElementById(s));
@@ -68,6 +78,12 @@ export function drawCircle(a) {
   }
 }
 
+/**
+ * Takes an annotation object and highlights
+ * the selected elements in the notation panel.
+ * Adds the description text as tool tip.
+ * @param {Object} a annotation object
+ */
 export function drawDescribe(a) {
   if ('selection' in a && 'description' in a) {
     const els = a.selection.map((s) => document.getElementById(s));
@@ -86,6 +102,12 @@ export function drawDescribe(a) {
   }
 }
 
+/**
+ * Takes an annotation object and highlights
+ * the selected elements in the notation panel.
+ * Adds a hyperlink (opening in a new tab).
+ * @param {Object} a annotation object
+ */
 export function drawLink(a) {
   if ('selection' in a && 'url' in a) {
     const els = a.selection.map((s) => document.getElementById(s));
@@ -106,7 +128,15 @@ export function drawLink(a) {
   }
 }
 
-// functions to create annotations
+//#endregion
+
+//#region functions to create annotations
+// called by addAnnotationHandlers() in the Tools tab of the Enrichment panel
+
+/**
+ * Creates a standoff Identify annotation, based on Music Annotation Ontology.
+ * @param {Event} e event
+ */
 export const createIdentify = (e) => {
   if (solid.getDefaultSession().info.isLoggedIn) {
     const label = window.prompt('Add label for identified object (optional)');
@@ -138,6 +168,11 @@ export const createIdentify = (e) => {
     document.getElementById('solidButton').click();
   }
 };
+
+/**
+ * Creates a new highlight annotation either inline or standoff.
+ * @param {Event} e event
+ */
 export const createHighlight = (e) => {
   const a = {
     id: generateXmlId('annot', v.xmlIdStyle),
@@ -148,6 +183,11 @@ export const createHighlight = (e) => {
   writeInlineIfRequested(a);
   writeStandoffIfRequested(a);
 };
+
+/**
+ * Creates a new circle annotation either inline or standoff.
+ * @param {Event} e event
+ */
 export const createCircle = (e) => {
   const a = {
     id: generateXmlId('annot', v.xmlIdStyle),
@@ -158,6 +198,11 @@ export const createCircle = (e) => {
   writeInlineIfRequested(a);
   writeStandoffIfRequested(a);
 };
+
+/**
+ * Creates a new describe annotation either inline or standoff.
+ * @param {Event} e event
+ */
 export const createDescribe = (e) => {
   // TODO improve UX!
   const desc = window.prompt(translator.lang.askForDescription.text);
@@ -171,6 +216,11 @@ export const createDescribe = (e) => {
   writeInlineIfRequested(a);
   writeStandoffIfRequested(a);
 };
+
+/**
[…]
