### 2023-10-18 55e98b9
duració changed for durada twice
