+++ app/static/lang/lang.ca.js
@@ -167,8 +167,8 @@ export const lang = {
   pitchOctaveDownText: { text: 'Disminueix 1 octava' },
   staffUpText: { text: 'Element 1 pauta amunt' },
   staffDownText: { text: 'Element 1 pauta avall' },
-  increaseDurText: { text: 'Augmenta duració' },
-  decreaseDurText: { text: 'Disminueix duració' },
+  increaseDurText: { text: 'Augmenta durada' },
+  decreaseDurText: { text: 'Disminueix durada' },
   cleanAccidText: { text: 'Verificar @accid.ges' },
   renumberMeasuresTestText: { text: 'Reenumera compassos (prova)' },
   renumberMeasuresExecText: { text: 'Reenumera compassos (execució)' },
