+++ app/static/lang/lang.ca.js
@@ -22,24 +22,23 @@ export const lang = {
       </p>
       <p>
         Tot i que mei-friend és una aplicació basada en el navegador, les teves dades personals (incloent la codificació que
-        estàs editant, la configuració de l'aplicació i els detalls actuals de l'inici de sessió, si n'hi ha) es
-        emmagatzemen al
+        estàs editant, la configuració de l'aplicació i els detalls actuals de l'inici de sessió, si n'hi ha) s'emmagatzemen a la
         <a href="https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage" target="_blank"
-          >magatzematge local</a
-        > del teu navegador i no es transmeten ni es emmagatzemen als nostres servidors.
+          >memòria local</a
+        > del teu navegador i no es transmeten ni s'emmagatzemen als nostres servidors.
       </p>
       <p>
         Les dades es transmeten a GitHub només quan ho demanes explícitament (per exemple, quan inicies la sessió a GitHub,
         carreguis la teva codificació des d'un repositori GitHub o quan demanis que s'executi un flux de treball GitHub
         Action). De manera similar, les dades es transmeten al teu proveïdor Solid escollit només quan ho demanes explícitament
-        (per exemple, quan inicies la sessió a Solid o carreguis o desis anotacions stand-off).
+        (per exemple, quan iniciïs la sessió a Solid o carreguis o desis anotacions stand-off).
       </p>
       <p>
         Fem servir <a href="https://matomo.org/" target="_blank">Matomo</a>
         per recopilar estadístiques d'ús anònimes. Aquestes inclouen la teva adreça IP truncada (que permet la geolocalització a
-        nivell de país però sense identificació addicional), el teu navegador i sistema operatiu, d'on vas arribar (és a dir, el
-        lloc web de referència), l'hora i durada de la teva visita i les pàgines que vas visitar. Aquesta informació es
-        emmagatzema a la instància de Matomo que s'executa als servidors de la Universitat de Música i Arts Escèniques de
+        nivell de país però sense identificació addicional), el teu navegador i sistema operatiu, des d'on has obert la pàgina (és a dir, el
+        lloc web de referència), l'hora i durada de la teva visita i les pàgines que hi has visitat. Aquesta informació s'emmagatzema 
+        a la instància de Matomo que s'executa als servidors de la Universitat de Música i Arts Escèniques de
         Viena i no es comparteix amb cap tercer.
       </p>
       <p>
@@ -62,12 +61,12 @@ export const lang = {
         <a href="https://spdx.org/licenses/AGPL-3.0-or-later.html" target="_blank"
           >Llicència Pública General Affero de GNU versió 3.0 (GNU AGPLv3)</a
         >. Consulta la nostra <a href="https://mei-friend.github.io/about/" target="_blank">pàgina d'agraïments</a> per a
-        obtenir més informació sobre els col·laboradors i els components de codi obert reutilitzats en el nostre projecte.
-        Agraim als nostres companys pel seu contribut i orientació.
+        obtenir més informació sobre les persones que hi han col·labort i els components de codi obert reutilitzats en el nostre projecte.
+        Agraim als nostres companys i companyes per la seva contribució i orientació.
       </p>
       <p>
-        El desenvolupament de l'aplicació web de mei-friend està finançat per la
-        <a href="https://fwf.ac.at" target="_blank">Fons d'Investigació d'Àustria (FWF)</a> en els projectes
+        El desenvolupament de l'aplicació web de mei-friend està finançat per 
[…]
