### 2023-10-18 eafe517
Changes in the Insert menu. Catalan. Not complete yet
