+++ app/static/lang/lang.ca.js
@@ -115,7 +115,7 @@ export const lang = {
   importHumdrum: { text: 'Importa Humdrum' },
   importPae: { text: 'Importa PAE, ABC' },
   saveMeiText: { text: 'Desa MEI' },
-  saveMeiBasicText: { text: 'Desa come MEI Basic' },
+  saveMeiBasicText: { text: 'Desa com a MEI Basic' },
   saveSvg: { text: 'Desa SVG' },
   saveMidi: { text: 'Desa MIDI' },
   printPreviewText: { text: 'Visualitza PDF' },
@@ -133,7 +133,7 @@ export const lang = {
   indentSelectionText: { text: 'Indentar selecció' },
   surroundWithTagsText: { text: 'Envoltar amb tags' },
   surroundWithLastTagText: { text: 'Envoltar amb ' },
-  jumpToLineText: { text: 'Salta a línia' },
+  jumpToLineText: { text: 'Salta a la línia' },
   toMatchingTagText: { text: 'Anar al tag coincident' },
   manualValidateText: { text: 'Valida' },
 
@@ -187,20 +187,20 @@ export const lang = {
   addTempoText: { text: 'Tempo' },
   addDirectiveText: { text: 'Directiva' },
   addDynamicsText: { text: 'Dinàmica' },
-  addSlurText: { text: 'Lligadura' },
-  addTieText: { text: 'Lligall' },
+  addSlurText: { text: 'Lligadura (slur)' },
+  addTieText: { text: 'Lligadura (tie)' },
   addCrescendoHairpinText: { text: 'Crescendo' },
   addDiminuendoHairpinText: { text: 'Diminuendo' },
-  addBeamText: { text: 'Lligat' },
-  addBeamSpanText: { text: 'Amplada del lligat' },
-  addSuppliedText: { text: 'Afegit' },
+  addBeamText: { text: 'Barra (unir pliques)' },
+  addBeamSpanText: { text: 'Amplada de la barra' },
+  addSuppliedText: { text: 'Afegit (supplied)' },
   addSuppliedArticText: { text: 'Afegit (articulació)' },
   addSuppliedAccidText: { text: 'Afegit (accidentals)' },
   addArpeggioText: { text: 'Arpegi' },
   addFermataText: { text: 'Fermata' },
   addGlissandoText: { text: 'Glissando' },
-  addPedalDownText: { text: 'Pedal (abaix)' },
-  addPedalUpText: { text: 'Pedal (amunt)' },
+  addPedalDownText: { text: 'Pedal (activar)' },
+  addPedalUpText: { text: 'Pedal (desactivar)' },
   addTrillText: { text: 'Trinat' },
   addTurnText: { text: 'Tornada' },
   addTurnLowerText: { text: 'Baixa la volta' },
