### 2023-10-18 ec844d5
Last changes in Insert menu (Catalan)
