+++ app/static/lang/lang.ca.js
@@ -185,7 +185,7 @@ export const lang = {
   addFlatText: { html: 'Bemoll &#9837;' },
   addDoubleFlatText: { html: 'Doble bemoll &#119083;' },
   addTempoText: { text: 'Tempo' },
-  addDirectiveText: { text: 'Directiva' },
+  addDirectiveText: { text: 'Anotació de sistema (directive)' },
   addDynamicsText: { text: 'Dinàmica' },
   addSlurText: { text: 'Lligadura (slur)' },
   addTieText: { text: 'Lligadura (tie)' },
@@ -196,14 +196,14 @@ export const lang = {
   addSuppliedText: { text: 'Afegit (supplied)' },
   addSuppliedArticText: { text: 'Afegit (articulació)' },
   addSuppliedAccidText: { text: 'Afegit (accidentals)' },
-  addArpeggioText: { text: 'Arpegi' },
+  addArpeggioText: { text: 'Arpeggio' },
   addFermataText: { text: 'Fermata' },
   addGlissandoText: { text: 'Glissando' },
-  addPedalDownText: { text: 'Pedal (activar)' },
-  addPedalUpText: { text: 'Pedal (desactivar)' },
+  addPedalDownText: { text: 'Pedal (accionar)' },
+  addPedalUpText: { text: 'Pedal (desaccionar)' },
   addTrillText: { text: 'Trinat' },
-  addTurnText: { text: 'Tornada' },
-  addTurnLowerText: { text: 'Baixa la volta' },
+  addTurnText: { text: 'Grupetto' },
+  addTurnLowerText: { text: 'Grupetto invertit' },
   addMordentText: { text: 'Mordent' },
   addMordentUpperText: { text: 'Mordent superior' },
   addOctave8AboveText: { text: 'Octava (8va per sobre)' },
