### 2023-10-18 521624c
small language changes before line 366
