+++ app/static/lib/viewer.js
@@ -507,7 +507,9 @@ export default class Viewer {
     } else {
       // set cursor position in buffer
       let found = utils.setCursorToId(cm, itemId);
-      if (!found) this.showMissingIdsWarning();
+      if (!found) {
+        this.showMissingIdsWarning(e.currentTarget.classList.item(0));
+      }
       this.selectedElements = [];
       this.selectedElements.push(itemId);
       msg += 'newly created: ' + itemId + ', size: ' + this.selectedElements.length;
@@ -526,16 +528,31 @@ export default class Viewer {
     this.allowCursorActivity = true;
   } // handleClickOnNotation()
 
-  showMissingIdsWarning() {
-    this.showAlert(
-      translator.lang.missingIdsWarningAlert.text +
-        ' (' +
-        translator.lang.manipulateMenuTitle.text +
-        '&mdash;' +
-        translator.lang.addIdsText.text +
-        ')',
-      'warning'
-    );
+  showMissingIdsWarning(nodeName = '') {
+    if (
+      [
+        'note',
+        'rest',
+        'chord',
+        'mRest',
+        'multiRest',
+        'beam',
+        'tuplet',
+        'accid',
+        'artic',
+        att.modelControlEvents,
+      ].includes(nodeName)
+    ) {
+      this.showAlert(
+        translator.lang.missingIdsWarningAlert.text +
+          ' (' +
+          translator.lang.manipulateMenuTitle.text +
+          '&mdash;' +
+          translator.lang.addIdsText.text +
+          ')',
+        'warning'
+      );
+    }
   } // showMissingIdsWarning()
 
   // when cursor pos in editor changed, update notation location / highlight
