### 2023-10-26 fe4b036
Add markup menu to enrichment panel
