+++ app/static/lang/lang.en.js
@@ -352,7 +352,8 @@ export const lang = {
   annotationCloseButtonText: { text: 'Close Annotations Panel' },
   hideAnnotationPanelButton: { description: 'Close Annotations Panel' },
   closeAnnotationPanelButton: { description: 'Close Annotations Panel' },
-  annotationToolsButton: { text: 'Tools', description: 'Annotation tools' },
+  markupToolsButton: { text: 'Markup', description: 'Markup tools' },
+  annotationToolsButton: { text: 'Annotate', description: 'Annotation tools' },
   annotationListButton: { text: 'List', description: 'List annotations' },
   writeAnnotStandoffText: { text: 'Web Annotation' },
   annotationToolsIdentifyTitle: { text: 'Identify' },
