### 2023-10-26 7a0f182
Move addTranscriptionLikeElement from editor to markup
