+++ app/static/lib/editor.js
@@ -9,7 +9,6 @@ import * as utils from './utils.js';
 import { loadFacsimile } from './facsimile.js';
 import { cmd, handleEditorChanges, translator, version, versionDate } from './main.js';
 import Viewer from './viewer.js';
-import { wrapGroupWithMarkup } from './markup.js';
 
 /**
  * Smart indents selected region in editor, if none, do all
@@ -929,161 +928,6 @@ export function addOctaveElement(v, cm, disPlace = 'above', dis = '8') {
   v.updateData(cm, false, true);
   v.allowCursorActivity = true; // update notation again
 } // addOctaveElement()
-
-/**
- * Surrounds selected elements with a markup element
- * (and a responsibility statement from v.respId derived
- * from mei-friend settings.
- *
- * If attrName is specified, it searches for those attributes,
- * inserts them as new elements, and surrounds them with supplied
- * elements. If there is already such a artic/accid child element,
- * take it and surround it.
- *
- * mElName contains the element name for the element supplied.
- * The element must be a member of att.modelTranscriptionLike.
- * By default, the added markup element will be <supplied>.
- * @param {Viewer} v
- * @param {CodeMirror} cm
- * @param {string} attrName ('artic', 'accid')
- * @param {string} mElName name of markup element to apply
- * @returns
- */
-export function addTranscriptionLikeElement(v, cm, attrName = 'none', mElName = 'supplied') {
-  v.loadXml(cm.getValue());
-  v.selectedElements = speed.filterElements(v.selectedElements, v.xmlDoc);
-  v.selectedElements = utils.sortElementsByScorePosition(v.selectedElements);
-  if (v.selectedElements.length < 1) return;
-  if (!att.modelTranscriptionLike.includes(mElName)) return;
-  v.allowCursorActivity = false;
-
-  let uuids = [];
-  let elementGroups = [];
-
-  // this loop updated the list of selected elements to be wrapped by markup
-  // and will separate the list into groups of adjacent elements to be wrapped into a single markup element
-  while (v.selectedElements.length > 0) {
-    let id = v.selectedElements[0];
-    let el = v.xmlDoc.querySelector("[*|id='" + id + "']");
-    if (!el) {
-      console.warn('No such element in xml document: ' + id);
-      // remove element from list of selected elements if it cannot be found
-      v.selectedElements.splice(v.selectedElements.indexOf(id), 1);
-    } else {
-      // convert attrName artic|accid to note|chord element and surround that
-      if (
-        (['note', 'chord'].includes(el.nodeName) && attrName === 'artic') ||
-        (el.nodeName === 'note' && attrName === 'accid')
-      ) {
-        if (!el.hasAttribute(attrName)) {
-          let childElement;
-          el.childNodes.forEach((ch) => {
-            if (ch.nodeName === attrName) childElement = ch;
-          });
-          if (childElement) {
-            // push childElement to list of groups to wrap
-            // and remove element from selectedElements to avoid double processing
-            elementGroups.push([childElement.getAttribute('xml:id')]);
-            v.selectedElements.splice(v.selectedElements.indexOf(id), 1);
-          } else {
-            const msg =
-              'No ' + attrName + ' attribute or child node found in element ' + el.nodeName + ' (' + id + ').';
-            console.log(msg);
-            v.showAlert(msg, 'warning');
-            return;
-          }
-        } else {
-          // make new element out of attribute and handle it as element to be surrounded
-          let attrValue = el.getAttribute(attrName);
-          let attrEl = document.createElementNS(dutils.meiNameSpace, attrName);
-          let uuid = mintSuppliedId(id, attrName);
[…]
