+++ app/static/lib/markup.js
@@ -5,9 +5,13 @@
  * retrieve different options.
  */
 
-import * as dutils from './dom-utils.js';
-import * as utils from './utils.js';
+import * as att from './attribute-classes.js';
 import { replaceInEditor } from './editor.js';
+import { cmd } from './main.js';
+import * as dutils from './dom-utils.js';
+import * as speed from './speed.js';
+import * as utils from './utils.js';
+import Viewer from './viewer.js';
 
 /**
  * [Wrapper for selectApparatus() and selectChoice().]
@@ -103,7 +107,164 @@ function firstChildElement(parent) {
   return null;
 } // firstChildElement()
 
-export function wrapGroupWithMarkup(v, cm, groupIds, mElName, parentEl) {
+
+/**
+ * Surrounds selected elements with a markup element
+ * (and a responsibility statement from v.respId derived
+ * from mei-friend settings.
+ *
+ * If attrName is specified, it searches for those attributes,
+ * inserts them as new elements, and surrounds them with supplied
+ * elements. If there is already such a artic/accid child element,
+ * take it and surround it.
+ *
+ * mElName contains the element name for the element supplied.
+ * The element must be a member of att.modelTranscriptionLike.
+ * By default, the added markup element will be <supplied>.
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {string} attrName ('artic', 'accid')
+ * @param {string} mElName name of markup element to apply
+ * @returns
+ */
+export function addTranscriptionLikeElement(v, cm, attrName = 'none', mElName = 'supplied') {
+  v.loadXml(cm.getValue());
+  v.selectedElements = speed.filterElements(v.selectedElements, v.xmlDoc);
+  v.selectedElements = utils.sortElementsByScorePosition(v.selectedElements);
+  if (v.selectedElements.length < 1) return;
+  if (!att.modelTranscriptionLike.includes(mElName)) return;
+  v.allowCursorActivity = false;
+
+  let uuids = [];
+  let elementGroups = [];
+
+  // this loop updated the list of selected elements to be wrapped by markup
+  // and will separate the list into groups of adjacent elements to be wrapped into a single markup element
+  while (v.selectedElements.length > 0) {
+    let id = v.selectedElements[0];
+    let el = v.xmlDoc.querySelector("[*|id='" + id + "']");
+    if (!el) {
+      console.warn('No such element in xml document: ' + id);
+      // remove element from list of selected elements if it cannot be found
+      v.selectedElements.splice(v.selectedElements.indexOf(id), 1);
+    } else {
+      // convert attrName artic|accid to note|chord element and surround that
+      if (
+        (['note', 'chord'].includes(el.nodeName) && attrName === 'artic') ||
+        (el.nodeName === 'note' && attrName === 'accid')
+      ) {
+        if (!el.hasAttribute(attrName)) {
+          let childElement;
+          el.childNodes.forEach((ch) => {
+            if (ch.nodeName === attrName) childElement = ch;
+          });
+          if (childElement) {
+            // push childElement to list of groups to wrap
+            // and remove element from selectedElements to avoid double processing
+            elementGroups.push([childElement.getAttribute('xml:id')]);
+            v.selectedElements.splice(v.selectedElements.indexOf(id), 1);
+          } else {
+            const msg =
+              'No ' + attrName + ' attribute or child node found in element ' + el.nodeName + ' (' + id + ').';
+            console.log(msg);
+            v.showAlert(msg, 'warning');
+            return;
+          }
+        } else {
+          // make new element out of attribute and handle it as element to be surrounded
+          let attrValue = el.getAttribute(attrName);
[…]
