### 2023-10-26 abf83c2
Move markup handler from main to enrichment_panel
