+++ app/static/lib/main.js
@@ -53,6 +53,7 @@ import {
 } from './resizer.js';
 import {
   addAnnotationHandlers,
+  addMarkupHandlers,
   clearListItems,
   getSolidIdP,
   readListItemsFromXML,
@@ -739,6 +740,7 @@ function onLanguageLoaded() {
 
   addEventListeners(v, cm);
   addAnnotationHandlers();
+  addMarkupHandlers();
   addNotationResizerHandlers(v, cm);
   addFacsimilerResizerHandlers(v, cm);
   let doit;
@@ -1591,29 +1593,6 @@ export let cmd = {
   addFClefChangeAfter: () => e.addClefChange(v, cm, 'F', '4', false),
   addBeam: () => e.addBeamElement(v, cm),
   addBeamSpan: () => e.addBeamSpan(v, cm),
-  addSupplied: () => e.addTranscriptionLikeElement(v, cm, null, 'supplied'),
-  addSuppliedAccid: () => e.addTranscriptionLikeElement(v, cm, 'accid', 'supplied'),
-  addSuppliedArtic: () => e.addTranscriptionLikeElement(v, cm, 'artic', 'supplied'),
-  addUnclear: () => e.addTranscriptionLikeElement(v, cm, null, 'unclear'),
-  addUnclearAccid: () => e.addTranscriptionLikeElement(v, cm, 'accid', 'unclear'),
-  addUnclearArtic: () => e.addTranscriptionLikeElement(v, cm, 'artic', 'unclear'),
-  addSic: () => e.addTranscriptionLikeElement(v, cm, null, 'sic'),
-  addSicAccid: () => e.addTranscriptionLikeElement(v, cm, 'accid', 'sic'),
-  addSicArtic: () => e.addTranscriptionLikeElement(v, cm, 'artic', 'sic'),
-  addCorr: () => e.addTranscriptionLikeElement(v, cm, null, 'corr'),
-  addCorrAccid: () => e.addTranscriptionLikeElement(v, cm, 'accid', 'corr'),
-  addCorrArtic: () => e.addTranscriptionLikeElement(v, cm, 'artic', 'corr'),
-  addOrig: () => e.addTranscriptionLikeElement(v, cm, null, 'orig'),
-  addOrigAccid: () => e.addTranscriptionLikeElement(v, cm, 'accid', 'orig'),
-  addOrigArtic: () => e.addTranscriptionLikeElement(v, cm, 'artic', 'orig'),
-  addReg: () => e.addTranscriptionLikeElement(v, cm, null, 'reg'),
-  addRegAccid: () => e.addTranscriptionLikeElement(v, cm, 'accid', 'reg'),
-  addRegArtic: () => e.addTranscriptionLikeElement(v, cm, 'artic', 'reg'),
-  addAdd: () => e.addTranscriptionLikeElement(v, cm, null, 'add'),
-  addAddAccid: () => e.addTranscriptionLikeElement(v, cm, 'accid', 'add'),
-  addAddArtic: () => e.addTranscriptionLikeElement(v, cm, 'artic', 'add'),
-  addDel: () => e.addTranscriptionLikeElement(v, cm, null, 'del'),
-  addDelAccid: () => e.addTranscriptionLikeElement(v, cm, 'accid', 'del'),
   addDelArtic: () => e.addTranscriptionLikeElement(v, cm, 'artic', 'del'),
   correctAccid: () => e.checkAccidGes(v, cm),
   renumberMeasuresTest: () => e.renumberMeasures(v, cm, false),
@@ -1965,30 +1944,6 @@ function addEventListeners(v, cm) {
   document.getElementById('addDimHairpin').addEventListener('click', cmd.addDimHairpin);
[…]
