+++ app/static/lang/lang.en.js
@@ -113,9 +113,6 @@ export const lang = {
   addDiminuendoHairpinText: { text: 'Diminuendo hairpin' },
   addBeamText: { text: 'Beam' },
   addBeamSpanText: { text: 'Beamspan' },
-  addSuppliedText: { text: 'Supplied' },
-  addSuppliedArticText: { text: 'Supplied (Artic)' },
-  addSuppliedAccidText: { text: 'Supplied (Accid)' },
   addArpeggioText: { text: 'Arpeggio' },
   addFermataText: { text: 'Fermata' },
   addGlissandoText: { text: 'Glissando' },
@@ -398,6 +395,39 @@ export const lang = {
     text1: 'Cannot write annotation as MEI anchor-point lacks xml:id.',
     text2: 'Please assign identifiers by selecting "Manipulate" -> "Re-render MEI (with ids)" and try again.',
   },
+  // MARKUP MENU
+  addSuppliedText: { text: '<supplied>' },
+  addSupplied: { text: 'Selected elements' },
+  addSuppliedArticText: { text: 'Articulation' },
+  addSuppliedAccidText: { text: 'Accidental' },
+  addUnclearText: { text: '<unclear>' },
+  addUnclear: { text: 'Selected elements' },
+  addUnclearArticText: { text: 'Articulation' },
+  addUnclearAccidText: { text: 'Accidental' },
+  addSicText: { text: '<sic>' },
+  addSic: { text: 'Selected elements' },
+  addSicArticText: { text: 'Articulation' },
+  addSicAccidText: { text: 'Accidental' },
+  addCorrText: { text: '<corr>' },
+  addCorr: { text: 'Selected elements' },
+  addCorrArticText: { text: 'Articulation' },
+  addCorrAccidText: { text: 'Accidental' },
+  addOrigText: { text: '<orig>' },
+  addOrig: { text: 'Selected elements' },
+  addOrigArticText: { text: 'Articulation' },
+  addOrigAccidText: { text: 'Accidental' },
+  addRegText: { text: '<reg>' },
+  addReg: { text: 'Selected elements' },
+  addRegArticText: { text: 'Articulation' },
+  addRegAccidText: { text: 'Accidental' },
+  addAddText: { text: '<add>' },
+  addAdd: { text: 'Selected elements' },
+  addAddArticText: { text: 'Articulation' },
+  addAddAccidText: { text: 'Accidental' },
+  addDelText: { text: '<del>' },
+  addDel: { text: 'Selected elements' },
+  addDelArticText: { text: 'Articulation' },
+  addDelAccidText: { text: 'Accidental' },
 
   // MIDI
   midiSpeedmodeIndicator: {
