### 2023-10-27 04d69f1
Add missing reference after moving addTranscriptionLikeElement
