### 2023-10-27 b349b51
Add svg icons to enrichment panel tab buttons
