+++ app/static/lang/lang.en.js
@@ -349,9 +349,9 @@ export const lang = {
   annotationCloseButtonText: { text: 'Close Annotations Panel' },
   hideAnnotationPanelButton: { description: 'Close Annotations Panel' },
   closeAnnotationPanelButton: { description: 'Close Annotations Panel' },
-  markupToolsButton: { text: 'Markup', description: 'Markup tools' },
-  annotationToolsButton: { text: 'Annotate', description: 'Annotation tools' },
-  annotationListButton: { text: 'List', description: 'List annotations' },
+  markupToolsButton: { description: 'Markup tools' },
+  annotationToolsButton: { description: 'Annotation tools' },
+  annotationListButton: { description: 'List annotations' },
   writeAnnotStandoffText: { text: 'Web Annotation' },
   annotationToolsIdentifyTitle: { text: 'Identify' },
   annotationToolsIdentifySpan: { text: 'Identify Musical Object' },
