+++ app/static/lang/lang.de.js
@@ -110,9 +110,6 @@ export const lang = {
   addDiminuendoHairpinText: { text: 'Diminuendo-Gabel' },
   addBeamText: { text: 'Balken' },
   addBeamSpanText: { text: 'Balkenspanner' },
-  addSuppliedText: { text: 'Supplied' },
-  addSuppliedArticText: { text: 'Supplied (Artic)' },
-  addSuppliedAccidText: { text: 'Supplied (Accid)' },
   addArpeggioText: { text: 'Arpeggio' },
   addFermataText: { text: 'Fermate' },
   addGlissandoText: { text: 'Glissando' },
@@ -143,7 +140,7 @@ export const lang = {
   // HELP MENU ITEM
   helpMenuTitle: { text: 'Hilfe' },
   goToHelpPage: { text: 'mei-friend Hilfeseiten' },
-  goToCheatSheet: {text: 'mei-friend Spickzettel'},
+  goToCheatSheet: { text: 'mei-friend Spickzettel' },
   showChangelog: { text: 'mei-friend Änderungsliste (changelog)' },
   goToGuidelines: { text: 'MEI Guidelines' },
   consultGuidelinesForElementText: { text: 'Guideline-Eintrag für aktuelles Element' },
@@ -359,8 +356,9 @@ export const lang = {
   annotationCloseButtonText: { text: 'Panel für Annotationen schließen' },
   hideAnnotationPanelButton: { description: 'Panel für Annotationen schließen' },
   closeAnnotationPanelButton: { description: 'Panel für Annotationen schließen' },
-  annotationToolsButton: { text: 'Tools', description: 'Annotationswerkzeuge' },
-  annotationListButton: { text: 'Liste', description: 'Annotationen auflisten' },
+  markupToolsButton: { description: 'Werkzeuge für editorische Auszeichnungen' },
+  annotationToolsButton: { description: 'Annotationswerkzeuge' },
+  annotationListButton: { description: 'Annotationen auflisten' },
   writeAnnotStandoffText: { text: 'Web Annotationen' },
   annotationToolsIdentifyTitle: { text: 'Identifizieren' },
   annotationToolsIdentifySpan: { text: 'Musikalisches Objekt identifizieren' },
@@ -405,6 +403,39 @@ export const lang = {
     text1: 'Annotation kann nicht geschrieben werden, da MEI-Ankerpunkt kein xml:id Attribut hat.',
     text2: 'Bitte xml:ids zuweisen ("Verändern" -> "Ids zu MEI hinzufügen") und erneut versuchen.',
   },
+  // MARKUP MENU
+  addSuppliedText: { text: '<supplied>' },
+  addSupplied: { text: 'Ausgewählte Elemente' },
+  addSuppliedArticText: { text: 'Artikulation' },
+  addSuppliedAccidText: { text: 'Vorzeichen' },
+  addUnclearText: { text: '<unclear>' },
+  addUnclear: { text: 'Ausgewählte Elemente' },
+  addUnclearArticText: { text: 'Artikulation' },
+  addUnclearAccidText: { text: 'Vorzeichen' },
+  addSicText: { text: '<sic>' },
+  addSic: { text: 'Ausgewählte Elemente' },
+  addSicArticText: { text: 'Artikulation' },
+  addSicAccidText: { text: 'Vorzeichen' },
+  addCorrText: { text: '<corr>' },
+  addCorr: { text: 'Ausgewählte Elemente' },
+  addCorrArticText: { text: 'Artikulation' },
+  addCorrAccidText: { text: 'Vorzeichen' },
+  addOrigText: { text: '<orig>' },
[…]
