### 2023-10-27 e1aabbd
Add translations for markup tools tab
