+++ app/static/lang/lang.de.js
@@ -406,36 +406,36 @@ export const lang = {
   // MARKUP MENU
   addSuppliedText: { text: '<supplied>' },
   addSupplied: { text: 'Ausgewählte Elemente' },
-  addSuppliedArticText: { text: 'Artikulation' },
-  addSuppliedAccidText: { text: 'Vorzeichen' },
+  addSuppliedArtic: { text: 'Artikulation' },
+  addSuppliedAccid: { text: 'Vorzeichen' },
   addUnclearText: { text: '<unclear>' },
   addUnclear: { text: 'Ausgewählte Elemente' },
-  addUnclearArticText: { text: 'Artikulation' },
-  addUnclearAccidText: { text: 'Vorzeichen' },
+  addUnclearArtic: { text: 'Artikulation' },
+  addUnclearAccid: { text: 'Vorzeichen' },
   addSicText: { text: '<sic>' },
   addSic: { text: 'Ausgewählte Elemente' },
-  addSicArticText: { text: 'Artikulation' },
-  addSicAccidText: { text: 'Vorzeichen' },
+  addSicArtic: { text: 'Artikulation' },
+  addSicAccid: { text: 'Vorzeichen' },
   addCorrText: { text: '<corr>' },
   addCorr: { text: 'Ausgewählte Elemente' },
-  addCorrArticText: { text: 'Artikulation' },
-  addCorrAccidText: { text: 'Vorzeichen' },
+  addCorrArtic: { text: 'Artikulation' },
+  addCorrAccid: { text: 'Vorzeichen' },
   addOrigText: { text: '<orig>' },
   addOrig: { text: 'Ausgewählte Elemente' },
-  addOrigArticText: { text: 'Artikulation' },
-  addOrigAccidText: { text: 'Vorzeichen' },
+  addOrigArtic: { text: 'Artikulation' },
+  addOrigAccid: { text: 'Vorzeichen' },
   addRegText: { text: '<reg>' },
   addReg: { text: 'Ausgewählte Elemente' },
-  addRegArticText: { text: 'Artikulation' },
-  addRegAccidText: { text: 'Vorzeichen' },
+  addRegArtic: { text: 'Artikulation' },
+  addRegAccid: { text: 'Vorzeichen' },
   addAddText: { text: '<add>' },
   addAdd: { text: 'Ausgewählte Elemente' },
-  addAddArticText: { text: 'Artikulation' },
-  addAddAccidText: { text: 'Vorzeichen' },
+  addAddArtic: { text: 'Artikulation' },
+  addAddAccid: { text: 'Vorzeichen' },
   addDelText: { text: '<del>' },
   addDel: { text: 'Ausgewählte Elemente' },
-  addDelArticText: { text: 'Artikulation' },
-  addDelAccidText: { text: 'Vorzeichen' },
+  addDelArtic: { text: 'Artikulation' },
+  addDelAccid: { text: 'Vorzeichen' },
 
   // MIDI
   midiSpeedmodeIndicator: {
