+++ app/static/lang/lang.en.js
@@ -398,36 +398,36 @@ export const lang = {
   // MARKUP MENU
   addSuppliedText: { text: '<supplied>' },
   addSupplied: { text: 'Selected elements' },
-  addSuppliedArticText: { text: 'Articulation' },
-  addSuppliedAccidText: { text: 'Accidental' },
+  addSuppliedArtic: { text: 'Articulation' },
+  addSuppliedAccid: { text: 'Accidental' },
   addUnclearText: { text: '<unclear>' },
   addUnclear: { text: 'Selected elements' },
-  addUnclearArticText: { text: 'Articulation' },
-  addUnclearAccidText: { text: 'Accidental' },
+  addUnclearArtic: { text: 'Articulation' },
+  addUnclearAccid: { text: 'Accidental' },
   addSicText: { text: '<sic>' },
   addSic: { text: 'Selected elements' },
-  addSicArticText: { text: 'Articulation' },
-  addSicAccidText: { text: 'Accidental' },
+  addSicArtic: { text: 'Articulation' },
+  addSicAccid: { text: 'Accidental' },
   addCorrText: { text: '<corr>' },
   addCorr: { text: 'Selected elements' },
-  addCorrArticText: { text: 'Articulation' },
-  addCorrAccidText: { text: 'Accidental' },
+  addCorrArtic: { text: 'Articulation' },
+  addCorrAccid: { text: 'Accidental' },
   addOrigText: { text: '<orig>' },
   addOrig: { text: 'Selected elements' },
-  addOrigArticText: { text: 'Articulation' },
-  addOrigAccidText: { text: 'Accidental' },
+  addOrigArtic: { text: 'Articulation' },
+  addOrigAccid: { text: 'Accidental' },
   addRegText: { text: '<reg>' },
   addReg: { text: 'Selected elements' },
-  addRegArticText: { text: 'Articulation' },
-  addRegAccidText: { text: 'Accidental' },
+  addRegArtic: { text: 'Articulation' },
+  addRegAccid: { text: 'Accidental' },
   addAddText: { text: '<add>' },
   addAdd: { text: 'Selected elements' },
-  addAddArticText: { text: 'Articulation' },
-  addAddAccidText: { text: 'Accidental' },
+  addAddArtic: { text: 'Articulation' },
+  addAddAccid: { text: 'Accidental' },
   addDelText: { text: '<del>' },
   addDel: { text: 'Selected elements' },
-  addDelArticText: { text: 'Articulation' },
-  addDelAccidText: { text: 'Accidental' },
+  addDelArtic: { text: 'Articulation' },
+  addDelAccid: { text: 'Accidental' },
 
   // MIDI
   midiSpeedmodeIndicator: {
