+++ app/static/lang/lang.ca.js
@@ -408,36 +408,36 @@ export const lang = {
   // MENÚ DE MARCATGE
   addSuppliedText: { text: '<supplied>' },
   addSupplied: { text: 'Elements seleccionats' },
-  addSuppliedArticText: { text: 'Articulació' },
-  addSuppliedAccidText: { text: 'Accidental' },
+  addSuppliedArtic: { text: 'Articulació' },
+  addSuppliedAccid: { text: 'Accidental' },
   addUnclearText: { text: '<unclear>' },
   addUnclear: { text: 'Elements seleccionats' },
-  addUnclearArticText: { text: 'Articulació' },
-  addUnclearAccidText: { text: 'Accidental' },
+  addUnclearArtic: { text: 'Articulació' },
+  addUnclearAccid: { text: 'Accidental' },
   addSicText: { text: '<sic>' },
   addSic: { text: 'Elements seleccionats' },
-  addSicArticText: { text: 'Articulació' },
-  addSicAccidText: { text: 'Accidental' },
+  addSicArtic: { text: 'Articulació' },
+  addSicAccid: { text: 'Accidental' },
   addCorrText: { text: '<corr>' },
   addCorr: { text: 'Elements seleccionats' },
-  addCorrArticText: { text: 'Articulació' },
-  addCorrAccidText: { text: 'Accidental' },
+  addCorrArtic: { text: 'Articulació' },
+  addCorrAccid: { text: 'Accidental' },
   addOrigText: { text: '<orig>' },
   addOrig: { text: 'Elements seleccionats' },
-  addOrigArticText: { text: 'Articulació' },
-  addOrigAccidText: { text: 'Accidental' },
+  addOrigArtic: { text: 'Articulació' },
+  addOrigAccid: { text: 'Accidental' },
   addRegText: { text: '<reg>' },
   addReg: { text: 'Elements seleccionats' },
-  addRegArticText: { text: 'Articulació' },
-  addRegAccidText: { text: 'Accidental' },
+  addRegArtic: { text: 'Articulació' },
+  addRegAccid: { text: 'Accidental' },
   addAddText: { text: '<add>' },
   addAdd: { text: 'Elements seleccionats' },
-  addAddArticText: { text: 'Articulació' },
-  addAddAccidText: { text: 'Accidental' },
+  addAddArtic: { text: 'Articulació' },
+  addAddAccid: { text: 'Accidental' },
   addDelText: { text: '<del>' },
   addDel: { text: 'Elements seleccionats' },
-  addDelArticText: { text: 'Articulació' },
-  addDelAccidText: { text: 'Accidental' },
+  addDelArtic: { text: 'Articulació' },
+  addDelAccid: { text: 'Accidental' },
 
   // MIDI
   midiSpeedmodeIndicator: {

[…]
