### 2023-11-01 5ed7bb8
Add public function addMarkup() to encapsulate markup functions
