### 2023-11-01 #6002 @editorialbot
[REVIEW]: mei-friend: An Interactive Web-based Editor for Digital Music Encodings
**Submitting author:** @wergo (<a href="http://orcid.org/0000-0002-1722-0718">Werner Goebl</a>)
**Repository:** https://github.com/mei-friend/mei-friend
**Branch with paper.md** (empty if default branch): joss2023
**Version:** v1.0.14
