+++ app/static/lib/enrichment_panel.js
@@ -20,6 +20,7 @@ import {
   rdf,
   speechBubble,
   symLinkFile,
+  codeScan,
 } from '../css/icons.js';
 import { loginAndFetch, solid, solidLogout, provider } from './solid.js';
 import { nsp } from './linked-data.js';
@@ -180,8 +181,9 @@ export function refreshAnnotationsInRendering(forceListRefresh = false) {
   rac.appendChild(annoSvg);
   if (document.getElementById('showAnnotations')?.checked) {
     // drawing handlers can draw into renderedAnnotationsSvg if they need to
+    // markup does not need to be drawn explicitly because this is already handled by Verovio
     listItems.forEach((a) => {
-      if ('type' in a) {
+      if ('type' in a && !('isMarkup' in a)) {
         switch (a.type) {
           case 'annotateIdentify':
             annot.drawIdentify(a);
@@ -202,7 +204,7 @@ export function refreshAnnotationsInRendering(forceListRefresh = false) {
             console.warn("Don't have a drawing function for this type of annotation", a);
         }
       } else {
-        console.warn('Skipping annotation without type: ', a);
+        console.warn('Skipping markup or annotation without type: ', a);
       }
     });
   }
@@ -260,28 +262,32 @@ function generateListItem(a) {
   const details = document.createElement('details');
   details.setAttribute('open', '');
   const summary = document.createElement('summary');
-
-  switch (a.type) {
-    case 'annotateHighlight':
-      summary.insertAdjacentHTML('afterbegin', highlight);
-      break;
-    case 'annotateCircle':
-      summary.insertAdjacentHTML('afterbegin', circle);
-      break;
-    case 'annotateLink':
-      summary.insertAdjacentHTML('afterbegin', link);
-      details.insertAdjacentHTML('afterbegin', a.url);
-      break;
-    case 'annotateDescribe':
-      summary.insertAdjacentHTML('afterbegin', pencil);
-      details.insertAdjacentHTML('afterbegin', a.description);
-      break;
-    case 'annotateIdentify':
-      summary.insertAdjacentHTML('afterbegin', identify);
-      details.insertAdjacentHTML('afterbegin', `<div class="mao-musMat" id="musMat_${a.id}"></div>`);
-      break;
-    default:
-      console.warn('Unknown type when drawing annotation in list: ', a);
+  if (a.isMarkup === true) {
+    summary.insertAdjacentHTML('afterbegin', codeScan);
+    details.insertAdjacentHTML('afterbegin', a.type);
+  } else {
+    switch (a.type) {
+      case 'annotateHighlight':
+        summary.insertAdjacentHTML('afterbegin', highlight);
+        break;
+      case 'annotateCircle':
+        summary.insertAdjacentHTML('afterbegin', circle);
+        break;
+      case 'annotateLink':
+        summary.insertAdjacentHTML('afterbegin', link);
+        details.insertAdjacentHTML('afterbegin', a.url);
+        break;
+      case 'annotateDescribe':
+        summary.insertAdjacentHTML('afterbegin', pencil);
+        details.insertAdjacentHTML('afterbegin', a.description);
+        break;
+      case 'annotateIdentify':
+        summary.insertAdjacentHTML('afterbegin', identify);
+        details.insertAdjacentHTML('afterbegin', `<div class="mao-musMat" id="musMat_${a.id}"></div>`);
+        break;
+      default:
[…]
