### 2023-11-01 5dd230f
Fix invalid DOI error
