+++ joss2023/paper.bib
@@ -173,7 +173,7 @@
     booktitle={Proceedings of the 6th International Conference on Digital Libraries for Musicology ({DLfM'19})},
     pages={84--88},
     year={2019},
-    doi={https://doi.org/10.1145/3358664.3358666}
+    doi={10.1145/3358664.3358666}
 }
 
 @inproceedings{Mansour_etal_2016,
