### 2023-11-01 #6002 @wergo
> @wergo happy to announce that we found two reviewers so quickly. While the reviewers are getting started, maybe you can already go ahead and try to fix the invalid DOI error?

Thanks for reviewing!

DOI error is fixed (superfluous https://doi.org/ removed).
