### 2023-11-02 d5e6aaf
Fix null reference error
