### 2023-11-02 3570eb3
Add brackets to markup types in bubbles
