### 2023-11-02 42c7cc3
Start with general deleteListItem functionality
