### 2023-11-02 #6002 @stefan-balke
Hi @wergo,

thanks for your contribution. I think the `mei-friend` is a very valuable contribution to the field.
The contribution as well as the paper are in very good shape. However, the documentation and the architecture of the actual implementation could be enhanced.

I would propose the following additions:
* Add the instructions for developers here: https://mei-friend.github.io/docs/advanced/dev/
  There are some in the INSTALL.md but they are mentioned nowhere and I found them by chance. Also: Is it really necessary to have a GitHub App registered for developing the core JS functionalities? Maybe there could be different levels of documentation (running locally, running locally + development, optional: Deploying on your own lab server --> links to other tutorials etc.)
* The architecture of the app itself is a client-server app using flask. However, the code itself is not split into client and server components. Could you lay out the architecture and the tasks of the client and the server respectively?
* The developer documentation is missing completely. I know, this is much work but will increase the value of the software tremendously. Start with a documentation of the folder structure and the files. Give pointers to the main files and where developers could find a handle. It could also be a short scree-cast showing people how to implement new features.
* Unit-Tests are missing. Start with testing core functionalities such as importing MusicXML files and checking against an expected MEI-output. These also give devs a good starting point to understand the inner functionalities of your software.

Keep up the great work. Since JOSS is focused on the Software aspect, please include the above mentioned, mainly docs, issues.
