### 2023-11-02 9fc6c0c
Add function to delete markup
