### 2023-11-02 6be2201
Started retrieval of markup from xmlDoc
