### 2023-11-03 041138e
Fix reading of annots and markup from xml
