### 2023-11-03 54a041a
Refactor situating of ListItems
Still buggy!
