+++ app/static/lib/markup.js
@@ -320,6 +320,10 @@ export function addTranscriptionLikeElement(v, cm, attrName = 'none', mElName = 
   if (uuids.length > 0) {
     v.selectedElements = [];
     uuids.forEach((u) => v.selectedElements.push(u));
+    if(v.selectedElements.length > 1) {
+      // add corresp attribute to markup elements
+      addCorrespAttr(v.selectedElements);
+    }
   }
 
   addApplicationInfo(v, cm);
@@ -361,6 +365,19 @@ function wrapGroupWithMarkup(v, cm, groupIds, mElName, parentEl) {
   cm.execCommand('indentAuto');
 
   return uuid;
+}
+
+function addCorrespAttr(grpIds) {
+  let grpLinks = grpIds.join(' #');
+  grpLinks = '#' + grpLinks;
+
+  grpIds.forEach((id) => {
+    let el = v.xmlDoc.querySelector("[*|id='" + id + "']");
+    if(!el.getAttribute('corresp')) {
+      el.setAttribute('corresp', grpLinks);
+      replaceInEditor(cm, el, true);
+    }
+  });
 }
 
 function mintSuppliedId(id, nodeName, v) {
