### 2023-11-06 mei-friend/mei-friend#58 @stefan-balke
Enhance installation docs
There is some information in the INSTALL.md but they are mentioned nowhere and I found them by chance.

Also: Is it really necessary to have a GitHub App registered for developing the core JS functionalities? Maybe there could be different levels of documentation (running locally, running locally + development, optional: Deploying on your own lab server --> links to other tutorials etc.)

Derived from the JOSS review in https://github.com/openjournals/joss-reviews/issues/6002#issuecomment-1790467282
