### 2023-11-06 mei-friend/mei-friend#59 @stefan-balke
Describe architecture and tasks of client and server
The architecture of the app itself is a client-server app using flask. However, the code itself is not split into client and server components.

Could you lay out the architecture and the tasks of the client and the server respectively?

Derived from the JOSS review in https://github.com/openjournals/joss-reviews/issues/6002#issuecomment-1790467282
