### 2023-11-06 mei-friend/mei-friend#60 @stefan-balke
Add developer documentation
The developer documentation is very sparse to non-existing currently :-).
I know, this is much work but will increase the value of the software tremendously. Start with a documentation of the folder structure and the files. Give pointers to the main files and where developers could find a handle. It could also be a short screen-cast showing people how to implement new features.

Derived from the JOSS review in https://github.com/openjournals/joss-reviews/issues/6002#issuecomment-1790467282
