### 2023-11-06 mei-friend/mei-friend#61 @stefan-balke
Add Unit-Tests for core features
Unit-Tests are currently missing. You could start with testing core functionalities such as importing MusicXML files and checking against an expected MEI-output. These also give devs a good starting point to understand the inner functionalities of your software.

Derived from the JOSS review in https://github.com/openjournals/joss-reviews/issues/6002#issuecomment-1790467282
