### 2023-11-06 mei-friend/mei-friend#62 @stefan-balke
Make codemirror a git-submodule
The manual download of codemirror5 could be automated, e.g. with a git submodule

Derived from the JOSS review in https://github.com/openjournals/joss-reviews/issues/6002#issuecomment-1790467282
