### 2023-11-07 7dfaa8c
Adding note with `n` based on selection
