### 2023-11-08 a9afd59
Make sure to add only one item object for related markup
Adds some utility functions to filter listItems and flatten arrays
