+++ app/static/lib/markup.js
@@ -14,6 +14,10 @@ import * as utils from './utils.js';
 import Viewer from './viewer.js';
 import { addListItem, isItemInList, refreshAnnotationsList, retrieveItemValuesByProperty } from './enrichment_panel.js';
 
+/**
+ * Reads markup elements from the XML document and creates
+ * list item objects for each group of corresponding markup elements.
+ */
 export function readMarkup() {
   // read all markup elements supported from the current document
   // create objects out of them
@@ -171,14 +175,25 @@ function firstChildElement(parent) {
   return null;
 } // firstChildElement()
 
+/**
+ * Calls the function to add markup and refreshes the list of items
+ * @param {string} attrName
+ * @param {string} mElName
+ */
 export function addMarkup(attrName = 'none', mElName = 'supplied') {
   addTranscriptionLikeElement(v, cm, attrName, mElName);
   //let successfullyAdded = xmlMarkupToListItem(v.selectedElements, mElName);
+  // Manually updating the item list is not necessary because refreshing the code in the editor triggers readMarkup()
   refreshAnnotationsList();
 }
-// TODO FIX!!!! Somehow, the markup is already added to the list by handling the editor changes. This is stupid, prevent this!!!!
-// This messes with markup that creates more than one xml element, e.g. for notes and control events!!!
 
+/**
+ * Creates a list item object for markup elements based on some basic data
+ * @param {string} currentElementId xml:id of current element -> item.id
+ * @param {string} mElName element name -> item.type
+ * @param {Array[string]} correspElements ids of corresponding elements (including self) -> item.selection
+ * @returns {boolean} adding item was successful
+ */
 function xmlMarkupToListItem(currentElementId, mElName, correspElements) {
   // one addMarkupAction might result in multiple markup elements
   // e.g. when selecting notes and control events
@@ -186,7 +201,6 @@ function xmlMarkupToListItem(currentElementId, mElName, correspElements) {
   // properties are similar to annotations
   // id is the xml:id of the first markup element
   // selection contains all markup elements created at the same time
-  // type uses capitalised element name, e.g. markupSupplied
 
   const markupItem = {
     id: currentElementId,
@@ -295,7 +309,6 @@ export function addTranscriptionLikeElement(v, cm, attrName = 'none', mElName = 
 
   // this loop iterates over the array of arrays of grouped ids
   // and wraps the markup around a whole group
-  // TODO (nice to have): Add @corresp to elements if more than one is created at once
   elementGroups.forEach((group) => {
     let parent = v.xmlDoc.querySelector("[*|id='" + group[0] + "']").parentNode;
 
@@ -350,14 +363,13 @@ export function addTranscriptionLikeElement(v, cm, attrName = 'none', mElName = 
       // add corresp attribute to markup elements
       addCorrespAttr(v.selectedElements);
     }
-    v.selectedElements.forEach(id => {
+    v.selectedElements.forEach((id) => {
       let el = v.xmlDoc.querySelector("[*|id='" + id + "']");
       let parentEl = el.parentNode;
       replaceInEditor(cm, parentEl, true);
       cm.execCommand('indentAuto');
       // refresh editor here after the markup insertion is finished
     });
-
   }
 
   addApplicationInfo(v, cm);
@@ -365,6 +377,15 @@ export function addTranscriptionLikeElement(v, cm, attrName = 'none', mElName = 
   v.allowCursorActivity = true; // update notation again
 } // addSuppliedElement()
 
+/**
[…]
