### 2023-11-08 c23a418
Fix bug in deleteListItem
