### 2023-11-08 784ce0c
Polish addNote code
