### 2023-11-09 4157d71
Merge speed.filterElements()
