### 2023-11-09 db55426
Toggle between notation and encoding
