### 2023-11-09 24a2b14
Duplicate note also for selected chord
