+++ app/static/lib/main.js
@@ -1597,6 +1597,7 @@ export let cmd = {
   redo: () => cm.redo(),
   // add note
   addNote: () => e.addNote(v, cm),
+  convertToChord: () => e.convertToChord(v, cm),
   toggleDots: () => e.toggleDots(v, cm),
   // add accidentals
   addDoubleSharp: () => e.addAccidental(v, cm, 'x'),
