### 2023-11-09 fea6825
Insert new note into empty layer
