### 2023-11-09 3ae26d6
Some refactoring at switch focus notation/encoding
