### 2023-11-09 e7c1506
Menu items for add note, toggle chord, dots
