### 2023-11-09 fdc3d42
Convert notes to rests with SHIFT + N
