+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # mei-friend-online CHANGELOG.md
 ### 1.0.5 (?) patch
-* Keyboard shortcut to duplicate a selected note or rest
-* Keyboard shortcut (`CTRL + SPACE`) to switch focus between notation and encoding panel
+* Keyboard shortcut (`CTRL + SPACE`) to switch focus between notation and encoding panel (#55)
+* Keyboard shortcut (`N`) to duplicate a selected chord, note or rest, or insert a new note into an empty layer (#57)
+* Keyboard shortcut (`SHIFT + N`) to convert selected notes to rests and rests to notes (#57)
 * Keyboard shortcut (`.`) to add or remove `dots="1"` to notes, chords, etc.
-* Keyboard shortcut (`C`) to embed selected notes inside a chord element, and vice versa
+* Keyboard shortcut (`C`) to embed selected notes inside a new chord element, and vice versa (remove chord and leave notes)
 
 ### 1.0.4 patch
 * Tweak CSS select and translations (German)
