### 2023-11-09 def7451
Improve selection after converting to chord
