+++ app/static/lang/lang.eo.js
@@ -162,6 +162,7 @@ export const lang = {
   staffDownText: { text: 'Elemento 1 personalo malsupren' },
   increaseDurText: { text: 'Pligrandigi daŭron' },
   decreaseDurText: { text: 'Malpligrandigi daŭron' },
+  toggleDotsText: { text: 'Ŝalti punktadon' },
   cleanAccidText: { text: 'Kontroli @accid.ges' },
   renumberMeasuresTestText: { text: ' Reenumeri mezurojn (testo)' },
   renumberMeasuresExecText: { text: ' Reenumeri mezurojn (ekzekuti)' },
