### 2023-11-10 303b9f9
Add missing insert notes translations
