### 2023-11-10 #6002 @wergo
Dear @stefan-balke, many thanks for the valuable comments and suggestions which we will be carefully including into our submission.

Just one question regarding the missing unit-tests: we would have thought about using the Selenium framework. Do you have an alternative suggestion for this?

Thanks a lot!
