### 2023-11-10 #6002 @faroit
@wergo to keep this review issue clean, please discuss unit test alternative in the respective issue
