### 2023-11-10 mei-friend/mei-friend#61 @wergo
@stefan-balke, we have started with this, but not elaborated carefully. We have used the Selenium framework for this. Would you have an alternative suggestion for this?
