### 2023-11-11 17a279e
Improve focus switching btw editor and notation
