### 2023-11-11 79f8573
Sort XML attributes ascending, but xml:id first
