### 2023-11-12 mei-friend/mei-friend#61 @stefan-balke
Hi @wergo,

for testing interaction and GUI elements, Selenium is one way to do it.
My suggestion aims more at the fundamental level which is easier to test, e.g., load an MXML file and make sure the MEI is what you expect, or save parameters and get what you expected it to save...or set or delete, not necessarily through user interaction. Could also be a call in a script you wrote.
