### 2023-11-13 33df13d
switch focus improvements
