### 2023-11-13 26c7fb5
Focus inside, not on, notation. Minor resizing fixes.
