### 2023-11-15 e3eca07
Add example openUrl unit test
