### 2023-11-15 ebeb80c
Refactor to add building blocks into testUtils.js
