### 2023-11-16 d4e849f
first attempt at settings test - WIP
