### 2023-11-16 c2efbcf
use assertions
