### 2023-11-17 #6002 @rlskoeser
Hi, @wergo, some feedback for you on the software paper. Your statement of need is clear, but I think a couple of aspects could be stated more explicitly:
- What is the intended audience? Is it primarily/only for scholarly use, and could you clarify that more, e.g. does that include research and education? You mention it in the intro that it provides a friendly, interactive interface - do you hope it will be useful for a more general public, or is MEI to technical/specialized for that?  You also mention uptake by the MEI community; does that overlap exactly with your target audience or do you hope it will have wider appeal?
- Please clarify the relationship to other available packages and solutions. You talk about MEI in comparison to MusicXML and other formats, but need to pay equal attention to your tool. You mention commercial notation software in passing, but don't name any - that may be fine, I suppose there are a lot of them and they serve a somewhat different purpose, but it could help to differentiate the goals a little more clearly. Am I understanding correctly that mei-friend is the only MEI editor of this type? If you so, you should definitely say that!
