### 2023-11-17 #6002 @rlskoeser
I've now worked through the installation instructions and was able to get mei-friend working locally - very cool to see the musical notation and the xml linked and editable. This seems like a really powerful and elegant interface, and I found myself wishing there was a tutorial to give me a little more of an orientation to the kinds of things I might be able to do, but I was able to poke around and try a few things out. When you update the installation documentation, you should make sure you link to the documentation you have on github pages about how to use it. A tutorial is a nice-to-have, not essential - you have thorough usage documentation already.

## installation

My feedback on the installation has significant overlap with what Stefan has already pointed out.

- Installation instructions should be easier to find; there should be a link in the readme. I don't feel strongly about whether it should be a markdown file or a page on your help pages, but it should be findable from both.
- I skipped the GitHub OAuth setup because I wanted to see what would happen. I didn't get any errors, but I don't think I've tried any functionality yet that requires it. It would be helpful to make it clearer what this is needed for and where it comes into play.
- It took me a couple of tries to get the codemirror download installed and put in the right place (at first I nested it incorrectly because you have an existing `CodeMirror/` directory already in the application). I'm not certain which codemirror components you're using (due to the code organization concerns that Stefan already raised), but it looks like you should be able to install it from npm; if so, that would be preferable to a git submodule.
- Consider making the application installable as a python package, at least from github even if you don't want to publish it on pypi. You can do this by adding a [pyproject.toml](https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/) file that describes your package, and includes the python requirements (this replaces requirements.txt). You will need to do configure the package to make sure the necessary non-python static files are included, and I think an npm install would still be needed for the js dependencies.
- I noticed you have a package.json file under an npm directory, but I'm not sure if you're using it; I see you also have a bunch of js files and licenses in a `deps` module. I recommend installing these dependencies from npm where possible.
- Your installation documentation includes an outdated reference bout setting the js mimetype, it looks like you've already added that.
- For the production deployment, I suggest linking to the default Flask deploy instructions and then adding any adjustments or configurations that are specific to your application.  You could do something similar for the python virtualenv instructions.

I second Stefan's comments about the code organization, developer documentation, and tests.

## community guidelines

I'm glad to see you're seeing issue templates, that's great. You might think about revamping your public repertoire submission template so that it looks more like a form with sections, and documentation with each section instead of all at the top. (Basically, structure them a bit more like the bug and feature request templates).

GitHub has documentation and examples for making a contributing guidelines document: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

I like that you have
