### 2023-11-18 #6002 @faroit
@rlskoeser Thanks for your in depth review.

Can you please open issues in the mei-friend repository and link back to this issue here so that we can better track down the progress. As I understand, some of your review points are just suggestions, so feel free to only convert "mandatory" points issues. 🙏
