+++ app/static/lib/drag-selector.js
@@ -192,10 +192,10 @@ export function addDragSelector(v, vp) {
           v.selectedElements.push(newEl);
         }
       });
-      console.debug('Drag-Selector selected elements: ', v.selectedElements)
+      // console.debug('Drag-Selector selected elements: ', v.selectedElements)
 
       if (!found) {
-        console.debug('Drag-Selector latest element: ', latest.el.classList.item(0));
+        // console.debug('Drag-Selector latest element: ', latest.el.classList.item(0));
         v.showMissingIdsWarning(latest.el.classList.item(0));
       }
       v.updateHighlight();
