+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Keyboard shortcut (`SHIFT + N`) to convert selected notes to rests and rests to notes (#57)
 * Keyboard shortcut (`.`) to add or remove `dots="1"` to notes, chords, etc.
 * Keyboard shortcut (`C`) to embed selected notes inside a new chord element, and vice versa (remove chord and leave notes)
+* Misleading warning when clicking on elements without ID fixed (#52)
 
 ### 1.0.4 patch
 * Tweak CSS select and translations (German)
