### 2023-11-20 26b0631
Make GH config optional
