### 2023-11-20 mei-friend/mei-friend#65 @musicog
Gracefully handle absence of GH OAuth application
As raised in by the reviews in https://github.com/openjournals/joss-reviews/issues/6002

It should be possible to run an mei-friend instance without having to set up a new GitHub OAuth application. The interface should gracefully detect and handle the absence of a server-side GH configuration by simply not showing the Login / Github menu button. The remainder of the application should otherwise work as normal.
