### 2023-11-20 #6002 @rlskoeser
@faroit thanks for the reminder, I'll create issues to document these (at least the essentials) when I'm able to do so.

@faroit question for you: how much of the functionality should I test? Should I try everything covered in the readme? I think the major points I haven't tried are the portions using github auth and annotation.
