### 2023-11-21 a0ae3bc
Complete base-40 system
