### 2023-11-21 7702963
Move pitch by key signature in base-40 system
