+++ app/static/lib/editor.js
@@ -923,7 +923,7 @@ export function toggleArtic(v, cm, artic = 'stacc') {
  * Shifts element (rests, note) up/down diatonically by pitch name (1 or 7 diatonic steps)
  * @param {Viewer} v
  * @param {CodeMirror} cm
- * @param {int} deltaPitch (-1, -12, +2)
+ * @param {int} deltaPitch (-1, -7, +2)
  * @param {boolean} shiftChromatically
  */
 export function shiftPitch(v, cm, deltaPitch = 0, shiftChromatically = false) {
@@ -2446,7 +2446,7 @@ function toggleArticForNote(note, artic, xmlIdStyle) {
 /**
  * Modifies an element's pitch up and down (i.e. manipulating @oct, @pname)
  * @param {Element} el
- * @param {number} deltaPitch (diatonic steps or chromatic steps)
+ * @param {number} deltaPitch (diatonic steps or chromatic steps, if flag below is set)
  * @param {boolean} shiftChromatically
  * @returns {Element} element with modified attributes
  * NOTE: when shifting chromatically, only semitones (+/-1) make sense!
@@ -2454,7 +2454,7 @@ function toggleArticForNote(note, artic, xmlIdStyle) {
 function pitchMover(v, el, deltaPitch, shiftChromatically = false) {
   let octValue = 4;
   let pnameValue = 'c';
-  let accidValue = '';
+  let accidValue = 'n';
   let octAttr;
   let pnameAttr;
   if (['note'].includes(el.nodeName)) {
@@ -2470,70 +2470,117 @@ function pitchMover(v, el, deltaPitch, shiftChromatically = false) {
   if (el.hasAttribute(octAttr)) octValue = parseInt(el.getAttribute(octAttr));
   if (el.hasAttribute(pnameAttr)) pnameValue = el.getAttribute(pnameAttr);
 
-  let affectedNotes = dutils.getAffectedNotesFromKeySig(el);
-
-  let pi;
-  if (shiftChromatically) {
-    accidValue =
-      el.getAttribute('accid') || el.querySelector('[accid]')?.getAttribute('accid') || el.getAttribute('accid.ges');
-
-    if (accidValue) accidValue = accidValue.slice(0, 1); // take only first character
-
-    // remove all possible accid information
-    el.removeAttribute('accid');
-    el.removeAttribute('accid.ges');
-    let accidChild = el.querySelector('accid');
-    if (accidChild) removeWithTextnodes(accidChild);
-
-    let b40i = att.pitchToBase40(pnameValue, accidValue, octValue);
-
-    pi = att.pnames.indexOf(pnameValue); // index in scale
-    let pitchesToBeAltered = att.sharps.slice(0, 5);
-    let accidSign = 's';
-    if (deltaPitch < 0) {
-      pitchesToBeAltered = att.flats.slice(0, 5);
-      accidSign = 'f';
+  // handle rests quickly
+  if (octAttr === 'oloc') {
+    let pi = att.pnames.indexOf(pnameValue) + deltaPitch;
+    // secure octave transistion
+    if (pi > att.pnames.length - 1) {
+      pi -= att.pnames.length;
+      octValue++;
+    } else if (pi < 0) {
+      pi += att.pnames.length;
+      octValue--;
     }
-    if (pitchesToBeAltered.includes(pnameValue)) {
-      if (accidValue === accidSign) {
-        pi += deltaPitch;
-        accidValue = '';
-      } else if (accidValue) {
-        accidValue = '';
+    el.setAttribute(octAttr, octValue);
+    el.setAttribute(pnameAttr, att.pnames[pi]);
+    return el;
+  } // treat rests
+
+  // get key signature and create scale steps in base-40 system
[…]
