### 2023-11-21 8f3f719
Fix notation and encoding panel width in resizer
