### 2023-11-21 dcfae74
finish settingsChangeTheme test - TODO figure out ECONNREFUSED
