### 2023-11-22 340a254
Improve base-40 code
