### 2023-11-22 39069a3
Set up split button behaviour
