### 2023-11-22 d106b36
Show/hide accid/.ges by keySig at shifting pitch
