### 2023-11-22 6c86ed5
Change affected notes to base-40
