### 2023-11-22 70853ff
Reduce verbosity in pitch mover
