### 2023-11-23 61b86d1
Update to 1.0.6, changelog, and date
