### 2023-11-23 8d79624
Streamlined addMarkup event integration
