+++ app/static/lib/main.js
@@ -71,6 +71,7 @@ import {
   generateSectionSelect,
 } from './control-menu.js';
 import { clock, unverified, xCircleFill } from '../css/icons.js';
+import { keymap } from '../keymaps/default-keymap.js';
 import { setCursorToId } from './utils.js';
 import { getInMeasure, navElsSelector, getElementAtCursor } from './dom-utils.js';
 import { addDragSelector } from './drag-selector.js';
@@ -164,7 +165,6 @@ const defaultCodeMirrorOptions = {
 
 // add all possible facsimile elements
 att.attFacsimile.forEach((e) => defaultVerovioOptions.svgAdditionalAttribute.push(e + '@facs'));
-const defaultKeyMap = `${root}keymaps/default-keymap.json`;
 const sampleEncodingsCSV = `${root}sampleEncodings/sampleEncodings.csv`;
 let freshlyLoaded = false; // flag to ignore a cm.on("changes") event on file load
 
@@ -381,8 +381,8 @@ async function suspendedValidate(text, updateLinting, options) {
 // when initial page content has been loaded
 document.addEventListener('DOMContentLoaded', function () {
   // disable GitHub menu if server-side configuration not available
-  if(!gitEnabled) { 
-    document.getElementById("GithubButton").disabled = true;
+  if (!gitEnabled) {
+    document.getElementById('GithubButton').disabled = true;
   }
 
   translator = new Translator();
@@ -770,7 +770,7 @@ function completeInitialLoad() {
     doit = setTimeout(() => setOrientation(cm, '', '', v, storage), 500);
   };
 
-  setKeyMap(defaultKeyMap);
+  setKeyMap(keymap);
 
   // remove URL parameters from URL
   // TODO: check handleURLParamSelect() occurrences, whether removing search parameters has an effect there.
@@ -2365,60 +2365,57 @@ function fillInSampleEncodings() {
  * Sets keymap JSON information to target element and defines listeners.
  * It loads all bindings in `#notation` to document.body, and the platform-specific
  * to both notation and editor panels (i.e. friendContainer).
- * @param {string} keyMapFilePath
  */
-function setKeyMap(keyMapFilePath) {
+function setKeyMap() {
   if (platform.startsWith('mac')) {
-    document.body.classList.add('platform-darwin');
+    document.body.classList.add('platform-darwin-all');
+    document.getElementById('notation').classList.add('platform-darwin-notation');
   }
   if (platform.startsWith('win')) {
-    document.body.classList.add('platform-win32');
+    document.body.classList.add('platform-win32-all');
+    document.getElementById('notation').classList.add('platform-win32-notation');
   }
   if (platform.startsWith('linux')) {
-    document.body.classList.add('platform-linux');
+    document.body.classList.add('platform-linux-all');
+    document.getElementById('notation').classList.add('platform-linux-notation');
   }
-  fetch(keyMapFilePath)
-    .then((resp) => {
-      return resp.json();
-    })
-    .then((keyMap) => {
-      // iterate all keys (element) in keymap.json
-      for (const [key, value] of Object.entries(keyMap)) {
-        document.querySelectorAll(key).forEach((el) => {
-          el.setAttribute('tabindex', '-1');
-          el.addEventListener('keydown', (ev) => {
-            if (['pagination2', 'selectTo', 'selectFrom', 'selectRange'].includes(document.activeElement.id)) {
-              return;
-            }
 
-            // at each keystroke: update cmd2key (CTRL on Mac, ALT on WIN/Linux)
-            v.cmd2KeyPressed = platform.startsWith('mac') ? ev.ctrlKey : ev.altKey;
+  // iterate all keys (element) in keymap.json
[…]
