+++ app/static/keymaps/default-keymap.js
@@ -1,0 +1,220 @@
+// default keymap 
+
+export const keymap = {
+  // gets mapped to notation only (for all operating systems)
+  '#notation': {
+    x: 'invertPlacement',
+    'shift-x': 'betweenPlacement',
+    v: 'addVerticalGroup',
+    right: 'nextNote',
+    left: 'previousNote',
+    up: 'layerUp',
+    down: 'layerDown',
+    'shift-up': 'shiftPitchChromaticallyUp',
+    'shift-down': 'shiftPitchChromaticallyDown',
+    'shift-alt-up': 'shiftPitchNameUp',
+    'shift-alt-down': 'shiftPitchNameDown',
+    'shift-left': 'decreaseDuration',
+    'shift-right': 'increaseDuration',
+    delete: 'delete',
+    backspace: 'delete',
+    pagedown: 'nextPage',
+    pageup: 'previousPage',
+    home: 'firstPage',
+    end: 'lastPage',
+
+    n: 'addNote',
+    'shift-n': 'convertNoteToRest',
+    c: 'convertToChord',
+    '.': 'toggleDots',
+
+    'shift-+': 'addDoubleSharp',
+    '+': 'addSharp',
+    '=': 'addNatural',
+    '-': 'addFlat',
+    'shift--': 'addDoubleFlat',
+
+    s: 'addSlur',
+    t: 'addTie',
+    h: 'addCresHairpin',
+    'shift-h': 'addDimHairpin',
+    f: 'addFermata',
+    i: 'addDirective',
+    d: 'addDynamics',
+    'shift-t': 'addTempo',
+    a: 'addArpeggio',
+    'ctrl-a': 'addArpeggioDown',
+    g: 'addGlissando',
+    p: 'addPedalDown',
+    'shift-p': 'addPedalUp',
+    m: 'addMordent',
+    'shift-m': 'addMordentUpper',
+    l: 'addTrill',
+    r: 'addTurn',
+    'shift-r': 'addTurnLower',
+
+    'shift-s': 'toggleStacc',
+    'shift-v': 'toggleAccent',
+    'shift-e': 'toggleTenuto',
+    'shift-a': 'toggleMarcato',
+    'shift-i': 'toggleStacciss',
+    'shift-y': 'toggleSpicc',
+
+    b: 'addBeam',
+    'shift-b': 'addBeamSpan',
+    u: 'addSupplied',
+    'ctrl-u': 'addSuppliedArtic',
+    'ctrl-shift-u': 'addSuppliedAccid',
+    o: 'addOctave8Above',
+    'ctrl-o': 'addOctave8Below',
+    'shift-o': 'addOctave15Above',
+    'ctrl-shift-o': 'addOctave15Below',
+
+    'shift-g': 'addGClefChangeBefore',
+    'shift-c': 'addCClefChangeBefore',
+    'shift-f': 'addFClefChangeBefore',
+
+    escape: 'escapeKeyPressed',
+
+    'shift-alt-f': 'indentSelection',
+
+    space: 'playPauseMidiPlayback',
+  },
+
+  // only mapped to notation under MAC OSX
+  '.platform-darwin-notation': {
+    'cmd-right': 'nextMeasure',
+    'cmd-left': 'previousMeasure',
+    'ctrl-cmd-right': 'nextPage',
+    'ctrl-cmd-left': 'previousPage',
+    'cmd-up': 'firstPage',
+    'cmd-down': 'lastPage',
+  },
+
+  // mapped to body under MAC OSX
[…]
