### 2023-11-23 2c1d862
Debug messaging enabled for switchFocus
