+++ app/static/lib/markup.js
@@ -65,6 +65,31 @@ export function readMarkup() {
 }
 
 /**
+ * Creates a list item object for markup elements based on some basic data
+ * @param {string} currentElementId xml:id of current element -> item.id
+ * @param {string} mElName element name -> item.type
+ * @param {Array[string]} correspElements ids of corresponding elements (including self) -> item.selection
+ * @returns {boolean} adding item was successful
+ */
+function xmlMarkupToListItem(currentElementId, mElName, correspElements) {
+  // one addMarkupAction might result in multiple markup elements
+  // e.g. when selecting notes and control events
+  // mostly because symbols in a score aren't necessarily close in the xml tree
+  // properties are similar to annotations
+  // id is the xml:id of the first markup element
+  // selection contains all markup elements created at the same time
+
+  const markupItem = {
+    id: currentElementId,
+    type: mElName,
+    isMarkup: true,
+    selection: correspElements,
+  };
+  let success = addListItem(markupItem);
+  return success;
+}
+
+/**
  * Returns for a markup item the page locations in the Verovio rendering
  * @param {Object} markupItem markupItem to situate
  * @returns {Promise} a promise containing the modified markupItem
@@ -184,35 +209,10 @@ export function addMarkup(event) {
   let mElName = eventTarget.dataset.elName;
   let attrName = eventTarget.dataset.selection;
   if (!att.modelTranscriptionLike.includes(mElName)) return;
-  addTranscriptionLikeElement(v, cm, attrName, mElName);
+  addMarkupToXML(v, cm, attrName, mElName);
   //let successfullyAdded = xmlMarkupToListItem(v.selectedElements, mElName);
   // Manually updating the item list is not necessary because refreshing the code in the editor triggers readMarkup()
   refreshAnnotationsList();
-}
-
-/**
- * Creates a list item object for markup elements based on some basic data
- * @param {string} currentElementId xml:id of current element -> item.id
- * @param {string} mElName element name -> item.type
- * @param {Array[string]} correspElements ids of corresponding elements (including self) -> item.selection
- * @returns {boolean} adding item was successful
- */
-function xmlMarkupToListItem(currentElementId, mElName, correspElements) {
-  // one addMarkupAction might result in multiple markup elements
-  // e.g. when selecting notes and control events
-  // mostly because symbols in a score aren't necessarily close in the xml tree
-  // properties are similar to annotations
-  // id is the xml:id of the first markup element
-  // selection contains all markup elements created at the same time
-
-  const markupItem = {
-    id: currentElementId,
-    type: mElName,
-    isMarkup: true,
-    selection: correspElements,
-  };
-  let success = addListItem(markupItem);
-  return success;
 }
 
 /**
@@ -234,12 +234,11 @@ function xmlMarkupToListItem(currentElementId, mElName, correspElements) {
  * @param {string} mElName name of markup element to apply
  * @returns
  */
-export function addTranscriptionLikeElement(v, cm, attrName = 'none', mElName = 'supplied') {
+function addMarkupToXML(v, cm, attrName = 'none', mElName, multiLayerContent = []) {
   v.loadXml(cm.getValue());
   v.selectedElements = speed.filterElements(v.selectedElements, v.xmlDoc);
   v.selectedElements = utils.sortElementsByScorePosition(v.selectedElements);
   if (v.selectedElements.length < 1) return;
-  if (!att.modelTranscriptionLike.includes(mElName)) return;
[…]
