### 2023-11-23 f24b3fd
Secure switchFocus under all conditions
