+++ app/static/lib/viewer.js
@@ -844,7 +844,7 @@ export default class Viewer {
    * @param {CodeMirror} cm
    */
   switchFocusBetweenNotationAndEncoding(cm) {
-    let vrvP = document.getElementById('verovio-panel');
+    let vrvP = document.getElementById('notation');
     let encoding = document.getElementById('encoding');
     if (document.activeElement.closest('#notation')) {
       console.debug('Viewer.switchFocusBetweenNotationAndEncoding(): Switching now to encoding.', cm);
