### 2023-11-23 875b09a
Some code polishing
