### 2023-11-24 4c19295
Basic support for choice
Needs refinement
