### 2023-11-24 e68f84b
Add dummy comment to empty alternative encoding after creation
