### 2023-11-24 82114c0
Fix selectChoice to allow highlighting in rendering
