### 2023-11-26 74a5ba9
Mark individual elements with SHIFT + arrows
