+++ app/static/lib/main.js
@@ -1475,6 +1475,12 @@ export let cmd = {
   layerUp: () => v.navigate(cm, 'layer', 'upwards'),
   layerDown: () => v.navigate(cm, 'layer', 'downwards'),
   notationTop: () => setOrientation(cm, 'top', '', v, storage),
+  markNextNote: () => v.navigate(cm, 'note', 'forwards', true),
+  markPreviousNote: () => v.navigate(cm, 'note', 'backwards', true),
+  markNextMeasure: () => v.navigate(cm, 'measure', 'forwards', true),
+  markPreviousMeasure: () => v.navigate(cm, 'measure', 'backwards', true),
+  markLayerUp: () => v.navigate(cm, 'layer', 'upwards', true),
+  markLayerDown: () => v.navigate(cm, 'layer', 'downwards', true),
   notationBottom: () => setOrientation(cm, 'bottom', '', v, storage),
   notationLeft: () => setOrientation(cm, 'left', '', v, storage),
   notationRight: () => setOrientation(cm, 'right', '', v, storage),
