### 2023-11-27 2c7afce
Read choice from xml
