### 2023-11-27 bd730d3
Delete markup keeps content or content of first child
