### 2023-11-27 ff07f9f
Keep choice in speed-mode to allow navigation
