### 2023-11-27 9a8cb66
Fix: multiLayeredContent is conditional
