### 2023-12-01 6606660
Update lang.ja.js
- Some mistranslations and incomprehensible expressions were corrected. Checked up to last line.
- Add "  // Warning for missing ids" that is not included lang.ja.js

(Sorry, but I'm not familiar to part of mei-friends features, so I cannot rule out the possibility of imperfection.)
