+++ app/static/lang/lang.ja.js
@@ -231,22 +231,22 @@ export const lang = {
   pdfPreviewNormalModeTitle: { text: 'PDFに保存するページ範囲を選択してください。' },
 
   // ファクシミリパネル
-  facsimileIcon: { description: 'ファクシミリパネル' },
+  facsimileIcon: { description: 'ファクシミリ画像パネル' },
   facsimileDecreaseZoomButton: { description: '画像を縮小' },
   facsimileZoom: { description: '画像のサイズを調整' },
   facsimileIncreaseZoomButton: { description: '画像を拡大' },
   facsimileFullPageLabel: { text: '全体表示', description: 'ファクシミリ画像の全体表示' },
   facsimileFullPageCheckbox: { description: 'ファクシミリ画像の全体表示' },
-  facsimileShowZonesLabel: { text: 'ゾーンボックスを表示', description: 'ファクシミリのゾーンボックスを表示' },
-  facsimileShowZonesCheckbox: { description: 'ファクシミリのゾーンボックスを表示' },
-  facsimileEditZonesCheckbox: { description: 'ファクシミリのゾーンを編集' },
-  facsimileEditZonesLabel: { text: 'ゾーンを編集', description: 'ファクシミリのゾーンを編集' },
-  facsimileCloseButton: { description: 'ファクシミリパネルを閉じる' },
-  facsimileDefaultWarning: { text: '表示するファクシミリコンテンツはありません。' },
+  facsimileShowZonesLabel: { text: '領域ボックスを表示', description: 'ファクシミリ画像の領域ボックスを表示' },
+  facsimileShowZonesCheckbox: { description: 'ファクシミリ画像の領域ボックスを表示' },
+  facsimileEditZonesCheckbox: { description: 'ファクシミリ画像の領域を編集' },
+  facsimileEditZonesLabel: { text: '領域を編集', description: 'ファクシミリ画像の領域を編集' },
+  facsimileCloseButton: { description: 'ファクシミリ画像パネルを閉じる' },
+  facsimileDefaultWarning: { text: '表示するファクシミリ画像はありません。' },
   facsimileNoSurfaceWarning: {
     text: 'このページにはsurface要素が見つかりません。\n(最初のpb要素がない可能性があります。)',
   },
-  facsimileNoZonesFullPageWarning: { text: 'ゾーンのないファクシミリは全体表示モードでのみ表示されます。' },
+  facsimileNoZonesFullPageWarning: { text: '領域指定のないファクシミリは全体表示モードでのみ表示されます。' },
   facsimileImgeNotLoadedWarning: { text: '画像を読み込めませんでした' },
 
   // Drag'n'drop
@@ -327,11 +327,11 @@ export const lang = {
   // Restore Solid session overlay
   solidExplanation: {
     description:
-      'Solidはリンクトデータ（RDF）を使用してスタンドオフ注釈を作成するための分散型プラットフォームです。Solidにログインしてリンクトデータアノテーションを作成できるようになります。',
+      'SolidはLinked Data（RDF）を使用してスタンドオフ注釈を作成するための分散型プラットフォームです。Solidにログインしてアノテーションを作成できるようになります。',
   },
-  solidProvider: { description: 'Solid Identity Provider (IdP)を選択または独自のIdPを指定してください。' },
+  solidProvider: { description: 'Solid Identity Provider (IdP)を選択、または独自のIdPを指定してください。' },
   solidLoginBtn: { text: 'ログイン' },
-  solidOverlayCancel: { html: 'Solidセッションを復元中 - <span>esc</span>を押すか、こちらをクリックしてキャンセル' },
[…]
