### 2023-12-01 mei-friend/mei-friend#58 @rlskoeser
additional comments about installation instructions based on my review for JOSS https://github.com/openjournals/joss-reviews/issues/6002:
- Install instructions should be linked from the readme and probably also from the documentation site; I recommend keeping it as a markdown file in the github repo for ease of findability and maintaining with the code.
- Make it clearer which setup steps are optional and what the impact is of skipping them; e.g., if you don't setup a GitHub app, will things break or will there just be some features that are not available to you?
- Remove the outdated reference about setting the js mimetype (it looks like this has been implemented already in the code).
- For production deploy instructions, I suggest linking to standard Flask deploy instructions and then adding any notes about adjustments or configurations specific to your application.
