### 2023-12-01 mei-friend/mei-friend#68 @rlskoeser
Add contributor guidelines 
per JOSS Review https://github.com/openjournals/joss-reviews/issues/6002

You need to provide contributor guidelines:
> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

GitHub has documentation and examples for making a contributing guidelines document, which should help you with getting started: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
