### 2023-12-01 mei-friend/mei-friend#69 @rlskoeser
consider making the application installable as a python package
Consider making the application installable as a python package, at least from GitHub even if you don't want to publish it on pypi. You can do this by adding a [pyproject.toml](https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/) file that describes your package, and includes the python requirements (this replaces requirements.txt). You will need to do configure the package to make sure the necessary non-python static files are included (js/npm dependencies would need to be handled separately).

suggestion from JOSS review https://github.com/openjournals/joss-reviews/issues/6002
