### 2023-12-01 mei-friend/mei-friend#70 @rlskoeser
consider using npm for js dependency management
I noticed you have a package.json file under an npm directory, but it doesn't seem like you're using it (there was no mention in the installation instructions). I see you also have a bunch of js files and licenses in a deps module. I recommend installing these dependencies from npm where possible.

It looks like it is also possible to install CodeMirror components from npm instead of using a git submodule. An npm install is preferable to git submodule because it's a more standard setup and lets you easily specify the appropriate versions.

You may also want to consider webpack or vite for compiling and minimizing the js code, but that would be a secondary step building on npm install, and potentially a more complicated change to the mei-friend codebase. There are probably existing flask plugins that would help with this.

suggestion from JOSS review https://github.com/openjournals/joss-reviews/issues/6002
