### 2023-12-01 mei-friend/mei-friend#71 @rlskoeser
suggestion: revise public repertoire submission template to make it more like a form
Consider revamping your public repertoire submission template so that it looks more like a form with sections, and documentation with each section instead of all at the top. Basically, structure them a bit more like the bug and feature request templates you are using so it is easier for people to fill out the necessary information, and have instructions in place as they do.

suggestion from JOSS review https://github.com/openjournals/joss-reviews/issues/6002
