### 2023-12-04 aa32123
Update public_repertoire.md to address #71
