### 2023-12-04 mei-friend/mei-friend#71 @wergo
Thank you for this suggestion!

Addressed in [aa32123](https://github.com/mei-friend/mei-friend/commit/aa32123970e75c9bd327e7d9b1ca7a93b15d80ad).
