### 2023-12-04 e070b2f
Add public repertoire test, WIP
