### 2023-12-04 6c23eda
Add public repertoire test utils (WIP)
