### 2023-12-05 97234c5
Update lang.ja.js
