+++ app/static/lang/lang.ja.js
@@ -14,65 +14,62 @@ export const lang = {
   },
   splashBody: {
     html: `
-      <p>
-        mei-friendは<a href="https://music-encoding.org">音楽エンコーディング</a>のエディタで、
-        <a href="https://mdw.ac.at" target="_blank">mdw - ウィーン音楽・演劇アート大学</a>でホストされています。
-        詳細な情報については<a href="https://mei-friend.github.io" target="_blank">詳細なドキュメンテーション</a>
-        をご覧ください。
+   <p>
+      mei-friendは、<a href="https://music-encoding.org">Music Encoding Initiative</a>のエディタで、
+      <a href="https://mdw.ac.at" target="_blank">ウィーン国立音楽大学</a>がホストするプラットフォームです。
+      詳細な情報は<a href="https://mei-friend.github.io" target="_blank">こちら</a>をご覧ください。
       </p>
       <p>
-        mei-friendはブラウザベースのアプリケーションですが、個人データ（編集中のエンコード、アプリケーションの設定、および
-        ログイン詳細など）はブラウザの
-        <a href="https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage" target="_blank"
-          >ローカルストレージ</a
-        >に保存され、サーバーに送信または保存されません。
+      mei-friendはブラウザベースのアプリケーションでありながら、個人データ（編集中の楽譜データ、アプリケーションの設定、
+      およびログイン詳細など）はブラウザの<a href="https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage" target="_blank" >ローカルストレージ</a>
+      に保存され、サーバーには送信または保存されません。
+
       </p>
       <p>
-        データはGitHubに明示的にリクエストした場合にのみ送信されます（たとえば、GitHubにログインし、GitHubのリポジトリからエンコードを読み込む、またはGitHub Actionのワークフローをリクエストする場合など）。同様に、選択したSolidプロバイダーに対しても、明示的にリクエストした場合にのみデータが送信されます（たとえば、Solidにログインし、スタンドオフの注釈を読み込むまたは保存する場合など）。
+
+      データは明示的なリクエスト時のみGitHubに送信されます。同様に、選択したSolidプロバイダーに対しても明示的なリクエスト時にのみデータが送信されます。
+
       </p>
       <p>
-        無記名の使用統計情報を収集するために<a href="https://matomo.org/" target="_blank">Matomo</a>
-        を使用しています。これには切り捨てられたIPアドレス（国のレベルでの地理的位置情報を許可しますが、それ以上の識別情報は含まれません）、ブラウザとオペレーティングシステム、リファリングウェブサイトからのアクセス先、訪問の時間と期間、および訪問したページが含まれます。この情報はmdw - ウィーン音楽・演劇アート大学のサーバーで実行されているMatomoインスタンスに保存され、第三者とは共有されません。
+
+      mei-friendは<a href="https://matomo.org/" target="_blank">Matomo</a>を使用して匿名の使用統計情報を収集しています。これには省略されたIPアドレス
+      （国のレベルでの地理的位置情報を許可しますが、それ以上の識別情報は含まれません）、ブラウザ、オペレーティングシステム、参照元ウェブサイト、訪問の時間と期間、
+      および訪問したページが含まれます。この情報はウィーン国立音楽大学のサーバーで実行されているMatomoインスタンスに保存され、第三者とは共有されることはありません。
       </p>
       <p>
-        Verovioツールキットは<a href="https://verovio.org" target="_blank">https://verovio.org</a>からロードされ、
-        <a href="https://rism.digital/" target="_blank">RISM Digital Switzerland</a>がホストしています。 
-        これにより、mei-friendは常に最新のツールキットバージョンを使用できるだけでなく、
[…]
