+++ tests/testUtils.js
@@ -33,7 +33,7 @@ module.exports = {
   },
   openSettingsTab: async function (driver, tab) {
     try {
-      const hideSettingsButton = driver.findElement(By.id('hideSettingsButton'))
+      const hideSettingsButton = driver.findElement(By.id('hideSettingsButton'));
       if (await hideSettingsButton.isDisplayed()) {
         console.log('Settings menu was already open!');
       } else {
@@ -41,7 +41,7 @@ module.exports = {
         console.log('Settings menu was closed - clicking to open!');
       }
       if (tab) {
-        console.log("trying to click tab", tab)
+        console.log('trying to click tab', tab);
         driver.findElement(By.id(tab)).click();
         console.log('Settings tab clicked: ', tab);
       }
@@ -59,17 +59,17 @@ module.exports = {
       await this.selectFromDropdown(driver, 'sampleEncodingsComposer', composer);
       await driver.wait(until.elementLocated(By.CSS_SELECTOR, `select[value='${composer}'`, 5000));
       await this.selectFromDropdown(driver, 'sampleEncodingsEncoding', encoding);
-      let sampcomp = await driver.findElement(By.id('sampleEncodingsComposer'))
+      let sampcomp = await driver.findElement(By.id('sampleEncodingsComposer'));
       let sampcompval = await sampcomp.getAttribute('value');
-      console.log("sampcompval", sampcompval)
+      console.log('sampcompval', sampcompval);
       driver.findElement(By.id('openUrlButton')).click();
       console.log('OpenURL button clicked!');
     } catch (error) {
       console.error('Error opening URL via UI: ', error);
     }
   },
-  selectFromDropdown: async function (driver, dropdownId, option, eventType = "change") {
-    console.log("Attempting to select from dropdown: ", dropdownId, option)
+  selectFromDropdown: async function (driver, dropdownId, option, eventType = 'change') {
+    console.log('Attempting to select from dropdown: ', dropdownId, option);
     try {
       await driver.executeScript(`
         const selectElement = document.getElementById("${dropdownId}");
