+++ tests/test.js
@@ -1,0 +1,2 @@
+const utils = require('./testUtils.js');
+open 'test_openUrl.js'
