### 2023-12-13 882bc8a
Bug fix for #73 to handle unknown language code
