### 2023-12-13 5dd7338
Hotfix for #73 to handle unsupported language code
