### 2023-12-13 ed27762
Add content of choice to details
