### 2023-12-14 a17f786
Added choice selector to Verovio control menu
No special filter for speed mode yet
