### 2023-12-15 10e6833
Update choice select according to current document state
