+++ app/static/lib/main.js
@@ -900,7 +900,7 @@ async function vrvWorkerEventsHandler(ev) {
       drawRightFooter();
       document.getElementById('statusBar').innerHTML = `Verovio ${tkVersion} ${translator.lang.verovioLoaded.text}.`;
       setBreaksOptions(tkAvailableOptions, defaultVerovioOptions.breaks);
-      setChoiceOptions();
+      setChoiceOptions('');
       if (!storage.supported || !meiFileName) {
         // open default mei file
         openFile();
@@ -960,6 +960,9 @@ async function vrvWorkerEventsHandler(ev) {
       } else if (v.speedMode && bs === 'auto' && Object.keys(v.pageBreaks).length > 0) {
         v.pageCount = Object.keys(v.pageBreaks).length;
       }
+      //update choiceSelect
+      let cs = document.getElementById('choiceSelect').value;
+      setChoiceOptions(cs);
       // update only if still same page
       if (v.currentPage === ev.data.pageNo || ev.data.forceUpdate || ev.data.computePageBreaks || v.pdfMode) {
         if (ev.data.forceUpdate) {
