+++ app/static/lang/lang.ca.js
@@ -282,6 +282,8 @@ export const lang = {
   breaksSelectLine: { text: 'Sistema' },
   breaksSelectEncoded: { text: 'Sistema i pàgina' },
   breaksSelectSmart: { text: 'Intel·ligent' },
+  choiceSelect: { description: 'Trieu el contingut mostrat pels elements de selecció' },
+  choiceDefault: { text: '(opció per defecte)' },
   updateControlsLabel: {
     text: 'Actualitzar',
     description: "Comportament d'actualització dels controls de la notació després de canvis en l'encoding",
