+++ app/static/lang/lang.ja.js
@@ -271,6 +271,8 @@ export const lang = {
   breaksSelectLine: { text: 'システム' },
   breaksSelectEncoded: { text: 'システムとページ' },
   breaksSelectSmart: { text: 'スマート' },
+  choiceSelect: { description: '選択要素の表示コンテンツを選択します' },
+  choiceDefault: { text: '(デフォルトの選択)' },
   updateControlsLabel: { text: '更新', description: 'エンコード変更後の楽譜の制御更新動作' },
   liveUpdateCheckbox: { description: 'エンコード変更後に楽譜を自動的に更新' },
   codeManualUpdateButton: { description: '楽譜を手動で更新' },
