### 2023-12-18 356c317
Add settings for default content of alternative encodings
+ convenience function: when copy, change view and select newly copied elements
