### 2023-12-18 a2fbe80
Add new setting for alternative encodings to language packs
