+++ app/static/lang/lang.it.js
@@ -839,6 +839,11 @@ export const lang = {
     text: 'Mostra elementi di markup editoriale',
     description: 'Evidenzia tutti gli elementi contenuti dagli elementi di markup editoriale',
   },
+  alternativeVersionContent: {
+    text: 'Seleziona il contenuto predefinito per le codifiche alternative',
+    description: 'Scegli se le codifiche alternative appena create sono vuote o copie della lettura originale',
+    labels: ['vuoto', 'copia'],
+  },
   suppliedColor: {
     text: 'Seleziona colore di evidenziazione per <supplied>',
     description: 'Seleziona colore di evidenziazione per <supplied>',
