### 2023-12-18 589b54a
Add settings for editoral content to language packs
