### 2023-12-18 68c067b
Fix bug for markup colours in language packs
