+++ app/static/lang/lang.uk.js
@@ -789,31 +789,31 @@ export const lang = {
     description: 'Вибір кольору підсвічування для <supplied>',
   },
   unclearColor: {
-    title: 'Вибір кольору підсвічування для <unclear>',
+    text: 'Вибір кольору підсвічування для <unclear>',
     description: 'Вибір кольору підсвічування для <unclear>',
   },
   sicColor: {
-    title: 'Вибір кольору підсвічування для <sic>',
+    text: 'Вибір кольору підсвічування для <sic>',
     description: 'Вибір кольору підсвічування для <sic>',
   },
   corrColor: {
-    title: 'Вибір кольору підсвічування для <corr>',
+    text: 'Вибір кольору підсвічування для <corr>',
     description: 'Вибір кольору підсвічування для <corr>',
   },
   origColor: {
-    title: 'Вибір кольору підсвічування для <orig>',
+    text: 'Вибір кольору підсвічування для <orig>',
     description: 'Вибір кольору підсвічування для <orig>',
   },
   regColor: {
-    title: 'Вибір кольору підсвічування для <reg>',
+    text: 'Вибір кольору підсвічування для <reg>',
     description: 'Вибір кольору підсвічування для <reg>',
   },
   addColor: {
-    title: 'Вибір кольору підсвічування для <add>',
+    text: 'Вибір кольору підсвічування для <add>',
     description: 'Вибір кольору підсвічування для <add>',
   },
   delColor: {
-    title: 'Вибір кольору підсвічування для <del>',
+    text: 'Вибір кольору підсвічування для <del>',
     description: 'Вибір кольору підсвічування для <del>',
   },
   // EDITOR SETTINGS / CODEMIRROR SETTINGS
