### 2023-12-19 fdbfaef
Start with generalization of choice select menu
