+++ app/static/lib/control-menu.js
@@ -658,6 +658,7 @@ export function setChoiceOptions(active) {
       choiceSelect[key] = new Option(el.label, el.value, false, false);
     }
     choiceSelect[key].id = el.id;
+    choiceSelect[key].data.setting = el.setting;
   });
 }
 
