### 2023-12-19 2938076
Fix start of generalization
