### 2023-12-20 e105e05
Modify luteconv splash screen text
