### 2023-12-20 aa91dc8
Fix italian language pack typo in key value
