### 2023-12-20 506dd2d
Fix missing key in Catalan language pack (thx Montserrat)
