### 2023-12-20 6233ef5
Luteconv splash screen texts for all other languages
