### 2023-12-20 9e46438
Update version number to 1.1.3
