+++ app/static/lib/viewer.js
@@ -367,11 +367,14 @@ export default class Viewer {
     if (fontSel) this.vrvOptions.font = fontSel.value;
     let bs = this.breaksSelect;
     if (bs) this.vrvOptions.breaks = bs.value;
-    let choiceSelect = this.choiceSelect.value;
-    if (choiceSelect != '') {
-      this.vrvOptions.choiceXPathQuery = ['./' + choiceSelect];
-    } else {
-      this.vrvOptions.choiceXPathQuery = [];
+    let choiceSelect = this.choiceSelect;
+    if (choiceSelect.selectedOptions.length > 0) {
+      let selectedChoice = choiceSelect.selectedOptions[0]; //always the first until type is changed to multiselect
+      if (selectedChoice.value != '') {
+        this.vrvOptions[selectedChoice.dataset.prop] = ['./' + selectedChoice.value];
+      } else {
+        this.vrvOptions[selectedChoice.dataset.prop] = [];
+      }
     }
 
     // update page dimensions, only if not in pdf mode
