### 2023-12-22 2508c36
Added changes in earlier versions to changelog
