+++ app/static/lib/viewer.js
@@ -1155,10 +1155,6 @@ export default class Viewer {
           checkedMarkup = document.getElementById('showMarkup').checked;
           setHighlightColorProperty('del', checkedMarkup, value, true);
           break;
-        case 'respSelect':
-          if (this.xmlDoc)
-            o.values = Array.from(this.xmlDoc.querySelectorAll('corpName[*|id]')).map((e) => e.getAttribute('xml:id'));
-          break;
         case 'controlMenuFontSelector':
           document.getElementById('engravingFontControls').style.display = value ? 'inherit' : 'none';
           break;
@@ -1401,9 +1397,6 @@ export default class Viewer {
             break;
           case 'delColor':
             setHighlightColorProperty('del', checkedMarkup, document.getElementById(option).value, false);
-            break;
-          case 'respSelect':
-            this.respId = document.getElementById('respSelect').value;
             break;
           case 'controlMenuFontSelector':
             document.getElementById('engravingFontControls').style.display = document.getElementById(
@@ -1908,6 +1901,10 @@ export default class Viewer {
     if (rs) {
       let value = rs.value;
       while (rs.length > 0) rs.options.remove(0);
+
+      //add empty default
+      rs.add(new Option('(none)', ''));
+
       let optEls = this.xmlDoc.querySelectorAll('corpName[*|id],persName[*|id]');
       optEls.forEach((el) => {
         if (el.closest('respStmt')) {
