### 2023-12-22 ca7db79
Complete integration of Markup Tools tab in language packs
