+++ app/static/lang/lang.en.js
@@ -479,38 +479,104 @@ export const lang = {
     text2: 'Please assign identifiers by selecting "Manipulate" -> "Re-render MEI (with ids)" and try again.',
   },
   // MARKUP MENU
-  addSuppliedText: { text: '<supplied>' },
-  addSupplied: { text: 'Selected elements' },
-  addSuppliedArtic: { text: 'Articulation' },
-  addSuppliedAccid: { text: 'Accidental' },
-  addUnclearText: { text: '<unclear>' },
-  addUnclear: { text: 'Selected elements' },
-  addUnclearArtic: { text: 'Articulation' },
-  addUnclearAccid: { text: 'Accidental' },
-  addSicText: { text: '<sic>' },
-  addSic: { text: 'Selected elements' },
-  addSicArtic: { text: 'Articulation' },
-  addSicAccid: { text: 'Accidental' },
-  addCorrText: { text: '<corr>' },
-  addCorr: { text: 'Selected elements' },
-  addCorrArtic: { text: 'Articulation' },
-  addCorrAccid: { text: 'Accidental' },
-  addOrigText: { text: '<orig>' },
-  addOrig: { text: 'Selected elements' },
-  addOrigArtic: { text: 'Articulation' },
-  addOrigAccid: { text: 'Accidental' },
-  addRegText: { text: '<reg>' },
-  addReg: { text: 'Selected elements' },
-  addRegArtic: { text: 'Articulation' },
-  addRegAccid: { text: 'Accidental' },
-  addAddText: { text: '<add>' },
-  addAdd: { text: 'Selected elements' },
-  addAddArtic: { text: 'Articulation' },
-  addAddAccid: { text: 'Accidental' },
-  addDelText: { text: '<del>' },
-  addDel: { text: 'Selected elements' },
-  addDelArtic: { text: 'Articulation' },
-  addDelAccid: { text: 'Accidental' },
+  respSelect: {
+    text: 'Select markup responsibility',
+    description: 'Select responsibility id',
+  },
+  alternativeEncodingsGrp: {
+    text: 'Alternative encodings',
+    description: 'Markup elements that contain multiple versions.',
+  },
+  addChoiceText: {
+    text: '<choice>',
+    description: 'Groups a number of alternative encodings for the same point in a text. ',
+  },
+  choiceSicCorr: { text: 'sic | corr', description: 'Put selection in <sic> and add <corr>.' },
+  choiceCorrSic: { text: 'corr | sic', description: 'Put selection in <corr> and add <sic>.' },
+  choiceOrigReg: { text: 'orig | reg', description: 'Put selection in <orig> and add <reg>.' },
+  choiceRegOrig: { text: 'reg | orig', description: 'Put selection in <reg> and add <orig>.' },
+  choiceContentTarget: {
+    text: '(select content)',
+    description: 'First, select content for this element by hovering over <choice>.',
+  },
+  addChoice: { text: 'Selected elements', description: 'Add markup to selected elements.' },
+  addChoiceArtic: { text: 'Articulation', description: 'Add markup to articulations within selection.' },
+  addChoiceAccid: { text: 'Accidental', description: 'Add markup to accidentals within selection.' },
+  addSubstText: {
+    text: '<subst>',
+    description:
+      '(substitution) – Groups transcriptional elements when the combination is to be regarded as a single intervention in the text.',
+  },
+  substAddDel: { text: 'add | del', description: 'Put selection in <add> and add <del>.' },
+  substDelAdd: { text: 'del | add', description: 'Put selection in <del> and add <add>.' },
+  substContentTarget: {
+    text: '(select content)',
+    description: 'First, select content for this element by hovering over <subst>.',
+  },
[…]
