### 2023-12-22 9c8f4d2
Added translations for Markup Tools
