+++ app/static/lang/lang.ja.js
@@ -462,38 +462,131 @@ export const lang = {
     text2: '「操作」->「MEIの再レンダリング（ID付き）」を選択して識別子を割り当て、もう一度お試しください。',
   },
   // Markup menu マークアップメニュー
-  addSuppliedText: { text: '<supplied>' },
-  addSupplied: { text: '選択された要素' },
-  addSuppliedArtic: { text: '発音' },
-  addSuppliedAccid: { text: '偶然' },
-  addUnclearText: { text: '<unclear>' },
-  addUnclear: { text: '選択された要素' },
-  addUnclearArtic: { text: '発音' },
-  addUnclearAccid: { text: '偶然' },
-  addSicText: { text: '<sic>' },
-  addSic: { text: '選択された要素' },
-  addSicArtic: { text: '発音' },
-  addSicAccid: { text: '偶然' },
-  addCorrText: { text: '<corr>' },
-  addCorr: { text: '選択された要素' },
-  addCorrArtic: { text: '発音' },
-  addCorrAccid: { text: '偶然' },
-  addOrigText: { text: '<orig>' },
-  addOrig: { text: '選択された要素' },
-  addOrigArtic: { text: '発音' },
-  addOrigAccid: { text: '偶然' },
-  addRegText: { text: '<reg>' },
-  addReg: { text: '選択された要素' },
-  addRegArtic: { text: '発音' },
-  addRegAccid: { text: '偶然' },
-  addAddText: { text: '<add>' },
-  addAdd: { text: '選択された要素' },
-  addAddArtic: { text: '発音' },
-  addAddAccid: { text: '偶然' },
-  addDelText: { text: '<del>' },
-  addDel: { text: '選択された要素' },
-  addDelArtic: { text: '発音' },
-  addDelAccid: { text: '偶然' },
+  respSelect: {
+    text: 'マークアップの責任を選択',
+    description: '責任のIDを選択',
+  },
+  alternativeEncodingsGrp: {
+    text: '代替エンコーディング',
+    description: '複数のバージョンを含むマークアップ要素。',
+  },
+  addChoiceText: {
+    text: '<choice>',
+    description: 'テキスト内の同じポイントの複数の代替エンコーディングをグループ化します。',
+  },
+  choiceSicCorr: { text: 'sic | corr', description: '<sic> に選択を入れて <corr> を追加します。' },
+  choiceCorrSic: { text: 'corr | sic', description: '<corr> に選択を入れて <sic> を追加します。' },
+  choiceOrigReg: { text: 'orig | reg', description: '<orig> に選択を入れて <reg> を追加します。' },
+  choiceRegOrig: { text: 'reg | orig', description: '<reg> に選択を入れて <orig> を追加します。' },
+  choiceContentTarget: {
+    text: '(コンテンツを選択)',
+    description: '<choice> の上にホバーしてこの要素のためにコンテンツを選択します。',
+  },
+  addChoice: { text: '選択された要素', description: '選択された要素にマークアップを追加します。' },
+  addChoiceArtic: {
+    text: 'アーティキュレーション',
+    description: '選択範囲内のアーティキュレーションにマークアップを追加します。',
+  },
+  addChoiceAccid: { text: 'アクシデンタル', description: '選択範囲内のアクシデンタルにマークアップを追加します。' },
+  addSubstText: {
+    text: '<subst>',
+    description: '(置換) - テキスト内での組み合わせが単一の介入と見なされる場合に、転写要素をグループ化します。',
+  },
+  substAddDel: { text: 'add | del', description: '<add> に選択を入れて <del> を追加します。' },
[…]
