+++ app/static/lang/lang.eo.js
@@ -187,9 +187,6 @@ export const lang = {
   addDiminuendoHairpinText: { text: 'Malgrandigo harpuno' },
   addBeamText: { text: 'Traveto' },
   addBeamSpanText: { text: 'Travetospaco' },
-  addSuppliedText: { text: 'Liverita' },
-  addSuppliedArticText: { text: 'Liverita (Artaĵo)' },
-  addSuppliedAccidText: { text: 'Liverita (Akĉento)' },
   addArpeggioText: { text: 'Arpegio' },
   addFermataText: { text: 'Fermato' },
   addGlissandoText: { text: 'Glisado' },
@@ -474,6 +471,105 @@ export const lang = {
     text1: 'Ne eblas enmeti noton ĉar la MEI ankorpunkto ne havas xml:id.',
     text2: 'Bonvolu aljuđi identigilojn elektante "Mankanĝi" -> "Re-rendi MEI (kun id-etoj)" kaj provi denove.',
   },
+  // Markup tools
+  respSelect: {
+    text: 'Elekti markad-respondecon',
+    description: 'Elekti identigilon de respondeco',
+  },
+  alternativeEncodingsGrp: {
+    text: 'Alternativaj kodoj',
+    description: 'Markigoj kiuj enhavas multoblajn versiojn.',
+  },
+  addChoiceText: {
+    text: '<choice>',
+    description: 'Grupigas nombron da alternativaj kodoj por la sama punkto en teksto.',
+  },
+  choiceSicCorr: { text: 'sic | corr', description: 'Metu elekton en <sic> kaj aldonu <corr>.' },
+  choiceCorrSic: { text: 'corr | sic', description: 'Metu elekton en <corr> kaj aldonu <sic>.' },
+  choiceOrigReg: { text: 'orig | reg', description: 'Metu elekton en <orig> kaj aldonu <reg>.' },
+  choiceRegOrig: { text: 'reg | orig', description: 'Metu elekton en <reg> kaj aldonu <orig>.' },
+  choiceContentTarget: {
+    text: '(elektu enhavon)',
+    description: 'Unue elektu enhavon por tiu elemento pasante la muson super <elekto>.',
+  },
+  addChoice: { text: 'Elektitaj elementoj', description: 'Aldonu markigon al elektitaj elementoj.' },
+  addChoiceArtic: { text: 'Artikulado', description: 'Aldonu markigon al artikuladoj en la elektado.' },
+  addChoiceAccid: { text: 'Okaza', description: 'Aldonu markigon al okazaĵoj en la elektado.' },
+  addSubstText: {
+    text: '<subst>',
+    description:
+      '(anstataŭigo) – Grupigas transskriptajn elementojn kiam la kombino devas esti rigardata kiel unuopa interveno en la teksto.',
+  },
+  substAddDel: { text: 'add | del', description: 'Metu elekton en <add> kaj aldonu <del>.' },
+  substDelAdd: { text: 'del | add', description: 'Metu elekton en <del> kaj aldonu <add>.' },
+  substContentTarget: {
+    text: '(elektu enhavon)',
+    description: 'Unue elektu enhavon por tiu elemento pasante la muson super <anstataŭigo>.',
+  },
+  addSubst: { text: 'Elektitaj elementoj', description: 'Aldonu markigon al elektitaj elementoj.' },
+  addSubstArtic: { text: 'Artikulado', description: 'Aldonu markigon al artikuladoj en la elektado.' },
+  addSubstAccid: { text: 'Okaza', description: 'Aldonu markigon al okazaĵoj en la elektado.' },
+  editInterventionsGrp: {
+    text: 'Redakciaj intervenoj',
+    description: 'Markigoj uzataj por enkodi redakciajn intervenojn.',
[…]
