### 2023-12-22 37c7529
Solve resp for choice and subst
