### 2023-12-22 f19b423
Add new xml:ids to copied descendants
... not just children
