+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add subst element to markup tools.
 * Moved responsibility select from mei-friend settings to markup tools.
 * Complete translations and tooltips for markup tools.
+* Fix generation of new xml:ids to copied elements.
 
 ### 1.1.3 <choice>
 
