### 2023-12-22 96c9a26
Add "Describe" function to markup list items
Pass selection as value to annot creators and start with sorting of list items
