### 2023-12-27 141f678
Refine situateListItems()
Runs asynchronously and and automatically sorts listItems when every item is situated. Also reduced rendundant calls.
