### 2023-12-27 1867f4d
Remove 'app' from alternativeEncodingElements
It is not supported yet because app needs special handling
