### 2023-12-27 b459f69
Refactored getX() and getY()
