### 2023-12-27 3d9d3a7
Remove dependent annotations if markup is deleted
