### 2023-12-27 46d7546
Improve sorting of list items
