+++ app/static/lang/lang.eo.js
@@ -426,8 +426,8 @@ export const lang = {
   annotationCloseButtonText: { text: 'Fermi la Panelon de Notoj' },
   hideAnnotationPanelButton: { description: 'Fermi la Panelon de Notoj' },
   closeAnnotationPanelButton: { description: 'Fermi la Panelon de Notoj' },
-  annotationToolsButton: { text: 'Iloj', description: 'Iloj de Notoj' },
-  annotationListButton: { text: 'Listo', description: 'Listigi Notojn' },
+  annotationToolsButton: { description: 'Iloj de Notoj' },
+  annotationListButton: { description: 'Listigi Notojn' },
   writeAnnotStandoffText: { text: 'Reteja Noto' },
   annotationToolsIdentifyTitle: { text: 'Identigi' },
   annotationToolsIdentifySpan: { text: 'Identigi Muzikan Objekton' },
@@ -444,6 +444,9 @@ export const lang = {
   loadWebAnnotationMessage2: { text: 'bonvolu reprovi' },
   noAnnotationsToDisplay: { text: 'Neniu notoj por montri' },
   flipPageToAnnotationText: { description: 'Turni paĝon al ĉi tiu noto' },
+  describeMarkup: { description: 'Priskribu ĉi tiun markadon' },
+  deleteMarkup: { description: 'Forigu ĉi tiun markadon' },
+  deleteMarkupConfirmation: { text: 'Ĉu vi certas, ke vi volas forigi ĉi tiun markadon?' },
   deleteAnnotation: { description: 'Forigi ĉi tiun noton' },
   deleteAnnotationConfirmation: { text: 'Ĉu vi certas ke vi volas forigi ĉi tiun noton?' },
   makeStandOffAnnotation: {
