+++ app/static/lang/lang.it.js
@@ -443,8 +443,8 @@ export const lang = {
   hideAnnotationPanelButton: { description: 'Nascondi la finestra di annotazione' },
   markupToolsButton: { description: 'Strumenti di marcatura' },
   closeAnnotationPanelButton: { description: 'Chiudi la finestra di annotazione' },
-  annotationToolsButton: { text: 'Strumenti', description: 'Strumenti di annotazione' },
-  annotationListButton: { text: 'Lista', description: 'Elenco delle annotazioni' },
+  annotationToolsButton: { description: 'Strumenti di annotazione' },
+  annotationListButton: { description: 'Elenco delle annotazioni' },
   writeAnnotStandoffText: { text: 'Annotazione Web' },
   annotationToolsIdentifyTitle: { text: 'Identifica' },
   annotationToolsIdentifySpan: { text: 'Identifica oggetto musicale' },
@@ -461,6 +461,9 @@ export const lang = {
   loadWebAnnotationMessage2: { text: 'per favore riprova' },
   noAnnotationsToDisplay: { text: 'Nessuna annotazione da visualizzare' },
   flipPageToAnnotationText: { description: 'Passa alla pagina di questa annotazione' },
+  describeMarkup: { description: 'Descrivi questo markup' },
+  deleteMarkup: { description: 'Elimina questo markup' },
+  deleteMarkupConfirmation: { text: 'Sei sicuro di voler eliminare questo markup?' },
   deleteAnnotation: { description: 'Elimina questa annotazione' },
   deleteAnnotationConfirmation: { text: 'Sei sicuro di voler eliminare questa annotazione?' },
   makeStandOffAnnotation: {
