+++ app/static/lang/lang.uk.js
@@ -421,8 +421,8 @@ export const lang = {
   annotationCloseButtonText: { text: 'Закрити панель анотацій' },
   hideAnnotationPanelButton: { description: 'Закрити панель анотацій' },
   closeAnnotationPanelButton: { description: 'Закрити панель анотацій' },
-  annotationToolsButton: { text: 'Інструменти', description: 'Інструменти анотацій' },
-  annotationListButton: { text: 'Список', description: 'Список анотацій' },
+  annotationToolsButton: { description: 'Інструменти анотацій' },
+  annotationListButton: { description: 'Список анотацій' },
   writeAnnotStandoffText: { text: 'Web анотація' },
   annotationToolsIdentifyTitle: { text: 'Визначити' },
   annotationToolsIdentifySpan: { text: "Визначити музичний об'єкт" },
@@ -439,6 +439,9 @@ export const lang = {
   loadWebAnnotationMessage2: { text: 'спробуйте ще раз' },
   noAnnotationsToDisplay: { text: 'Немає анотацій для відображення' },
   flipPageToAnnotationText: { description: 'Перейти до цієї анотації' },
+  describeMarkup: { description: 'Опишіть цей розмітка' },
+  deleteMarkup: { description: 'Видалити цей розмітка' },
+  deleteMarkupConfirmation: { text: 'Ви впевнені, що хочете видалити цей розмітка?' },
   deleteAnnotation: { description: 'Видалити цю анотацію' },
   deleteAnnotationConfirmation: { text: 'Ви впевнені, що хочете видалити цю анотацію?' },
   makeStandOffAnnotation: {
