### 2023-12-28 2797c67
Refresh list of all things on language change
