+++ app/static/lib/translator.js
@@ -6,6 +6,7 @@ import { isSafari } from './defaults.js';
 import * as l from '../lang/lang.en.js'; // default language
 import { translateLanguageSelection } from './language-selector.js';
 import { drawRightFooter, updateStatusBar } from './main.js';
+import { refreshAnnotationsList } from './enrichment_panel.js';
 import Viewer from './viewer.js';
 
 /**
@@ -92,6 +93,7 @@ export default class Translator {
     }
     updateStatusBar();
     drawRightFooter();
+    refreshAnnotationsList();
 
     this.handleLanguageExceptions();
 
