### 2024-01-03 d60f559
Improve choice select menu for files without choices
