+++ app/static/lang/lang.en.js
@@ -282,6 +282,7 @@ export const lang = {
   breaksSelectSmart: { text: 'Smart' },
   choiceSelect: { description: 'Choose displayed content for choice elements' },
   choiceDefault: { text: '(default choice)' },
+  noChoice: { text: '(no choice available)' },
   updateControlsLabel: { text: 'Update', description: 'Control update behavior of notation after changes in encoding' },
   liveUpdateCheckbox: { description: 'Automatically update notation after changes in encoding' },
   codeManualUpdateButton: { description: 'Update notation manually' },
