+++ app/static/lib/main.js
@@ -1,6 +1,6 @@
 // mei-friend version and date
-export const version = '1.1.3';
-export const versionDate = '20 December 2023'; // use full or 3-character english months, will be translated
+export const version = '1.1.4';
+export const versionDate = '03 January 2024'; // use full or 3-character english months, will be translated
 
 var vrvWorker;
 var spdWorker;
