### 2024-01-05 9771c03
Fix vanishing notation control bar (as in #a92284d)
