### 2024-01-16 b92cd38
Increment version to 1.0.7 and update changelog
