### 2024-01-17 fa48681
Fix choice selection in speed mode
