### 2024-01-18 948e06e
Include filtering of children for choice and subst in speed mode
