+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.1.5 patch
 
 * Include filtering for different versions inside `<choice>` or `<subst>` in speed mode.
+* Update midi on switching versions in the notation panel.
 
 ### 1.1.4 Refine Editorial Markup
 * Add subst element to markup tools.
