### 2024-01-19 ba955bb
Begin work on testing CMN file functions
