+++ e2e/testCMNFileFunctions.spec.js
@@ -1,0 +1,7 @@
+import { test, expect } from '@playwright/test';
+
+test.describe('Test CMN file functions', () => {
+  test('Test OpenURL function to load known CMN MEI', async ({ page }) => {
+    
+  });
+})
