### 2024-01-19 d3a0076
Add setup before each test Co-authored-by: @annplaksin Co-authored-by: @janjusolja Co-authored-by: @wergo Co-authored-by: @peterprovaznik Co-authored-by: @DILewis Co-authored-by: @simar0at Co-authored-by: @fraukevj
