+++ e2e/setup.ts
@@ -1,0 +1,17 @@
+const defaultUrl = 'http://localhost:5001/';
+
+export async function setupPage(page, url = defaultUrl) {
+  await page.goto(url);
+  await page.click('#splashConfirmButton');
+}
+
+export async function openUrl(page, url) {
+  await page.click('#fileMenuTitle');
+  // click on the element with ID "openUrl"
+  await page.click('#openUrl');
+  // focus on the element with ID "openUrlInput"
+  await page.focus('#openUrlInput');
+  // input known URL into the focused element
+  await page.keyboard.type(url);
+  await page.click('#openUrlButton');
+}
