### 2024-01-19 8003b31
Delete unused js files Co-authored-by: @annplaksin Co-authored-by: @janjusolja Co-authored-by: @wergo Co-authored-by: @peterprovaznik Co-authored-by: @DILewis Co-authored-by: @simar0at Co-authored-by: @fraukevj
