### 2024-01-19 1d1fda6
Set baseURL in playwrigth.config Co-authored-by: @peterprovaznik Co-authored-by: simar0at
