### 2024-01-20 1c5194f
Reload and redraw facsimile after keystroke
