### 2024-01-20 b9221dd
Revert reloading xmlDoc in notationUpdated()
