### 2024-01-21 a80932b
All search selectors for xml:id strings with quotes
