### 2024-01-21 f80113e
Fix bug in github menu that prevents reacting on cursor activity
