### 2024-01-21 0b7654b
Draw facsimile svg bounding boxes only with non-zero values
