### 2024-01-21 baddc37
Sort xml attributes ascending, but @xml:id first
