### 2024-01-22 75966f1
Include facsimile panel UI elements to local storage, incomplete commit
