### 2024-01-24 2c0e665
Connect facsimile zoom to local storage
