### 2024-01-24 b31d913
Bind facsimile panel UI inputs to settings inputs to assure local storage update
