### 2024-01-24 607ac65
Restore local storage values for Verovio settings
