### 2024-01-24 6f4d1af
Report min, max, and default values for Verovio number options
