### 2024-01-24 7df1edd
Increment version number to 1.0.9, update changelog and date
