+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Include filtering for different versions inside `<choice>` or `<subst>` in speed mode.
 * Update midi on switching versions in the notation panel.
+* Honrizontal layout for markup panel added.
 
 ### 1.1.4 Refine Editorial Markup
 * Add subst element to markup tools.
