### 2024-01-30 ea91643
Normalize and fix query selectors for xml:id search
