+++ app/static/lib/main.js
@@ -1,6 +1,6 @@
 // mei-friend version and date
 export const version = '1.0.9';
-export const versionDate = '24 January 2024'; // use full or 3-character english months, will be translated
+export const versionDate = '30 January 2024'; // use full or 3-character english months, will be translated
 
 var vrvWorker;
 var spdWorker;
