### 2024-01-31 6c07529
Add playwright installation instructions
