### 2024-01-31 bd65b03
Add default MEI file load test
