+++ e2e/testCMNFileFunctions.spec.ts
@@ -7,17 +7,6 @@ test.beforeEach(async ({ page }) => {
 });
 
 test.describe('Test CMN file functions', () => {
-  /*
-  // TODO - decide whether we want to keep this, given that the default MEI might change
-
-  test('Test default MEI (Beethoven WoO70) loads and renders as notation', async ({ page }) => {
-    const svgElement = page.locator('#rend-0000002081006402>tspan>tspan');
-    await expect(svgElement).toBeVisible();
-    const innerHTML = await svgElement.evaluate(el => el.innerHTML);
-    expect(innerHTML).toBe("SECHS VARIATIONEN");
-  });
-  */
-
   test('Test that default MEI loads and renders as notation', async ({ page }) => {
     await expect(page.locator('g.note').first()).toBeVisible(); // can we see the first note?
     await expect(page.locator('g.note').first()).toHaveClass("note highlighted"); // first note highlighted?
