### 2024-01-31 41b46ef
Reactivate cursor activity after edit/navigate operations
