### 2024-01-31 5c84991
Add initial settings panel test
