+++ app/static/css/icons.js
@@ -58,6 +58,9 @@ export const chevronFirst =
 export const chevronLast =
   '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16"><path transform="rotate(-90 8 8) translate(0,-2)" fill-rule="evenodd" d="M12.78 6.22a.75.75 0 010 1.06l-4.25 4.25a.75.75 0 01-1.06 0L3.22 7.28a.75.75 0 011.06-1.06L8 9.94l3.72-3.72a.75.75 0 011.06 0zM3.75 13a.75.75 0 000 1.5h8.5a.75.75 0 000-1.5h-8.5z"/></svg>';
 
+export const codebergLogo = 
+  '<svg id="profile_codeberg_logo" version="1.1" viewBox="-6 0 75.976 69.999897" height="19" width="19" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" xmlns:svg="http://www.w3.org/2000/svg" xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:cc="http://creativecommons.org/ns#" xmlns:dc="http://purl.org/dc/elements/1.1/"> <title id="title11271">Codeberg</title> <defs id="defs6983"> <linearGradient xlink:href="#linearGradient6924" id="linearGradient6918" x1="42519.285" y1="-7078.7891" x2="42575.336" y2="-6966.9307" gradientUnits="userSpaceOnUse" /> <linearGradient id="linearGradient6924"> <stop style="stop-color:#ffffff;stop-opacity:0" offset="0" id="stop6920" /> <stop id="stop6926" offset="0.49517274" style="stop-color:#71c2ff;stop-opacity:1" /> <stop style="stop-color:#39aaff;stop-opacity:1" offset="1" id="stop6922" /> </linearGradient> </defs> <metadata id="metadata6986"> <rdf:RDF> <cc:Work rdf:about=""> <dc:format>image/svg+xml</dc:format> <dc:type rdf:resource="http://purl.org/dc/dcmitype/StillImage" /> <dc:title>Codeberg logo</dc:title> <dc:creator> <cc:Agent> <dc:title>Robert Martinez</dc:title> </cc:Agent> </dc:creator> <cc:license rdf:resource="http://creativecommons.org/licenses/by/4.0/" /> <dc:date>2020-05-41</dc:date> <dc:rights> <cc:Agent> <dc:title>Codeberg and the Codeberg Logo are trademarks of Codeberg e.V.</dc:title> </cc:Agent> </dc:rights> </cc:Work> <cc:License rdf:about="http://creativecommons.org/licenses/by/4.0/"> <cc:permits rdf:resource="http://creativecommons.org/ns#Reproduction" /> <cc:permits rdf:resource="http://creativecommons.org/ns#Distribution" /> <cc:requires rdf:resource="http://creativecommons.org/ns#Notice" /> <cc:requires rdf:resource="http://creativecommons.org/ns#Attribution" /> <cc:permits rdf:resource="http://creativecommons.org/ns#Deriva
[…]
