+++ app/static/lang/lang.eo.js
@@ -89,7 +89,7 @@ export const lang = {
   },
 
   // Ĉefmenuo
-  githubLoginLink: { text: 'Ensaluti' },
+  githubLoginLinkText: { text: 'git' },
 
   month: {
     jan: 'Januaro',
