### 2024-02-02 b8f22fe
Fix version selection in speed mode
