### 2024-02-02 88f0263
Fix deleteMarkup while a reading is selected
