### 2024-02-09 faa60ba
Fix version selection and parsing in speed mode
