### 2024-02-12 0c0982f
Refactor git menu for use with profile pane
