+++ app/static/lib/github-menu.js
@@ -771,9 +771,6 @@ export function logoutFromGithub() {
 } // logoutFromGithub()
 
 export function refreshGithubMenu() {
-  // display Github name
-  document.getElementById('profile_githubNameOrLoginMsg').innerText =
-    github.author.name === 'None' ? github.userLogin : github.author.name;
   // populate Github menu
   let githubMenu = document.getElementById('GithubMenu');
   githubMenu.classList.remove('loggedOut');
