+++ app/static/lib/main.js
@@ -2,6 +2,9 @@
 export const version = '1.0.10';
 export const versionDate = '31 January 2024'; // use full or 3-character english months, will be translated
 
+// HACK 20240212: temporarily make profile.github.obj a shared / exported variable until githubmenu can be refactored
+export let github;
+
 var vrvWorker;
 var spdWorker;
 var tkAvailableOptions;
@@ -19,7 +22,6 @@ export var cm;
 export var v; // viewer instance
 export var validator; // validator object
 export var rngLoader; // object for loading a relaxNG schema for hinting
-export let github; // github API wrapper object
 export let storage = new Storage();
 export var tkVersion = ''; // string of the currently loaded toolkit version
 export var tkUrl = ''; // string of the currently loaded toolkit origin
@@ -186,7 +188,7 @@ export function setFileChangedState(fileChangedState) {
     fileStatusElement.classList.remove('changed');
     fileChangedIndicatorElement.innerText = '';
   }
-  if (isLoggedIn && github && github.filepath && commitUI) {
+  if (isLoggedIn && profile.github.available && profile.github.obj.filepath && commitUI) {
     setCommitUIEnabledStatus();
   }
   if (storage.supported) {
@@ -289,16 +291,16 @@ export function updateLocalStorage(meiXml) {
 
 export function updateGithubInLocalStorage() {
   if (storage.supported && !storage.override && isLoggedIn) {
-    const author = github.author;
+    const author = profile.github.obj.author;
     const name = author.name;
     const email = author.email;
     storage.github = {
-      githubRepo: github.githubRepo,
-      githubToken: github.githubToken,
-      branch: github.branch,
-      commit: github.commit,
-      filepath: github.filepath,
-      userLogin: github.userLogin,
+      githubRepo: profile.github.obj.githubRepo,
+      githubToken: profile.github.obj.githubToken,
+      branch: profile.github.obj.branch,
+      commit: profile.github.obj.commit,
+      filepath: profile.github.obj.filepath,
+      userLogin: profile.github.obj.userLogin,
       userName: name,
       userEmail: email,
     };
@@ -388,7 +390,7 @@ document.addEventListener('DOMContentLoaded', function () {
     github: gitEnabled,
     gitlab: false, // TODO gitlab integration
     codeberg: false, // TODO codeberg integration
-  })
+  });
 
   if (!gitEnabled) {
     document.getElementById('GithubButton').disabled = true;
@@ -562,7 +564,7 @@ function completeInitialLoad() {
     if (storage.github) {
       // use github object from local storage if available
       isLoggedIn = true;
-      github = new Github(
+      profile.github.obj = new Github(
         storage.github.githubRepo,
         storage.github.githubToken,
         storage.github.branch,
@@ -575,14 +577,14 @@ function completeInitialLoad() {
       //document.querySelector("#fileLocation").innerText = meiFileLocationPrintable;
     } else if (isLoggedIn) {
       // initialise and store new github object
-      github = new Github('', githubToken, '', '', '', userLogin, userName, userEmail);
+      profile.github.obj = new Github('', githubToken, '', '', '', userLogin, userName, userEmail);
       storage.github = {
-        githubRepo: github.githubRepo,
-        githubToken: github.githubToken,
-        branch: github.branch,
-        commit: github.commit,
-        filepath: github.filepath,
-        userLogin: github.userLogin,
+        githubRepo: profile.github.obj.githubRepo,
+        githubToken: profile.github.obj.githubToken,
+        branch: profile.github.obj.branch,
+        commit: profile.github.obj.commit,
+        filepath: profile.github.obj.filepath,
[…]
