### 2024-02-12 cd8405a
politely indicate git menu when git profile clicked
