### 2024-02-13 0c6bfb6
add profile options to settings (WIP)
