### 2024-02-13 348e86a
add login status spans and buttons
