### 2024-02-14 6512ec4
Update settings status to reflect profile login state (WIP)
