### 2024-02-14 #6002 @wergo
We are on our way, but need some time to implement a comprehensive test framework. We are implementing this using [Playwright](https://playwright.dev/). We are aiming to have the re-submission ready by early March.
