### 2024-02-14 820ef8f
Refactor scrollSvgTo()
