### 2024-02-14 d3ab4a4
Facsimile image scrolls into view if zone is there for element
