### 2024-02-15 6c26f58
Add semantic colouring of markup buttons
Split buttons need CSS fix
