### 2024-02-15 c9780ef
Rename testSettingsPanel to settings.spec.ts, start work on Verovio settings tab.
