### 2024-02-15 93870f3
Fix adding facsimile zones in non-fullpage mode
