### 2024-02-16 a5fb63a
Reloading facsimile upon xmlDoc reload
