### 2024-02-16 35cbe96
Deal with double notes in tab notation SVG
