### 2024-02-16 4a01ba5
Trying to trace double images in SVG (WIP)
