+++ app/static/lib/facsimile.js
@@ -14,7 +14,6 @@ import { defaultFacsimileRectangleColor, defaultFacsimileRectangleLineWidth } fr
 
 var facs = {}; // facsimile structure in MEI file
 var sourceImages = {}; // object of source images
-var oldFacsimile; // previous facsimile element
 const rectangleLineWidth = defaultFacsimileRectangleLineWidth; // width of bounding box rectangles in px
 const rectangleColor = defaultFacsimileRectangleColor; // color of zone rectangles
 var listenerHandles = {};
@@ -163,7 +162,7 @@ export async function drawFacsimile() {
   );
 
   if (svgFacs && fullPage) {
-    let firstZone = svgFacs.item(0);
+    let firstZone = svgFacs.at(0);
     if (firstZone && firstZone.hasAttribute('data-facs')) zoneId = rmHash(firstZone.getAttribute('data-facs'));
   } else {
     svgFacs.forEach((f) => {
@@ -195,7 +194,6 @@ export async function drawFacsimile() {
     }
   }
   if (zoneId && facs[zoneId]) {
-    svg.innerHTML = '';
     // find the correct path of the image file
     let img;
     let imgName = facs[zoneId].target;
@@ -237,7 +235,9 @@ export async function drawFacsimile() {
       img = await loadImage(imgName);
       sourceImages[imgName] = img;
     }
+
     //
+    Array.from(svg.children).forEach((c) => svg.removeChild(c));
     if (img) {
       console.log('Appending child image:', img);
       svg.appendChild(img);
@@ -314,21 +314,21 @@ export async function drawFacsimile() {
  */
 function drawBoundingBox(zoneId) {
   if (facs[zoneId]) {
-    let pointerId = facs[zoneId]['pointerId'];
-    let pointerN = facs[zoneId]['pointerN'];
+    let svg = document.getElementById('sourceImageSvg');
     let rect = document.createElementNS(svgNameSpace, 'rect');
+    svg.appendChild(rect);
     rect.setAttribute('rx', rectangleLineWidth / 2);
     rect.setAttribute('ry', rectangleLineWidth / 2);
     rect.addEventListener('click', (e) => v.handleClickOnNotation(e, cm));
     let editFacsimileZones = document.getElementById('editFacsimileZones').checked;
-    let svg = document.getElementById('sourceImageSvg');
-    svg.appendChild(rect);
     let x = parseFloat(facs[zoneId].ulx);
     let y = parseFloat(facs[zoneId].uly);
     let width = parseFloat(facs[zoneId].lrx) - x;
     let height = parseFloat(facs[zoneId].lry) - y;
     updateRect(rect, x, y, width, height, rectangleColor, rectangleLineWidth, 'none');
+    let pointerId = facs[zoneId]['pointerId']; // id of pointing element (e.g., measure)
     if (pointerId) rect.id = editFacsimileZones ? zoneId : pointerId;
+    let pointerN = facs[zoneId]['pointerN']; // number-like
     if (pointerN) {
       // draw number-like info from element (e.g., measure)
       let txt = document.createElementNS(svgNameSpace, 'text');
@@ -434,8 +434,8 @@ export function highlightZone(rect) {
   if (document.getElementById('editFacsimileZones').checked) listenerHandles = addZoneResizer(v, rect);
 
   // highlight rectangle
-  // rect.classList.add('highlighted');
-  // rect.querySelectorAll('g').forEach((g) => g.classList.add('highlighted'));
+  rect.classList.add('highlighted');
+  rect.querySelectorAll('g').forEach((g) => g.classList.add('highlighted'));
 } // highlightZone()
 
 /**
@@ -446,12 +446,7 @@ export function highlightZone(rect) {
[…]
