### 2024-02-19 debf654
Secure variables in facsimile and optimize code
