### 2024-02-19 db38fd3
Add verbosity to facsimile envelope coordinates
