### 2024-02-19 c0a6b72
Account for Firefox bug in retrieving screenCTM in facsimile zone drawing
