### 2024-02-20 14ef99d
add click listeners and descriptions to profile settings login / logout buttons
