### 2024-02-20 ca45e02
Optimize facsimile drawing code
