### 2024-02-20 d172132
Keep zones highlighted from notation during zone editing
