### 2024-02-20 8afab71
Hide login status label when not logged in
