### 2024-02-22 e4b2418
Add 4.1.0 to supported Verovio versions and update date
