### 2024-02-23 61a83ea
add initial mei-friend settings tab tests
