### 2024-02-23 2a1493e
Some progress in input
