### 2024-03-01 e3319e9
Fix error if result of selectMarkup is empty
