### 2024-03-01 0085c75
Move selection for markup to a separate select in markup panel
