+++ app/static/lang/lang.ja.js
@@ -481,6 +481,16 @@ export const lang = {
     text: 'マークアップの責任を選択',
     description: '責任のIDを選択',
   },
+  selectionSelect: {
+    text: 'マークアップのデフォルト選択',
+    description: '新しく作成されたマークアップが選択された要素、アーティキュレーション、またはアクシデンタルを囲むかどうかを選択してください',
+    labels: ['選択された要素','アーティキュレーション', 'アクシデンタル'],
+    valuesDescriptions: [
+      '選択された要素にマークアップを追加します。', 
+      '選択範囲内のアーティキュレーションにマークアップを追加します。',
+      '選択範囲内のアクシデンタルにマークアップを追加します。'
+    ],
+  },
   alternativeEncodingsGrp: {
     text: '代替エンコーディング',
     description: '複数のバージョンを含むマークアップ要素。',
@@ -504,12 +514,6 @@ export const lang = {
     text: '(コンテンツを選択)',
     description: '<choice> の上にホバーしてこの要素のためにコンテンツを選択します。',
   },
-  addChoice: { text: '選択された要素', description: '選択された要素にマークアップを追加します。' },
-  addChoiceArtic: {
-    text: 'アーティキュレーション',
-    description: '選択範囲内のアーティキュレーションにマークアップを追加します。',
-  },
-  addChoiceAccid: { text: 'アクシデンタル', description: '選択範囲内のアクシデンタルにマークアップを追加します。' },
   addSubstText: {
     text: '<subst>',
     description: '(置換) - テキスト内での組み合わせが単一の介入と見なされる場合に、転写要素をグループ化します。',
@@ -524,12 +528,6 @@ export const lang = {
     text: '(コンテンツを選択)',
     description: '<subst> の上にホバーしてこの要素のためにコンテンツを選択します。',
   },
-  addSubst: { text: '選択された要素', description: '選択された要素にマークアップを追加します。' },
-  addSubstArtic: {
-    text: 'アーティキュレーション',
-    description: '選択範囲内のアーティキュレーションにマークアップを追加します。',
-  },
-  addSubstAccid: { text: 'アクシデンタル', description: '選択範囲内のアクシデンタルにマークアップを追加します。' },
   editInterventionsGrp: {
     text: '編集介入',
     description: '編集介入をエンコードするために使用されるマークアップ要素。',
@@ -538,81 +536,33 @@ export const lang = {
     text: '<supplied>',
     description: '転写者または編集者によって供給された素材を含みます。',
   },
-  addSupplied: { text: '選択された要素', description: '選択された要素にマークアップを追加します。' },
-  addSuppliedArtic: {
-    text: 'アーティキュレーション',
-    description: '選択範囲内のアーティキュレーションにマークアップを追加します。',
-  },
-  addSuppliedAccid: { text: 'アクシデンタル', description: '選択範囲内のアクシデンタルにマークアップを追加します。' },
   addUnclearText: {
     text: '<unclear>',
     description: '原文で確実に転写できない素材を含みます（読めないか聞き取れない）。',
   },
-  addUnclear: { text: '選択された要素', description: '選択された要素にマークアップを追加します。' },
-  addUnclearArtic: {
-    text: 'アーティキュレーション',
-    description: '選択範囲内のアーティキュレーションにマークアップを追加します。',
-  },
-  addUnclearAccid: { text: 'アクシデンタル', description: '選択範囲内のアクシデンタルにマークアップを追加します。' },
   addSicText: { text: '<sic>', description: '明らかに不正確または誤った素材を含みます。' },
-  addSic: { text: '選択された要素', description: '選択された要素にマークアップを追加します。' },
-  addSicArtic: {
[…]
