+++ app/static/lang/lang.ca.js
@@ -511,6 +511,16 @@ export const lang = {
     text: 'Seleccionar la responsabilitat de la marca',
     description: "Seleccionar l'ID de la responsabilitat",
   },
+  selectionSelect: {
+    text: 'Selecció per defecte per a la marca',
+    description: 'Trieu si la marca de nova creació hauria d\'envoltar els elements seleccionats, les articulacions o els accidentals',
+    labels: ['Elements seleccionats','Articulació', 'Accidental'],
+    valuesDescriptions: [
+      'Afegeix marcatge als elements seleccionats.', 
+      'Afegeix marcatge a les articulacions dins la selecció.', 
+      'Afegeix marcatge als accidentals dins la selecció.'
+    ],
+  },
   alternativeEncodingsGrp: {
     text: 'Codificacions alternatives',
     description: 'Elements de marcatge que contenen múltiples versions.',
@@ -535,9 +545,6 @@ export const lang = {
     text: '(selecciona contingut)',
     description: 'Primer, selecciona contingut per a aquest element passant el ratolí sobre <choice>.',
   },
-  addChoice: { text: 'Elements seleccionats', description: 'Afegeix marcatge als elements seleccionats.' },
-  addChoiceArtic: { text: 'Articulació', description: 'Afegeix marcatge a les articulacions dins la selecció.' },
-  addChoiceAccid: { text: 'Accidental', description: 'Afegeix marcatge als accidentals dins la selecció.' },
   addSubstText: {
     text: '<subst>',
     description:
@@ -553,9 +560,6 @@ export const lang = {
     text: '(selecciona contingut)',
     description: 'Primer, selecciona contingut per a aquest element passant el ratolí sobre <subst>.',
   },
-  addSubst: { text: 'Elements seleccionats', description: 'Afegeix marcatge als elements seleccionats.' },
-  addSubstArtic: { text: 'Articulació', description: 'Afegeix marcatge a les articulacions dins la selecció.' },
-  addSubstAccid: { text: 'Accidental', description: 'Afegeix marcatge als accidentals dins la selecció.' },
   editInterventionsGrp: {
     text: 'Intervencions editorials',
     description: 'Elements de marcatge utilitzats per codificar intervencions editorials.',
@@ -564,57 +568,33 @@ export const lang = {
     text: '<supplied>',
     description: 'Conté material subministrat pel transcriptor o editor per qualsevol motiu.',
   },
-  addSupplied: { text: 'Elements seleccionats', description: 'Afegeix marcatge als elements seleccionats.' },
-  addSuppliedArtic: { text: 'Articulació', description: 'Afegeix marcatge a les articulacions dins la selecció.' },
-  addSuppliedAccid: { text: 'Accidental', description: 'Afegeix marcatge als accidentals dins la selecció.' },
   addUnclearText: {
     text: '<unclear>',
     description: 'Conté material que no es pot transcriure amb certesa perquè és il·legible o inaudible a la font.',
   },
-  addUnclear: { text: 'Elements seleccionats', description: 'Afegeix marcatge als elements seleccionats.' },
-  addUnclearArtic: { text: 'Articulació', description: 'Afegeix marcatge a les articulacions dins la selecció.' },
-  addUnclearAccid: { text: 'Accidental', description: 'Afegeix marcatge als accidentals dins la selecció.' },
   addSicText: { text: '<sic>', description: 'Conté material aparentment incorrecte o inexacte.' },
-  addSic: { text: 'Elements seleccionats', description: 'Afegeix marcatge als elements seleccionats.' },
-  addSicArtic: { text: 'Articulació', description: 'Afegeix marcatge a les articulacions dins la selecció.' },
[…]
