+++ app/static/lib/markup.js
@@ -336,7 +336,7 @@ function firstChildElement(parent) {
 export function addMarkup(event) {
   let eventTarget = event.currentTarget;
   let mElName = eventTarget.dataset.elName;
-  let attrName = eventTarget.dataset.selection;
+  let attrName = document.getElementById('selectionSelect').value;;
   let multiLayerContent = eventTarget.dataset.content?.split(',');
   if (!att.modelTranscriptionLike.includes(mElName) && !att.alternativeEncodingElements.includes(mElName)) return;
 

[…]
