### 2024-03-06 d90ac4d
Properly toggle markup button background colors by settings
