### 2024-03-07 8f9714b
Update INSTALL and README (#58)
