### 2024-03-07 a736d55
Fix typo in README
