### 2024-03-07 5769f96
Some fixes in README
