### 2024-03-07 mei-friend/mei-friend#58 @wergo
We have extended the descriptions on INSTALL.md and made the registration as GitHub app optional.

There is a reference to INSTALL.md on README.md and on the documentation pages (Advanced – Developing mei-friend).
