+++ CONTRIBUTING.md
@@ -1,0 +1,41 @@
+# Contributing to mei-friend
+
+We are grateful to community contributions to the mei-friend project. You may either raise and issue for ...
+
+## Steps for creating good issues or pull requests
+
+1. Provide a short but descriptive title and description of the issue or pull request.
+2. For pull requests, add a description of the changes made.
+3. Use markdown to format the text.
+4. Provide links to any related issues or pull requests.
+5. Add labels for the issue or pull request.
+6. Include a screenshot or animated GIF if the pull request is for a visual change.
+
+## Links to external documentation
+
+* [README](README.md)
+* [Installation](INSTALL.md)
+* [mei-friend documentation pages](https://mei-friend.github.io/)
+
+## Code contributions
+
+We welcome code contributions for new features to mei-friend. If you would like to have them included into the main branch of mei-friend, please get in contact with the mei-friend developer team before you raise a Pull Request on GitHub.
+
+* Please use Prettier to format the javascript code.
+ 
+
+## Community and behavioral expectations.
+
+1. Be respectful to others and their opinions.
+2. Be open to feedback and criticism.
+3. Be helpful and supportive to others in the community.
+
+
+
+
+
+
+
+
+
+    
