+++ CONTRIBUTING.md
@@ -1,41 +1,49 @@
 # Contributing to mei-friend
 
-We are grateful to community contributions to the mei-friend project. You may either raise and issue for ...
+mei-friend is developed in an open-source environment and we welcome contributions from the community. This document provides guidelines for contributing to mei-friend.
 
-## Steps for creating good issues or pull requests
+## How can I contribute?
 
-1. Provide a short but descriptive title and description of the issue or pull request.
-2. For pull requests, add a description of the changes made.
-3. Use markdown to format the text.
-4. Provide links to any related issues or pull requests.
-5. Add labels for the issue or pull request.
-6. Include a screenshot or animated GIF if the pull request is for a visual change.
+We are grateful for contributions to the mei-friend project in the following areas: bug reports, feature requests, contributing encodings to the public repertoire, contributing translation fixes and new translations, and contributing to the mei-friend codebase.
 
-## Links to external documentation
+## Reporting Bugs
 
-* [README](README.md)
-* [Installation](INSTALL.md)
-* [mei-friend documentation pages](https://mei-friend.github.io/)
+If you find a bug in mei-friend, please raise an issue on the [mei-friend GitHub repository's issue tracker](https://github.com/mei-friend/mei-friend/issues/new/choose). Please follow the template provided, entering as much information as possible, including the version of mei-friend you are using, the operating system you are using, and the steps to reproduce the bug.
 
-## Code contributions
+## Requesting new features
 
-We welcome code contributions for new features to mei-friend. If you would like to have them included into the main branch of mei-friend, please get in contact with the mei-friend developer team before you raise a Pull Request on GitHub.
+If you have an idea for a new feature in mei-friend, please raise an issue on the [mei-friend GitHub repository's issue tracker](https://github.com/mei-friend/mei-friend/issues/new/choose). Again, please follow the provided template, clearly describing the problem to be solved, your proposed solution, and any additional context.
 
-* Please use Prettier to format the javascript code.
- 
+## Contributing encodings to the public repertoire
 
-## Community and behavioral expectations.
+If you have would like to contribute a publicly licenced encoding to mei-friend's public repertoire (accessible in mei-friend under `File` -> `Public repertoire`), please raise an issue on the [mei-friend GitHub repository's issue tracker](https://github.com/mei-friend/mei-friend/issues/new/choose). The relevant template will provide you with the required contribution form. Please note that we can only accept contributions that meet the following requirements:
 
-1. Be respectful to others and their opinions.
-2. Be open to feedback and criticism.
-3. Be helpful and supportive to others in the community.
+- Supported encoding format: MEI, Kern, MusicXML (raw and compressed), ABC, PAE.
+- Available via public Github repositories.
+- Available under a public (e.g., a Creative Commons) license.
+- The licensing of the encoding must be explicitly stated in the GitHub repository and/or in the encoding itself, e.g. using the MEI [useRestrict](https://music-encoding.org/guidelines/v5/elements/useRestrict) element.
 
+## Contributing translation fixes and new translations
 
+mei-friend's interface supports a growing number of natural languages, which can be selected from the 'globe' menu and the 'mei-friend settings' dialog. These translations are maintained as JavaScript objects within files in a [dedicated folder in mei-friend's codebase](https://github.com/mei-friend/mei-friend/tree/develop/app/static/lang). To contribute a new translation or fix an existing translation, please first [fork the mei-friend repository](https://github.com/mei-friend/mei-friend/fork), and then make your changes to the relevant language file(s).
 
[…]
