### 2024-03-07 e17b9cc
add info on providing feedback from within mei-friend
