### 2024-03-07 mei-friend/mei-friend#59 @musicog
Dear @stefan-balke, thanks for this issue and apologies for the rather slow response.

mei-friend is implemented as an open-source single-page Web application written in “vanilla” JavaScript. While it is indeed served by a minimal Python Flask server, this server's tasks are minimal -- it serves the main HTML template and the JavaScript static files, and also handles basic OAuth authentication for the GitHub integration.

Major re-used JavaScript components include [CodeMirror](https://codemirror.net/5/), a highly configurable text editor providing the basis for mei-friend’s encoding panel, and [Verovio](https://verovio.org), a powerful MEI engraving engine used for a number of purposes within our application, most significantly to provide the SVG renderings displayed in the notation panel, and to handle input and output functionalities including conversions from other formats.

A comprehensive description of all components re-used within mei-friend is available in the README.md file within the editor’s code repository.

Do you think this is sufficient to fulfil the information need required? Or does it need to be spelled out more clearly elsewhere?
