### 2024-03-07 mei-friend/mei-friend#68 @wergo
We have added a CONTRIBUTION.md to the root folder of the repo containing the five ways how the community may contribute to the mei-friend project. This file is also linked from the [documentation pages](https://mei-friend.github.io/docs/advanced/contributing/).
