### 2024-03-07 88edfe5
Increment to 1.0.11, update changelog and date
