### 2024-03-07 mei-friend/mei-friend#60 @wergo
We have started a page for developers at https://mei-friend.github.io/docs/advanced/structure/ where the main folders and files are described. As requested by #68, we have added a more detailed documentation for contributions to the mei-friend project at https://mei-friend.github.io/docs/advanced/contributing/.

We will continue explaining the files in app/static/lib in more detail shortly.
