### 2024-03-07 mei-friend/mei-friend#59 @wergo
In addition, we have explained some of the software architecture in the documentation at https://mei-friend.github.io/docs/advanced/structure/. More details about the files in `app/static/lib/` to follow.
