### 2024-03-11 mei-friend/mei-friend#69 @musicog
Thank you for this suggestion. We have improved the build instructions within `INSTALL.md` and streamlined the install process, e.g. to make GitHub configuration optional. The dependencies have been moved into a 'deps' folder and no longer need building; the only remaining npm requirement is for the (newly added) playwright end-to-end tests, as also described in `INSTALL.md`.

Our expectation is that most users of our application will make use of the hosted version at https://mei-friend.mdw.ac.at. Do you think our newly streamlined installation procedure is sufficiently straight-forward now to cover those users who wish to self-host? Or is a move to a installable python package vastly preferable?
