+++ joss2023/paper.md
@@ -27,6 +27,7 @@ Digital music encodings are machine-readable representations of music score docu
 
 mei-friend is a friendly, interactive, browser-based editor for music encodings. It opens music encodings in a variety of formats and converts these to MEI, the XML-based format developed by the Music Encoding Initiative, a community of music scholars, librarians, and technologists. Alongside a rich set of interactive editing operations, mei-friend offers specialized functionalities targeted at music scholars, including panels to display facsimile score images and to author score annotations. Collaborative encoding practices are supported through integration with GitHub and Solid, and by remote configuration capabilities via hyperlink (URL parameters). Important MEI community resources are also integrated, supporting direct reference to documentation [@MEI_Guidelines] for currently-selected elements, and auto-completion and validation using the MEI-XML schema definitions. Also provided are curated links to openly-licensed reference music encodings.
 
+The editor targets an audience of music scholars, students, librarians, performers, and enthusiasts. It is not conceived as a replacement for commercial music notation software, but rather as a complement opening up the affordances of machine-readable semantic music encodings to a broader audience. Our tool is the only currently available editor with comprehensive, native MEI-editing support in Common Western Music Notation (CMN). Editing directly in MEI avoids conversion errors (due to bugs or inherent differences in supported features) from other formats, while supporting interactive use of those capabilities of the MEI schema that other formats simply do not support, such as editorial markup, annotation, and facsimile integration. 
 
 # Statement of Need
 
