### 2024-03-11 39340d9
Update settings for building paper
