### 2024-03-11 daccc0f
Update screenshot figure
