+++ joss2023/paper.md
@@ -34,7 +34,7 @@ Music encodings are machine-readable representations of music scores with import
 
 MEI, the family of XML schemas promoted and continually developed by the community of the Music Encoding Initiative, is noted for its richly comprehensive representation of music semantics, including facilities for digital scholarly edition, analytical markup, and capturing catalogue-level metadata and source descriptions alongside the music content; and, for its support for a variety of notations including tablatures, neumes, and mensural notation alongside common western music notation [@Crawford2016music]. These capabilities set MEI apart for scholarly use from less richly elaborated but more widespread competitors such as MusicXML, which, unlike MEI, can be exported by most commercial notation software [@Pugin2018interaction]. Conversely, the retention of XML as a technological basis, with advantages particularly in terms of interoperability and addressability, speak for MEI in favour of other research-focused formats such as Humdrum Kern in the context of FAIR data management [@Weigl_etal_EMR2021]. 
 
-![mei-friend interface: MEI encoding of Beethoven's WoO 57 (right panel), digital score rendering (top-left panel), and associated facsimile image of the source edition (bottom-left panel), with the currently-selected measure highlighted in each modality.\label{fig:example}](mei-friend-WoO57-facsimile-2.png){ width="100%" }
+![mei-friend interface: MEI encoding of Beethoven's WoO 57 (right panel), digital score rendering (top-left panel), and associated facsimile image of the source edition (bottom-left panel), with the currently-selected measure highlighted in each modality.\label{fig:example}](mei-friend-WoO57-facsimile-3.png){ width="100%" }
 
 These properties have led to the use of MEI as a de-facto encoding standard in music scholarship and cultural heritage preservation in recent years, as evidenced by its adoption as a recommended preservation format for digital scores by the @loc2019mei. However, the relatively scarce availability of high-quality, publicly-licensed music encodings remains a limiting factor, and the creation of such encodings, particularly at scale, is a high-effort activity with a relatively steep learning curve. 
 
