### 2024-03-11 b107b0b
Update screenshot and paper
