+++ joss2023/paper.md
@@ -52,7 +52,7 @@ The official application instance, available at <https://mei-friend.mdw.ac.at/>,
 The mei-friend development process operates over three distinct environments: *production*, served from the `main` branch of the mei-friend code repository; *staging*, served from the `staging` branch and accessible for public testing purposes at <https://staging.mei-friend.mdw.ac.at>; and *develop*, served from the `develop` branch. The application adjusts to its environment context, running tests when in develop and altering the logo to alert users when in staging.
 
 The editor has seen significant and growing adoption by the MEI community, currently receiving more than five-hundred distinct visitor sessions each month on its hosted instance.
-The predecessor Atom plugin was reviewed in the context of editorial work for the *Digital Interactive Mozart Edition* [@Sapov_2022] and used in the edition of a large collection of Beethoven solo-piano pieces [@weigl2019interweaving].
+The predecessor Atom plugin was reviewed as part of editorial work for the *Digital Interactive Mozart Edition* [@Sapov_2022] and used to edit a large collection of Beethoven solo-piano pieces [@weigl2019interweaving].
 The Web application is central to ongoing Digital Musicology research projects analysing and comparing orchestral performances aligned with digital score encodings [@Weigl_etal_DLfM2022] and establishing a digital edition of German lute tablatures of the 15th and 16th centuries [@elaute2023mec]. 
 A further community-based project is underway to extend mei-friend's support for scholarly edition and markup [@nfdi4culture_Plaksin2022]. 
 mei-friend has been used to teach music encoding at universities in Boston, Vienna, and Würzburg; academic events, including the Digital Humanities @ Oxford Summer School and the *Semana HD* organized by South American Digital Humanities associations; and conferences, including the *Music Encoding Conference* (MEC), the *Medieval and Renaissance Music Conference* (MedRen), and the *Extended Semantic Web Conference* (ESWC). 
