### 2024-03-11 39fdecd
Update bibtex file
