+++ joss2023/paper.bib
@@ -62,6 +62,24 @@
     DOI={10.1145/3543882.3543892}
 }
 
+@inproceedings{Weigl_etal_DLfM2023,
+author = {Weigl, David M. and VanderHart, Chanda and Rammler, Delilah and Pescoller, Matth\"{a}us and Goebl, Werner},
+title = {Listen Here! A Web-native digital musicology environment for machine-assisted close listening},
+year = {2023},
+isbn = {9798400708336},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3625135.3625144},
+doi = {10.1145/3625135.3625144},
+abstract = {Close listening is a mainstay of the musicological study of performance recordings, but paying focused, critical attention carries high cognitive overhead, making the application of this approach difficult when dealing with large corpora. In Signature&nbsp;Sound&nbsp;Vienna, a project investigating the Vienna Philharmonic Orchestra’s New Year’s Concert series by means of collected concert recordings and associated information, we have developed a Web-native digital musicology environment to help address this issue. We apply the Music Annotation Ontology, recently proposed in the literature, to capture linked data annotations on abstracted representations of music objects. For the first time, this approach is applied to information in both encoded score and recorded audio modalities, using an extended version of the mei-friend music encoding editor, and the Listen Here close-listening tool newly developed for this purpose. Together, this environment enables scholars to annotate digital music scores and immediately access the corresponding playback positions in collections of performance recordings. All tools are semantic Web applications, applying the Web-native Solid platform for social linked data for authentication and user-controlled data storage. The architecture integrates the different tools on the data level, permitting the future addition of further externally-developed tooling. Here, we motivate the developed environment within our research context, detail its implementation, and report on experiences applying the tooling within our music scholarship.},
+booktitle = {Proceedings of the 10th International Conference on Digital Libraries for Musicology},
+pages = {109–118},
+numpages = {10},
+keywords = {close listening, digital musicology, linked data, music annotation, music encoding, performance recordings},
+location = {<conf-loc>, <city>Milan</city>, <country>Italy</country>, </conf-loc>},
+series = {DLfM '23}
+}
+
  @article{Crawford2016music, 
     year={2016},
     author={Crawford, Tim and Lewis, Richard}, 
