### 2024-03-11 05752b9
Limit construction of guidelines URI to correct first element
