### 2024-03-11 7833a2c
secure elementAtCursor.previousSibling
