### 2024-03-12 71a9395
Open minimal facsimile and test absence of elements
