### 2024-03-12 34afc34
Move zone and check coordinates
