### 2024-03-12 6edbc16
Delete zone at the end
