+++ app/static/lib/main.js
@@ -1,6 +1,6 @@
 // mei-friend version and date
 export const version = '1.0.11';
-export const versionDate = '7 March 2024'; // use full or 3-character english months, will be translated
+export const versionDate = '12 March 2024'; // use full or 3-character english months, will be translated
 
 var vrvWorker;
 var spdWorker;
