### 2024-03-13 64b8098
add initial playback tests
