### 2024-03-13 2886d2f
Refactor panel tests
