### 2024-03-13 7816dcb
Test MIDI playback highlights notes
