### 2024-03-13 51b7352
test auto scrolling during playback
