### 2024-03-13 d517b13
test auto paging during playback
