### 2024-03-14 0d28d17
fix appInfo test
