### 2024-03-14 279a696
make playback tests less flaky
