### 2024-03-14 cac120c
test playback with speedmode
