### 2024-03-14 86ab8cf
Add basic translator checks
