+++ e2e/editing.spec.ts
@@ -7,35 +7,61 @@ test.beforeEach(async ({ page }) => {
 });
 
 test.describe('Test editing functionality.', () => {
-
-  // TODO: Split into test.step 
   test('Add, modify, and delete slur to selected elements', async ({ page }) => {
-    // open file from URL
-    await openUrl(
-      page,
-      'https://raw.githubusercontent.com/trompamusic-encodings/Beethoven_WoO70_BreitkopfHaertel/master/Beethoven_WoO70-Breitkopf.mei'
-    );
-    // click on #nextPageButton
-    await page.locator('#nextPageButton').click();
+    let uuid: string | null; // id of the new slur
 
-    // click on decreaseScaleButton
-    await page.locator('#decreaseScaleButton').click({
-      clickCount: 3,
+    await test.step('Open MEI file, turn page, and descrease scaling', async () => {
+      // open file from URL
+      await openUrl(
+        page,
+        'https://raw.githubusercontent.com/trompamusic-encodings/Beethoven_WoO70_BreitkopfHaertel/master/Beethoven_WoO70-Breitkopf.mei'
+      );
+      // click on #nextPageButton
+      await page.locator('#nextPageButton').click();
+
+      // check whether note note-0000001568544877 is visible
+      await expect(page.locator('#note-0000001568544877')).toBeVisible();
+
+      // retrieve width from svg inside verovio-panel
+      const height = await page.locator('#verovio-panel svg').first().getAttribute('height');
+      console.log('Height before: ', height);
+
+      // click on decreaseScaleButton
+      await page.locator('#decreaseScaleButton').click({
+        clickCount: 1,
+      });
+
+      // poll until width attribute in svg is smaller than width before
+      await expect(async () => {
+        const newHeight = await page.locator('#verovio-panel svg').first().getAttribute('height');
+        expect(newHeight).not.toEqual(height);
+        console.log('Height after: ', newHeight);
+      }).toPass({ timeout: 5000 });
     });
 
-    // check first note on second page
-    await expect(page.locator('#note-0000001568544877')).toBeVisible();
+    await test.step('Select two elements and add slur to selected elements', async () => {
+      // select two elements and add slur to selected elements
+      let elements = ['note-0000001751146466', 'note-0000001042243679'];
+      await selectElements(page, elements);
+      uuid = await insertElement(page, elements, 'slur', 'addSlur');
+      console.log('New slur inserted: ', uuid);
 
-    // select two elements and add slur to selected elements
-    let elements = ['note-0000001751146466', 'note-0000001042243679'];
-    let uuid = await insertElement(page, elements, 'slur', 'addSlur');
-    console.log('New slur inserted: ', uuid);
+      // test whether the new slur has attributes startid and endid to the first and last elements respectively
+      const slur = page.locator('g#' + uuid);
+      await expect(slur).toBeVisible();
+      const currentLine = page.locator('span[role="presentation"]', { has: page.getByText(uuid!) });
+      expect(currentLine).toBeVisible();
+      let re = new RegExp('startid="#' + elements[0]);
+      expect(currentLine).toHaveText(re);
+      re = new RegExp('endid="#' + elements.at(-1));
+      expect(currentLine).toHaveText(re);
+    });
 
-    // delete this slur now
-    if (uuid) {
+    await test.step('Delete slur', async () => {
+      // delete this slur now
       console.log('Deleting slur: ', uuid);
       await deleteElement(page, [uuid]);
-    }
+    });
   });
 });
 
@@ -53,9 +79,10 @@ async function selectElements(page: Page, ids: string[]) {
     if (i > 0) {
       mod['modifiers'] = ['Meta'];
     }
-    let note = page.locator('g#' + id);
[…]
