### 2024-03-15 d56a3c9
Increase waiting time in speed mode playback test
