+++ e2e/playback.spec.ts
@@ -109,7 +109,7 @@ test.describe('Test that MIDI playback works correctly', () => {
     // ... something should be highlighted ...
     await expect(page.locator('.currently-playing').first()).toBeVisible();
     // ... then, after 2 seconds of playback (i.e., past end of page) ...
-    await page.waitForTimeout(2000);
+    await page.waitForTimeout(4000);
     // ... nothing should be highlighted ...
     await expect(page.locator('.currently-playing')).not.toBeVisible();
     // ... and the page indicator should not have changed
