### 2024-03-15 0af1992
Small fix for facsimile zone delete test
