### 2024-03-15 2e0f2c0
Add slur to one element and check correct selection of second
