+++ e2e/editing.spec.ts
@@ -7,7 +7,75 @@ test.beforeEach(async ({ page }) => {
 });
 
 test.describe('Test editing functionality.', () => {
-  test('Add, modify, and delete slur to selected elements', async ({ page }) => {
+  test('Add, modify, and delete slur to one selected element', async ({ page }) => {
+    let uuid: string | null; // id of the new slur
+
+    await test.step('Open MEI file, flip to last page', async () => {
+      // open file from URL
+      await openUrl(
+        page,
+        'https://raw.githubusercontent.com/trompamusic-encodings/Beethoven_WoO70_BreitkopfHaertel/master/Beethoven_WoO70-Breitkopf.mei'
+      );
+      // click on #nextPageButton
+      await page.locator('#lastPageButton').click();
+
+      // check whether first note on last page 'note-0000000604712286' is visible
+      await expect(page.locator('#note-0000000604712286')).toBeVisible();
+
+      // retrieve width from svg inside verovio-panel
+      const height = await page.locator('#verovio-panel svg').first().getAttribute('height');
+      console.log('Height before: ', height);
+    });
+
+    await test.step('Select one element and add slur to that element', async () => {
+      // select one element and add slur to that element (expecting next note to be selected)
+      let elements = ['note-0000002071553041'];
+      await selectElements(page, elements);
+      uuid = await insertElement(page, elements, 'slur', 'addSlur');
+      console.log('New slur inserted: ', uuid);
+
+      // test whether the new slur has attributes startid and endid to the first and last elements respectively
+      const slur = page.locator('g#' + uuid);
+      await expect(slur).toBeVisible();
+      const currentLine = page.locator('span[role="presentation"]', { has: page.getByText(uuid!) });
+      await expect(currentLine).toBeVisible();
+      await expect(currentLine).toHaveText(new RegExp('startid="#' + elements[0]));
+      await expect(currentLine).toHaveText(new RegExp('endid="#note-0000001587074923"')); // expected next note
+    });
+
+    await test.step('Modify slur', async () => {
+      // modify this slur now
+      console.log('Modifying slur: ', uuid);
+
+      // get notation panel into focus
+      await page.keyboard.press('Shift+Space');
+      // check whether notation has class focus-visible
+      const notation = page.locator('#notation');
+      await expect(notation).toHaveClass(/focus-visible/);
+
+      // press X to flip slur direction
+      await page.keyboard.press('x');
+      // check encoding whether slur has curvedir="above"
+      let currentLine = page.locator('span[role="presentation"]', { has: page.getByText(uuid!) });
+      await expect(currentLine).toBeVisible();
+      await expect(currentLine).toHaveText(new RegExp('curvedir="above"'));
+
+      // press X to flip slur direction
+      await page.keyboard.press('x');
+      // check encoding whether slur has curvedir="above"
+      currentLine = page.locator('span[role="presentation"]', { has: page.getByText(uuid!) });
+      await expect(currentLine).toBeVisible();
+      await expect(currentLine).toHaveText(new RegExp('curvedir="below"'));
+    });
+
+    await test.step('Delete slur', async () => {
+      // delete this slur now
+      console.log('Deleting slur: ', uuid);
+      await deleteElement(page, [uuid!]);
+    });
+  });
+
+  test('Add and delete slur to two selected elements', async ({ page }) => {
     let uuid: string | null; // id of the new slur
 
     await test.step('Open MEI file, turn page, and descrease scaling', async () => {
@@ -19,7 +87,7 @@ test.describe('Test editing functionality.', () => {
       // click on #nextPageButton
       await page.locator('#nextPageButton').click();
 
[…]
