### 2024-03-15 2c21e77
Make playback tests run on slow machine with Chromium
