### 2024-04-02 c61b8c3
Make facsimile tests pass with Firefox
