### 2024-04-02 e92ed6e
Bump to 1.0.12
