+++ app/static/lib/main.js
@@ -1,5 +1,5 @@
 // mei-friend version and date
-export const version = '1.0.11';
+export const version = '1.0.12';
 export const versionDate = '2 April 2024'; // use full or 3-character english months, will be translated
 
 var vrvWorker;
