### 2024-04-03 f7dc2c5
Add comments to facsimile tests
