### 2024-04-03 8452a3f
Update insertElement for testing editing (firefox still flaky)
