### 2024-04-03 187428c
More editing code polishing, firefox still flaky
