### 2024-04-04 e4bacfd
Add test for inserting notes, rests, chords, accidentals
