+++ e2e/editing.spec.ts
@@ -6,8 +6,261 @@ test.beforeEach(async ({ page }) => {
   await setupPage(page);
 });
 
-test.describe('Test insertion functionality.', () => {
-  test('Add, modify, and delete slur to one selected element', async ({ page }) => {
+test.describe('Insert notes, rests, and accidentals.', () => {
+  test('Insert new note', async ({ page }) => {
+    let uuid: string | null; // id of the new slur
+    const noteId = 'note-0000001117852400'; // selected note
+
+    await test.step('Open MEI file and select a note in first measure', async () => {
+      // open file from URL
+      await openUrl(
+        page,
+        'https://raw.githubusercontent.com/trompamusic-encodings/Beethoven_WoO70_BreitkopfHaertel/master/Beethoven_WoO70-Breitkopf.mei'
+      );
+
+      // check whether page loaded successfully
+      await expect(page.locator('#note-0000001631474113')).toBeVisible();
+
+      // select #note-0000001117852400
+      await selectElements(page, [noteId]);
+
+      // check whether note is highlighted
+      await expect(page.locator('#' + noteId)).toHaveClass('note highlighted');
+    });
+
+    await test.step('Insert a note and check its content', async () => {
+      let uuid = await insertElement(page, [noteId], 'note', 'addNote');
+      console.log('New note inserted: ', uuid);
+
+      // check that the new note is visible
+      await expect(page.locator(`#${uuid}`)).toBeVisible();
+
+      const currentLine = page.locator('span[role="presentation"]', { has: page.getByText(uuid!) });
+      await expect(currentLine).toBeVisible();
+      // test attributes of the new note
+      await expect(currentLine).toHaveText(new RegExp('pname="a"'));
+      await expect(currentLine).toHaveText(new RegExp('oct="4"'));
+      await expect(currentLine).toHaveText(new RegExp('dur="8"'));
+    });
+  });
+
+  test('Convert new notes to chord', async ({ page }) => {
+    const ids = ['note-0000001117852400']; // first selected note
+    let chordId: string | null; // id of the new chord
+
+    await test.step('Open MEI file and select a note in first measure', async () => {
+      // open file from URL
+      await openUrl(
+        page,
+        'https://raw.githubusercontent.com/trompamusic-encodings/Beethoven_WoO70_BreitkopfHaertel/master/Beethoven_WoO70-Breitkopf.mei'
+      );
+
+      // check whether page loaded successfully
+      await expect(page.locator('#note-0000001631474113')).toBeVisible();
+
+      // select #note-0000001117852400
+      await selectElements(page, ids);
+
+      // check whether note is highlighted
+      await expect(page.locator('#' + ids)).toHaveClass('note highlighted');
+    });
+
+    await test.step('Add a note and change its pitch', async () => {
+      let uuid = await insertElement(page, ids, 'note', 'addNote');
+      console.log('First new note inserted: ', uuid);
+
+      // check that the new note is visible
+      await expect(page.locator(`#${uuid}`)).toBeVisible();
+
+      // click on note to select it
+      await selectElements(page, [uuid!]);
+
+      // change pitch up of the new note
+      await clickManipulate(page, 'pitchUpDiat');
+
+      ids.push(uuid!);
+
+      uuid = await insertElement(page, ids, 'note', 'addNote');
+      console.log('Second new note inserted: ', uuid);
+
+      // check that the new note is visible
+      await expect(page.locator(`#${uuid}`)).toBeVisible();
+
+      // click on note to select it
+      await selectElements(page, [uuid!]);
+
+      // change pitch up of the new note
+      await clickManipulate(page, 'pitchOctaveDown');
+
+      ids.push(uuid!);
+    });
+
[…]
