### 2024-04-04 e05ec95
Test 5 accidentals and make editor tests less flaky
