+++ e2e/editing.spec.ts
@@ -25,6 +25,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await selectElements(page, [noteId]);
 
       // check whether note is highlighted
+      await expect(page.locator('#' + noteId)).toBeVisible();
       await expect(page.locator('#' + noteId)).toHaveClass('note highlighted');
     });
 
@@ -62,6 +63,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await selectElements(page, ids);
 
       // check whether note is highlighted
+      await expect(page.locator('#' + ids)).toBeVisible();
       await expect(page.locator('#' + ids)).toHaveClass('note highlighted');
     });
 
@@ -150,16 +152,14 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await selectElements(page, [noteId]);
 
       // check whether note is highlighted
+      await expect(page.locator('#' + noteId)).toBeVisible();
       await expect(page.locator('#' + noteId)).toHaveClass('note highlighted');
     });
 
     await test.step('Convert new note to rest', async () => {
       // convert the new note to a rest
       console.log('Converting note to rest: ', uuid);
-      await page.locator('#manipulateMenuTitle').click();
-      await page.locator('#manipulateMenuTitle').hover();
-      await expect(page.locator('#convertNoteToRest')).toBeVisible();
-      await page.locator('#convertNoteToRest').click();
+      await clickManipulate(page, 'convertNoteToRest');
 
       // find the new rest
       const rest = page.locator('g.rest.highlighted');
@@ -194,6 +194,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await selectElements(page, ids);
 
       // check whether notes are highlighted
+      await expect(page.locator('#' + ids[0])).toBeVisible();
       await expect(page.locator('#' + ids[0])).toHaveClass('note highlighted');
       await expect(page.locator('#' + ids[1])).toHaveClass('note highlighted');
       await expect(page.locator('#' + ids[2])).toHaveClass('note highlighted');
@@ -204,7 +205,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
     });
 
     await test.step('Delete selected elements', async () => {
-      let uuid = await deleteElement(page, ids);
+      let uuid = await deleteElements(page, ids);
       console.log('Deleted elements: ', ids);
 
       // check that there are no child notes in the layer anymore
@@ -212,9 +213,16 @@ test.describe('Insert notes, rests, and accidentals.', () => {
     });
   });
 
-  test('Insert/delete double sharp', async ({ page }) => {
-    const ids = ['note-0000000417567361']; // selected note
-    let uuid: string | null; // id of the new accidental
+  test('Insert/delete accidentals', async ({ page }) => {
+    const ids = [
+      'note-0000001403952294',
+      'note-0000002094055458',
+      'note-0000001828075570',
+      'note-0000000417567361',
+      'note-0000000078077931',
+    ]; // selected note
+    const accids = ['addDoubleSharp', 'addSharp', 'addNatural', 'addFlat', 'addDoubleFlat'];
+    let uuids: string[] = []; // id of the new accidental
 
     await test.step('Open MEI file and select notes in first measure', async () => {
       // open file from URL
@@ -225,36 +233,49 @@ test.describe('Insert notes, rests, and accidentals.', () => {
 
       // check whether page loaded successfully
       await expect(page.locator('g#note-0000001631474113')).toBeVisible();
-
-      // select #note-0000000417567361
-      await selectElements(page, ids);
-
-      // check whether notes are highlighted
-      await expect(page.locator('g#' + ids[0])).toHaveClass('note highlighted');
     });
 
[…]
