+++ e2e/editing.spec.ts
@@ -22,7 +22,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await expect(page.locator('#note-0000001631474113')).toBeVisible();
 
       // select #note-0000001117852400
-      await selectElements(page, [noteId]);
+      await selectElementsByClicking(page, [noteId]);
 
       // check whether note is highlighted
       await expect(page.locator('#' + noteId)).toBeVisible();
@@ -60,7 +60,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await expect(page.locator('#note-0000001631474113')).toBeVisible();
 
       // select #note-0000001117852400
-      await selectElements(page, ids);
+      await selectElementsByClicking(page, ids);
 
       // check whether note is highlighted
       await expect(page.locator('#' + ids)).toBeVisible();
@@ -75,7 +75,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await expect(page.locator(`#${uuid}`)).toBeVisible();
 
       // click on note to select it
-      await selectElements(page, [uuid!]);
+      await selectElementsByClicking(page, [uuid!]);
 
       // change pitch up of the new note
       await clickManipulate(page, 'pitchUpDiat');
@@ -89,7 +89,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await expect(page.locator(`#${uuid}`)).toBeVisible();
 
       // click on note to select it
-      await selectElements(page, [uuid!]);
+      await selectElementsByClicking(page, [uuid!]);
 
       // change pitch up of the new note
       await clickManipulate(page, 'pitchOctaveDown');
@@ -99,7 +99,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
 
     await test.step('Select all three notes and convert to chord', async () => {
       // select elements
-      await selectElements(page, ids);
+      await selectElementsByClicking(page, ids);
 
       // check whether notes are highlighted
       await expect(page.locator('#' + ids[0])).toHaveClass('note highlighted');
@@ -149,7 +149,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await expect(page.locator('#note-0000001631474113')).toBeVisible();
 
       // select #note-0000001117852400
-      await selectElements(page, [noteId]);
+      await selectElementsByClicking(page, [noteId]);
 
       // check whether note is highlighted
       await expect(page.locator('#' + noteId)).toBeVisible();
@@ -191,7 +191,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await expect(page.locator('#note-0000001631474113')).toBeVisible();
 
       // select #note-0000001117852400
-      await selectElements(page, ids);
+      await selectElementsByClicking(page, ids);
 
       // check whether notes are highlighted
       await expect(page.locator('#' + ids[0])).toBeVisible();
@@ -235,10 +235,10 @@ test.describe('Insert notes, rests, and accidentals.', () => {
       await expect(page.locator('g#note-0000001631474113')).toBeVisible();
     });
 
-    await test.step('Insert accidentals', async () => {
+    await test.step('Insert and delete accidentals', async () => {
       for (let i = 0; i < ids.length; i++) {
         // select #note-0000000417567361
-        await selectElements(page, ids.slice(i, i + 1));
+        await selectElementsByClicking(page, ids.slice(i, i + 1));
 
         // find 'g#' + ids[0] and check whether it has an accid as child
         await expect(async () => {
@@ -252,6 +252,7 @@ test.describe('Insert notes, rests, and accidentals.', () => {
 
         // insert double sharp
         await clickInsert(page, accids[i]);
+        console.log('Accidental inserted: ', accids[i]);
 
         // check whether note has an accid as child (not working in Firefox)
[…]
