### 2024-04-16 mei-friend/mei-friend#61 @wergo
Hi @stefan-balke thank you for your suggestion regarding testing. We have now implemented comprehensive end-to-end tests using PlayWright. Currently, there are 44 individual tests covereing different aspects of the application relating to editing, input, page load, facsimile display and editing, MIDI playback, application settings, UI translation, and URL parameters. These cover the major functionalities of mei-friend.

We will continue developing further tests to cover edge cases and new features as they come along. We are also still working making the existing tests more robust, as some of them can currently exhibit flaky behavior (run well on one occasion, but not on another).

Independently of these end-to-end tests, mei-friend also is able to run internal tests when deployed in the development environment. Currently we use this to validate the completeness of the language packs.
