### 2024-04-19 6e9bfab
Clean up code in drawFacsimile()
