+++ app/static/lib/facsimile.js
@@ -67,58 +67,66 @@ export function clearFacsimile() {
  * each containing own coordinates (ulx, uly, lrx, lry)
  * and containing surface info (target, width, height)
  * @param {Document} xmlDoc
- * Modifies global variable facs
+ *
+ * Note:
+ * 1) Clears and creates global variable facs.
+ * 2) Loads only first facsimile element in xmlDoc.
  */
 export function loadFacsimile(xmlDoc) {
   let facsimile = xmlDoc.querySelector('facsimile');
-  if (facsimile) {
-    facs = {};
-    // look for surface elements
-    let surfaces = facsimile.querySelectorAll('surface');
-    surfaces.forEach((s) => {
-      let id;
-      let { target, width, height } = fillGraphic(s.querySelector('graphic'));
-      if (s.hasAttribute('xml:id')) id = s.getAttribute('xml:id');
-      if (id) {
-        facs[id] = {};
-        if (target) facs[id]['target'] = target;
-        if (width) facs[id]['width'] = width;
-        if (height) facs[id]['height'] = height;
-        facs[id]['type'] = 'surface';
-      }
-    });
-    // look for zone elements
-    let zones = facsimile.querySelectorAll('zone');
-    console.debug('facsimile.loadFacsimile(): loading ' + zones.length + ' zones.');
-    zones.forEach((z) => {
-      let id, ulx, uly, lrx, lry;
-      if (z.hasAttribute('xml:id')) id = z.getAttribute('xml:id');
-      if (z.hasAttribute('ulx')) ulx = z.getAttribute('ulx');
-      if (z.hasAttribute('uly')) uly = z.getAttribute('uly');
-      if (z.hasAttribute('lrx')) lrx = z.getAttribute('lrx');
-      if (z.hasAttribute('lry')) lry = z.getAttribute('lry');
-      let { target, width, height } = fillGraphic(z.parentElement.querySelector('graphic'));
-      if (id) {
-        facs[id] = {};
-        facs[id]['type'] = 'zone';
-        if (target) facs[id]['target'] = target;
-        if (width) facs[id]['width'] = width;
-        if (height) facs[id]['height'] = height;
-        if (ulx) facs[id]['ulx'] = ulx;
-        if (uly) facs[id]['uly'] = uly;
-        if (lrx) facs[id]['lrx'] = lrx;
-        if (lry) facs[id]['lry'] = lry;
-        let measure = xmlDoc.querySelector('[facs="#' + id + '"]');
-        if (measure) {
-          if (measure.hasAttribute('xml:id')) facs[id]['pointerId'] = measure.getAttribute('xml:id');
-          if (measure.hasAttribute('n')) facs[id]['pointerN'] = measure.getAttribute('n');
+  if (!facsimile) {
+    console.info('facsimile.loadFacsimile(): No facsimile element found.');
+    return;
+  }
+  facs = {};
+
+  // look for surface elements
+  let surfaces = facsimile.querySelectorAll('surface');
+  surfaces.forEach((s) => {
+    let id;
+    let { target, width, height } = fillGraphic(s.querySelector('graphic'));
+    if (s.hasAttribute('xml:id')) id = s.getAttribute('xml:id');
+    if (id) {
+      facs[id] = {};
+      if (target) facs[id]['target'] = target;
+      if (width) facs[id]['width'] = width;
+      if (height) facs[id]['height'] = height;
+      facs[id]['type'] = 'surface';
+    }
+  });
+
+  // look for zone elements
+  let zones = facsimile.querySelectorAll('zone');
+  console.debug('facsimile.loadFacsimile(): loading ' + zones.length + ' zones.');
+  zones.forEach((z) => {
+    let id, ulx, uly, lrx, lry;
+    if (z.hasAttribute('xml:id')) id = z.getAttribute('xml:id');
[…]
