### 2024-04-19 49bb967
Load facsimile multiple source images into separate SVGs
