+++ app/static/lib/facsimile.js
@@ -27,15 +27,15 @@ var ulx, uly;
  * @returns void
  */
 function showWarningText(txt = translator.lang.facsimileDefaultWarning.text) {
-  let svgContainer = document.getElementById('sourceImageContainer');
-  let svg = document.getElementById('sourceImageSvg');
-  if (!svg || !svgContainer) return;
+  // let svgContainer = document.getElementById('sourceImageContainer');
+  // let svg = document.getElementById('sourceImageSvg');
+  // if (!svg || !svgContainer) return;
 
-  // clear existing structures
-  svgContainer.removeAttribute('transform-origin');
-  svgContainer.removeAttribute('transform');
-  svg.removeAttribute('viewBox');
-  svg.innerHTML = '';
+  // // clear existing structures
+  // svgContainer.removeAttribute('transform-origin');
+  // svgContainer.removeAttribute('transform');
+  // svg.removeAttribute('viewBox');
+  // svg.innerHTML = '';
 
   let facsimileMessagePanel = document.getElementById('facsimileMessagePanel');
   facsimileMessagePanel.style.display = 'block';
@@ -82,7 +82,7 @@ export function loadFacsimile(xmlDoc) {
 
   // look for surface elements
   let surfaces = facsimile.querySelectorAll('surface');
-  surfaces.forEach((s) => {
+  surfaces.forEach((s, i) => {
     let id;
     let { target, width, height } = fillGraphic(s.querySelector('graphic'));
     if (s.hasAttribute('xml:id')) id = s.getAttribute('xml:id');
@@ -92,6 +92,8 @@ export function loadFacsimile(xmlDoc) {
       if (width) facs[id]['width'] = width;
       if (height) facs[id]['height'] = height;
       facs[id]['type'] = 'surface';
+      facs[id]['sourceImageNumber'] = i; // store number of image page
+      facs['sourceImage-' + i] = id;
     }
   });
 
@@ -105,6 +107,7 @@ export function loadFacsimile(xmlDoc) {
     if (z.hasAttribute('uly')) uly = z.getAttribute('uly');
     if (z.hasAttribute('lrx')) lrx = z.getAttribute('lrx');
     if (z.hasAttribute('lry')) lry = z.getAttribute('lry');
+    let parentId = z.parentElement.getAttribute('xml:id');
     let { target, width, height } = fillGraphic(z.parentElement.querySelector('graphic'));
     if (id) {
       facs[id] = {};
@@ -116,6 +119,7 @@ export function loadFacsimile(xmlDoc) {
       if (uly) facs[id]['uly'] = uly;
       if (lrx) facs[id]['lrx'] = lrx;
       if (lry) facs[id]['lry'] = lry;
+      if (parentId) facs[id]['surfaceId'] = parentId;
       let pointingElement = xmlDoc.querySelector('[facs="#' + id + '"]');
       if (pointingElement) {
         if (pointingElement.hasAttribute('xml:id')) {
@@ -156,9 +160,17 @@ export async function drawFacsimile() {
   let fullPage = document.getElementById('showFacsimileFullPage').checked;
   let facsimileMessagePanel = document.getElementById('facsimileMessagePanel');
   facsimileMessagePanel.style.display = 'none';
-  let svgContainer = document.getElementById('sourceImageContainer');
-  let svg = document.getElementById('sourceImageSvg');
-  if (!svg || !svgContainer) return;
+  let facsimilePanel = document.getElementById('facsimile-panel');
+  // let svgContainer = document.getElementById('sourceImageContainer');
+  // let svg = document.getElementById('sourceImageSvg');
+  // if (!svg || !svgContainer) return;
+
+  let zoomFactor = document.getElementById('facsimileZoomInput').value / 100;
[…]
