### 2024-04-19 60ba375
Scale multiple scoure images by width/height rather than transform
