+++ app/static/lib/facsimile.js
@@ -18,7 +18,7 @@ const rectangleLineWidth = defaultFacsimileRectangleLineWidth; // width of bound
 const rectangleColor = defaultFacsimileRectangleColor; // color of zone rectangles
 var listenerHandles = {};
 var resize = ''; // west, east, north, south, northwest, southeast etc
-var ulx, uly;
+var sourceImageBoxes = {}; // ulx, uly, lrx, lry of source image boxes per page
 
 /**
  * Show warning text to facsimile panel (as svg text element) and
@@ -94,6 +94,7 @@ export function loadFacsimile(xmlDoc) {
       facs[id]['type'] = 'surface';
       facs[id]['sourceImageNumber'] = i; // store number of image page
       facs['sourceImage-' + i] = id;
+      facs[id]['surfaceId'] = id;
     }
   });
 
@@ -167,15 +168,10 @@ export async function drawFacsimile() {
 
   let zoomFactor = document.getElementById('facsimileZoomInput').value / 100;
 
-  // clear svg and append image
+  // clear all svgs and sourceImageBoxes
   facsimilePanel.querySelectorAll('[*|id^="sourceImage-"]').forEach((c) => c.remove());
+  // c.querySelectorAll(':not(image)').forEach((e) => c.removeChild(e));
   // Array.from(svg.children).forEach((c) => svg.removeChild(c));
-
-  ulx = Number.MAX_VALUE; // boundary values for image envelope (left-upper corner is global)
-  uly = Number.MAX_VALUE;
-  let lrx = 0;
-  let lry = 0;
-  let zoneId = '';
 
   // list displayed zones (filter doubled note elements in tab notation, see Verovio issue #3600)
   let svgFacs = Array.from(document.querySelectorAll('[data-facs]')).filter(
@@ -184,59 +180,79 @@ export async function drawFacsimile() {
 
   // iterate over svgFacs (svg group elements with data-facs attributes) and retrieve zoneId
   for (let f of svgFacs) {
+    let zoneId = '';
     if (f.hasAttribute('data-facs')) {
       zoneId = rmHash(f.getAttribute('data-facs'));
     }
 
-    
     // retrieve source image number from surface/zone element
-    let sourceImageNumber = '';
-    let surfaceId = '';
-    surfaceId = facs[zoneId].type === 'zone' ? facs[zoneId]['surfaceId'] : zoneId;
+    let sourceImageNumber = -1;
+    let surfaceId = facs[zoneId].surfaceId;
     if (surfaceId && facs[surfaceId]) {
       sourceImageNumber = facs[surfaceId]['sourceImageNumber'];
     }
-    
+
     console.log('Facsimile zoneId: ' + zoneId + ' (' + f.id + ')' + ', type: ' + facs[zoneId]?.type);
     console.log('          surfaceId: ' + surfaceId + ', sourceImageNumber: ' + sourceImageNumber);
-    
+
     if (sourceImageNumber >= 0) {
+      // when not in full page mode, store envelope of source image per page
+      if (!fullPage) {
+        if (!sourceImageBoxes[sourceImageNumber])
+          sourceImageBoxes[sourceImageNumber] = { ulx: Number.MAX_VALUE, uly: Number.MAX_VALUE, lrx: 0, lry: 0 };
+        if (parseFloat(facs[zoneId].ulx) < sourceImageBoxes[sourceImageNumber].ulx)
+          sourceImageBoxes[sourceImageNumber].ulx = parseFloat(facs[zoneId].ulx) - rectangleLineWidth / 2;
+        if (parseFloat(facs[zoneId].uly) < sourceImageBoxes[sourceImageNumber].uly)
+          sourceImageBoxes[sourceImageNumber].uly = parseFloat(facs[zoneId].uly) - rectangleLineWidth / 2;
[…]
