### 2024-04-20 48224a1
Enable insert and edit callbacks
