+++ app/static/lib/facsimile.js
@@ -162,9 +162,6 @@ export async function drawFacsimile() {
   let facsimileMessagePanel = document.getElementById('facsimileMessagePanel');
   facsimileMessagePanel.style.display = 'none';
   let facsimilePanel = document.getElementById('facsimile-panel');
-  // let svgContainer = document.getElementById('sourceImageContainer');
-  // let svg = document.getElementById('sourceImageSvg');
-  // if (!svg || !svgContainer) return;
 
   let zoomFactor = document.getElementById('facsimileZoomInput').value / 100;
 
@@ -192,8 +189,8 @@ export async function drawFacsimile() {
       sourceImageNumber = facs[surfaceId]['sourceImageNumber'];
     }
 
-    console.log('Facsimile zoneId: ' + zoneId + ' (' + f.id + ')' + ', type: ' + facs[zoneId]?.type);
-    console.log('          surfaceId: ' + surfaceId + ', sourceImageNumber: ' + sourceImageNumber);
+    // console.log('Facsimile zoneId: ' + zoneId + ' (' + f.id + ')' + ', type: ' + facs[zoneId]?.type);
+    // console.log('          surfaceId: ' + surfaceId + ', sourceImageNumber: ' + sourceImageNumber);
 
     if (sourceImageNumber >= 0) {
       // when not in full page mode, store envelope of source image per page
@@ -216,6 +213,7 @@ export async function drawFacsimile() {
         // let div = document.createElement('div');
         let imageSvg = document.createElementNS(svgNameSpace, 'svg');
         imageSvg.id = 'sourceImage-' + sourceImageNumber;
+        imageSvg.setAttribute('data-sourceImageNumber', sourceImageNumber);
         // div.appendChild(imageSvg);
         facsimilePanel.appendChild(imageSvg);
 
@@ -556,7 +554,10 @@ export function zoomFacsimile(deltaPercent) {
  * @param {rect} rect
  */
 export function highlightZone(rect) {
-  let svg = document.getElementById('sourceImageSvg');
+  let svg = rect.closest('svg');
+  // console.log('Highlighting zone: ', rect.id, svg);
+  if (!svg) return;
+
   // remove event listerners
   for (let key in listenerHandles) {
     if (key === 'mousedown') {
@@ -568,9 +569,11 @@ export function highlightZone(rect) {
       if (ip) ip.removeEventListener(key, listenerHandles[key]);
     }
   }
-  // add zone resizer for selected zone box (only when linked to zone
+  // add zone resizer for selected zone box (only when linked to zone,
   // rather than to pointing element, ie. measure)
-  if (document.getElementById('editFacsimileZones').checked) listenerHandles = addZoneResizer(v, rect);
+  if (document.getElementById('editFacsimileZones').checked) {
+    listenerHandles = addZoneResizer(v, rect);
+  }
 
   // highlight rectangle
   rect.classList.add('highlighted');
@@ -587,7 +590,9 @@ export function highlightZone(rect) {
 export function addZoneResizer(v, rect) {
   let txt, txtX, txtY; // text element for zone number and its x/y coordinates
   var ip = document.getElementById('facsimile-panel');
-  var svg = document.getElementById('sourceImageSvg');
+  var svg = rect.closest('svg');
+  // console.log('Adding zone resizer for: ', rect.id, svg);
+  if (!ip || !svg) return;
   var start = {}; // starting point start.x, start.y
   var end = {}; // ending point
   var bb;
@@ -667,6 +672,7 @@ export function addZoneResizer(v, rect) {
       end.x = ev.clientX;
       end.y = ev.clientY;
 
+      if (!svg) return;
       var mx = svg.getScreenCTM().inverse();
[…]
