+++ app/static/lib/viewer.js
@@ -653,7 +653,7 @@ export default class Viewer {
         if (el) {
           el.forEach((e) => {
             // console.log('updatehighlight() e: ', e);
-            if (e.nodeName === 'rect' && e.closest('#sourceImageSvg')) {
+            if (e.nodeName === 'rect' && e.closest('#facsimile-panel')) {
               facs.highlightZone(e);
             } else if (e.hasAttribute('data-facs')) {
               let ref = utils.rmHash(e.getAttribute('data-facs'));
