### 2024-04-20 ddd2dfc
Avoid add double zone rectangles
