### 2024-04-20 a315346
Put source image svgs into div elements
