### 2024-04-20 f178ba8
Show image titles and loading message
