### 2024-04-20 7155713
Get image width/height from bounding box
