### 2024-04-20 7998021
Make splash screen confirm button focussed
