### 2024-04-20 77192d6
Add warning messages for incomplete facsimile encodings
