### 2024-04-20 e563a75
Fix viewBox envelope without full page
