+++ app/static/lib/facsimile.js
@@ -11,7 +11,6 @@ import { cm, fileLocationType, github, isCtrlOrCmd, meiFileLocation, translator,
 import { addZone, replaceInEditor } from './editor.js';
 import { attFacsimile } from './attribute-classes.js';
 import { defaultFacsimileRectangleColor, defaultFacsimileRectangleLineWidth, isFirefox } from './defaults.js';
-import { clock } from '../css/icons.js';
 
 var facs = {}; // facsimile structure in MEI file
 var sourceImages = {}; // object of source images
@@ -162,6 +161,7 @@ export async function drawFacsimile() {
 
   // clear all svgs and sourceImageBoxes
   clearSourceImages();
+  sourceImageBoxes = {};
   // c.querySelectorAll(':not(image)').forEach((e) => c.removeChild(e));
   // Array.from(svg.children).forEach((c) => svg.removeChild(c));
 
@@ -231,11 +231,10 @@ export async function drawFacsimile() {
           sourceImageBoxes[sourceImageNumber].lrx = parseFloat(facs[zoneId].lrx) + rectangleLineWidth / 2;
         if (parseFloat(facs[zoneId].lry) > sourceImageBoxes[sourceImageNumber].lry)
           sourceImageBoxes[sourceImageNumber].lry = parseFloat(facs[zoneId].lry) + rectangleLineWidth / 2;
+        // console.log('Facsimile envelope: ulx/uly; lrx/lry: ' + sourceImageBoxes[sourceImageNumber]);
       }
 
-      if (facsimilePanel.querySelector('#sourceImage-' + sourceImageNumber)) {
-        console.log('Image svg already appended: sourceImage-' + sourceImageNumber);
-      } else {
+      if (!facsimilePanel.querySelector('#sourceImage-' + sourceImageNumber)) {
         let imgName = createImageName(zoneId);
 
         // create new div and svg for source image
@@ -310,129 +309,18 @@ export async function drawFacsimile() {
         let viewBox = [];
         viewBox.push(sourceImageBoxes[sourceImageNumber].ulx);
         viewBox.push(sourceImageBoxes[sourceImageNumber].uly);
-        viewBox.push(sourceImageBoxes[sourceImageNumber].lrx - sourceImageBoxes[sourceImageNumber].ulx);
-        viewBox.push(sourceImageBoxes[sourceImageNumber].lry - sourceImageBoxes[sourceImageNumber].uly);
+        let width = sourceImageBoxes[sourceImageNumber].lrx - sourceImageBoxes[sourceImageNumber].ulx;
+        let height = sourceImageBoxes[sourceImageNumber].lry - sourceImageBoxes[sourceImageNumber].uly;
+        viewBox.push(width);
+        viewBox.push(height);
         svg.setAttribute('viewBox', viewBox.join(' '));
+        console.log('sourceImageBoxes: ', sourceImageBoxes[sourceImageNumber]);
+        console.log('Facsimile viewBox: ' + viewBox.join(' '));
+
+        svg.setAttribute('width', Math.round(width * zoomFactor));
+        svg.removeAttribute('height');
       }
     }
-    // else {
-    //   showWarningText(translator.lang.facsimileImgeNotLoadedWarning.text + ' \n(' + imgName + ').');
-    //   busy(false);
-    //   return;
-    // }
-  }
-
-  busy(false);
-  return;
-
-  if (svgFacs && fullPage) {
-    let firstZone = svgFacs.at(0);
-    if (firstZone && firstZone.hasAttribute('data-facs')) {
-      zoneId = rmHash(firstZone.getAttribute('data-facs'));
-    }
-  } else {
-    svgFacs.forEach((f) => {
[…]
