### 2024-04-20 388deaa
Optimize facsimile code
