### 2024-04-21 5088cbb
Add settings item for showing facsimile image titles in all translations
