+++ app/static/lib/defaults.js
@@ -733,6 +733,13 @@ export const meiFriendSettingsOptions = {
     type: 'bool',
     default: false,
   },
+  showFacsimileTitles: {
+    title: 'Show facsimile titles',
+    description: 'filled in by language packs',
+    type: 'bool',
+    default: true,
+  },
+
   // Supplied element
   titleSupplied: {
     title: 'Handle editorial content',
