### 2024-04-22 9a98f3c
Scroll to facsimile and/or notation, when clicking on zone or linked notation element
