+++ .github/workflows/playwright.yml
@@ -15,7 +15,6 @@ jobs:
           node-version: 18
       - name: Install dependencies
         run: npm ci
-        working-directory: e2e
       - name: Install Playwright Browsers
         run: npx playwright install --with-deps
         working-directory: e2e
