### 2024-05-08 #6002 @musicog
Dear @faroit, a quick update regarding our playwright tests: they are now automatically launched by a pre-push git-hook which executes on the 'testing', 'staging', and 'main' branches (i.e., those that are served by the publicly-visible *.mei-friend.mdw.ac.at service). The hook is supplied in the repository under a new `.githooks` directory. The `INSTALL.md` file has been updated with a single command (`git config`) that needs to be run to enable this.
