### 2024-05-10 mei-friend/mei-friend#61 @stefan-balke
Hi @wergo,

as far as I understand it, playwrights tests on the JS components. How about the Python side of things? Is there anything that should be tested regularly? There should be some endpoints which take only certain types (or should take) etc.

Best,
Stefan
