### 2024-05-10 mei-friend/mei-friend#59 @stefan-balke
Yes, this gives a first overview. Everything else needs to be explored through code or direct communictation, which is fine! Thanks!
