### 2024-05-10 #6002 @stefan-balke
Hey @musicog,

thanks for the efforts. I commented on the respective issues.

Thanks for all the work and sorry for the delay, I was pretty busy in the last months, but I hope, that this review process was also a little help for you to make this tool better. It's starting to get used by the community so this is a big win!

From my side, I would be happy to proceed to publication.
