### 2024-05-11 #6002 @musicog
@stefan-balke No need to apologise! Thank you very much (also to @rlskoeser!) for your in-depth feedback, this was very helpful to put mei-friend on a more sustainable footing. We'll continue to work further in this direction in future work.
