### 2024-05-12 mei-friend/mei-friend#70 @musicog
Thanks for this suggestion, and apologies for the late response. We've addressed this by:
* Including local copies of our dependencies (alongside licenses) in a new app/static/deps folder
* Adding CodeMirror as a submodule (rather than the suggested npm package) -- see below
* Using npm only to install the optional tests package (Playwright)

We chose to remain with the submodule rather than the npm dependency for CodeMirror because this removes npm entirely as a core requirement -- it is only necessary for Playwright installation.

We feel that the set-up now strikes a good balance between maintainability and usability (from the perspective of someone installing a local copy of mei-friend). It does mean that we will need to periodically check our dependencies for critical updates; we acknowledge this and will endeavour to do so regularly.
