### 2024-05-17 #6002 @rlskoeser
I was able to run through the updated installation instructions and run the application, and also installed and ran the end-to-end tests. They didn't all pass for me, but I don't know if that is expected or if I'm using the correct options.  The install still has a lot of manual steps that seems like that could be made easier with software package management, but I understand you made that decision deliberately so I'm ok signing off if everyone else is.

Was there an updated version of the software paper or did that not change?
