### 2024-05-21 #6002 @musicog
@rlskoeser Thanks again for your feedback! Regarding the paper, the following updates were undertaken: <https://github.com/mei-friend/mei-friend/compare/d4cb7..f6f43>

In summary:

- Expanded summary to more clearly identifying target audience(s) and key innovations that set the editor apart
- Add info on `testing` environment
- Updated screenshot of interface
- Update visitor count
- Add funding acknowledgement for Anna Plaksin's extension development
- Added citation to DLfM 2023 paper ("Listen Here!")
- Fixed DOI problem in another citation (DLfM 2019 "Interweaving...")
- Lightly edit to retain fit to 4 pages
