### 2024-05-23 #6002 @wergo
We have checked all missing DOIs. They are either bib items not referenced by the paper or legitimately do not have DOIs. We have added a URL for Geertinger, 2021 (Digital Encoding of Music Notation with MEI).
