### 2024-05-28 #6002 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
